% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_associate_fleet}
\alias{appstream_associate_fleet}
\title{Associates the specified fleet with the specified stack}
\usage{
appstream_associate_fleet(FleetName, StackName)
}
\arguments{
\item{FleetName}{[required] The name of the fleet.}

\item{StackName}{[required] The name of the stack.}
}
\value{
An empty list.
}
\description{
Associates the specified fleet with the specified stack.
}
\section{Request syntax}{
\preformatted{svc$associate_fleet(
  FleetName = "string",
  StackName = "string"
)
}
}

\keyword{internal}
