% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_create_app_block}
\alias{appstream_create_app_block}
\title{Creates an app block}
\usage{
appstream_create_app_block(
  Name,
  Description = NULL,
  DisplayName = NULL,
  SourceS3Location,
  SetupScriptDetails = NULL,
  Tags = NULL,
  PostSetupScriptDetails = NULL,
  PackagingType = NULL
)
}
\arguments{
\item{Name}{[required] The name of the app block.}

\item{Description}{The description of the app block.}

\item{DisplayName}{The display name of the app block. This is not displayed to the user.}

\item{SourceS3Location}{[required] The source S3 location of the app block.}

\item{SetupScriptDetails}{The setup script details of the app block. This must be provided for the
\code{CUSTOM} PackagingType.}

\item{Tags}{The tags assigned to the app block.}

\item{PostSetupScriptDetails}{The post setup script details of the app block. This can only be
provided for the \code{APPSTREAM2} PackagingType.}

\item{PackagingType}{The packaging type of the app block.}
}
\description{
Creates an app block.

See \url{https://www.paws-r-sdk.com/docs/appstream_create_app_block/} for full documentation.
}
\keyword{internal}
