% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_create_app_block_builder_streaming_url}
\alias{appstream_create_app_block_builder_streaming_url}
\title{Creates a URL to start a create app block builder streaming session}
\usage{
appstream_create_app_block_builder_streaming_url(
  AppBlockBuilderName,
  Validity = NULL
)
}
\arguments{
\item{AppBlockBuilderName}{[required] The name of the app block builder.}

\item{Validity}{The time that the streaming URL will be valid, in seconds. Specify a
value between 1 and 604800 seconds. The default is 3600 seconds.}
}
\description{
Creates a URL to start a create app block builder streaming session.

See \url{https://www.paws-r-sdk.com/docs/appstream_create_app_block_builder_streaming_url/} for full documentation.
}
\keyword{internal}
