% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_create_updated_image}
\alias{appstream_create_updated_image}
\title{Creates a new image with the latest Windows operating system updates,
driver updates, and AppStream 2}
\usage{
appstream_create_updated_image(existingImageName, newImageName,
  newImageDescription, newImageDisplayName, newImageTags, dryRun)
}
\arguments{
\item{existingImageName}{[required] The name of the image to update.}

\item{newImageName}{[required] The name of the new image. The name must be unique within the AWS
account and Region.}

\item{newImageDescription}{The description to display for the new image.}

\item{newImageDisplayName}{The name to display for the new image.}

\item{newImageTags}{The tags to associate with the new image. A tag is a key-value pair, and
the value is optional. For example, Environment=Test. If you do not
specify a value, Environment=.

Generally allowed characters are: letters, numbers, and spaces
representable in UTF-8, and the following special characters:

_ . : / = + \\ - @

If you do not specify a value, the value is set to an empty string.

For more information about tags, see \href{https://docs.aws.amazon.com/appstream2/latest/developerguide/tagging-basic.html}{Tagging Your Resources}
in the \emph{Amazon AppStream 2.0 Administration Guide}.}

\item{dryRun}{Indicates whether to display the status of image update availability
before AppStream 2.0 initiates the process of creating a new updated
image. If this value is set to \code{true}, AppStream 2.0 displays whether
image updates are available. If this value is set to \code{false}, AppStream
2.0 initiates the process of creating a new updated image without
displaying whether image updates are available.}
}
\value{
A list with the following syntax:\preformatted{list(
  image = list(
    Name = "string",
    Arn = "string",
    BaseImageArn = "string",
    DisplayName = "string",
    State = "PENDING"|"AVAILABLE"|"FAILED"|"COPYING"|"DELETING"|"CREATING"|"IMPORTING",
    Visibility = "PUBLIC"|"PRIVATE"|"SHARED",
    ImageBuilderSupported = TRUE|FALSE,
    ImageBuilderName = "string",
    Platform = "WINDOWS"|"WINDOWS_SERVER_2016"|"WINDOWS_SERVER_2019",
    Description = "string",
    StateChangeReason = list(
      Code = "INTERNAL_ERROR"|"IMAGE_BUILDER_NOT_AVAILABLE"|"IMAGE_COPY_FAILURE",
      Message = "string"
    ),
    Applications = list(
      list(
        Name = "string",
        DisplayName = "string",
        IconURL = "string",
        LaunchPath = "string",
        LaunchParameters = "string",
        Enabled = TRUE|FALSE,
        Metadata = list(
          "string"
        )
      )
    ),
    CreatedTime = as.POSIXct(
      "2015-01-01"
    ),
    PublicBaseImageReleasedDate = as.POSIXct(
      "2015-01-01"
    ),
    AppstreamAgentVersion = "string",
    ImagePermissions = list(
      allowFleet = TRUE|FALSE,
      allowImageBuilder = TRUE|FALSE
    ),
    ImageErrors = list(
      list(
        ErrorCode = "IAM_SERVICE_ROLE_MISSING_ENI_DESCRIBE_ACTION"|"IAM_SERVICE_ROLE_MISSING_ENI_CREATE_ACTION"|"IAM_SERVICE_ROLE_MISSING_ENI_DELETE_ACTION"|"NETWORK_INTERFACE_LIMIT_EXCEEDED"|"INTERNAL_SERVICE_ERROR"|"IAM_SERVICE_ROLE_IS_MISSING"|"MACHINE_ROLE_IS_MISSING"|"STS_DISABLED_IN_REGION"|"SUBNET_HAS_INSUFFICIENT_IP_ADDRESSES"|"IAM_SERVICE_ROLE_MISSING_DESCRIBE_SUBNET_ACTION"|"SUBNET_NOT_FOUND"|"IMAGE_NOT_FOUND"|"INVALID_SUBNET_CONFIGURATION"|"SECURITY_GROUPS_NOT_FOUND"|"IGW_NOT_ATTACHED"|"IAM_SERVICE_ROLE_MISSING_DESCRIBE_SECURITY_GROUPS_ACTION"|"FLEET_STOPPED"|"FLEET_INSTANCE_PROVISIONING_FAILURE"|"DOMAIN_JOIN_ERROR_FILE_NOT_FOUND"|"DOMAIN_JOIN_ERROR_ACCESS_DENIED"|"DOMAIN_JOIN_ERROR_LOGON_FAILURE"|"DOMAIN_JOIN_ERROR_INVALID_PARAMETER"|"DOMAIN_JOIN_ERROR_MORE_DATA"|"DOMAIN_JOIN_ERROR_NO_SUCH_DOMAIN"|"DOMAIN_JOIN_ERROR_NOT_SUPPORTED"|"DOMAIN_JOIN_NERR_INVALID_WORKGROUP_NAME"|"DOMAIN_JOIN_NERR_WORKSTATION_NOT_STARTED"|"DOMAIN_JOIN_ERROR_DS_MACHINE_ACCOUNT_QUOTA_EXCEEDED"|"DOMAIN_JOIN_NERR_PASSWORD_EXPIRED"|"DOMAIN_JOIN_INTERNAL_SERVICE_ERROR",
        ErrorMessage = "string",
        ErrorTimestamp = as.POSIXct(
          "2015-01-01"
        )
      )
    )
  ),
  canUpdateImage = TRUE|FALSE
)
}
}
\description{
Creates a new image with the latest Windows operating system updates,
driver updates, and AppStream 2.0 agent software.

For more information, see the "Update an Image by Using Managed
AppStream 2.0 Image Updates" section in \href{https://docs.aws.amazon.com/appstream2/latest/developerguide/administer-images.html}{Administer Your AppStream 2.0 Images},
in the \emph{Amazon AppStream 2.0 Administration Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_updated_image(
  existingImageName = "string",
  newImageName = "string",
  newImageDescription = "string",
  newImageDisplayName = "string",
  newImageTags = list(
    "string"
  ),
  dryRun = TRUE|FALSE
)
}
}

\keyword{internal}
