% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_app_block}
\alias{appstream_delete_app_block}
\title{Deletes an app block}
\usage{
appstream_delete_app_block(Name)
}
\arguments{
\item{Name}{[required] The name of the app block.}
}
\description{
Deletes an app block.

See \url{https://www.paws-r-sdk.com/docs/appstream_delete_app_block/} for full documentation.
}
\keyword{internal}
