% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_application}
\alias{appstream_delete_application}
\title{Deletes an application}
\usage{
appstream_delete_application(Name)
}
\arguments{
\item{Name}{[required] The name of the application.}
}
\description{
Deletes an application.

See \url{https://paws-r.github.io/docs/appstream/delete_application.html} for full documentation.
}
\keyword{internal}
