% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_entitlement}
\alias{appstream_delete_entitlement}
\title{Deletes the specified entitlement}
\usage{
appstream_delete_entitlement(Name, StackName)
}
\arguments{
\item{Name}{[required] The name of the entitlement.}

\item{StackName}{[required] The name of the stack with which the entitlement is associated.}
}
\description{
Deletes the specified entitlement.

See \url{https://paws-r.github.io/docs/appstream/delete_entitlement.html} for full documentation.
}
\keyword{internal}
