% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_image_builder}
\alias{appstream_delete_image_builder}
\title{Deletes the specified image builder and releases the capacity}
\usage{
appstream_delete_image_builder(Name)
}
\arguments{
\item{Name}{[required] The name of the image builder.}
}
\description{
Deletes the specified image builder and releases the capacity.

See \url{https://paws-r.github.io/docs/appstream/delete_image_builder.html} for full documentation.
}
\keyword{internal}
