% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_image_permissions}
\alias{appstream_delete_image_permissions}
\title{Deletes permissions for the specified private image}
\usage{
appstream_delete_image_permissions(Name, SharedAccountId)
}
\arguments{
\item{Name}{[required] The name of the private image.}

\item{SharedAccountId}{[required] The 12-digit identifier of the AWS account for which to delete image
permissions.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes permissions for the specified private image. After you delete
permissions for an image, AWS accounts to which you previously granted
these permissions can no longer use the image.
}
\section{Request syntax}{
\preformatted{svc$delete_image_permissions(
  Name = "string",
  SharedAccountId = "string"
)
}
}

\keyword{internal}
