% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_theme_for_stack}
\alias{appstream_delete_theme_for_stack}
\title{Deletes custom branding that customizes the appearance of the streaming
application catalog page}
\usage{
appstream_delete_theme_for_stack(StackName)
}
\arguments{
\item{StackName}{[required] The name of the stack for the theme.}
}
\description{
Deletes custom branding that customizes the appearance of the streaming application catalog page.

See \url{https://www.paws-r-sdk.com/docs/appstream_delete_theme_for_stack/} for full documentation.
}
\keyword{internal}
