% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_usage_report_subscription}
\alias{appstream_delete_usage_report_subscription}
\title{Disables usage report generation}
\usage{
appstream_delete_usage_report_subscription()
}
\description{
Disables usage report generation.

See \url{https://www.paws-r-sdk.com/docs/appstream_delete_usage_report_subscription/} for full documentation.
}
\keyword{internal}
