% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_user}
\alias{appstream_delete_user}
\title{Deletes a user from the user pool}
\usage{
appstream_delete_user(UserName, AuthenticationType)
}
\arguments{
\item{UserName}{[required] The email address of the user.

Users' email addresses are case-sensitive.}

\item{AuthenticationType}{[required] The authentication type for the user. You must specify USERPOOL.}
}
\description{
Deletes a user from the user pool.

See \url{https://www.paws-r-sdk.com/docs/appstream_delete_user/} for full documentation.
}
\keyword{internal}
