% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_describe_app_block_builder_app_block_associations}
\alias{appstream_describe_app_block_builder_app_block_associations}
\title{Retrieves a list that describes one or more app block builder
associations}
\usage{
appstream_describe_app_block_builder_app_block_associations(
  AppBlockArn = NULL,
  AppBlockBuilderName = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AppBlockArn}{The ARN of the app block.}

\item{AppBlockBuilderName}{The name of the app block builder.}

\item{MaxResults}{The maximum size of each page of results.}

\item{NextToken}{The pagination token used to retrieve the next page of results for this
operation.}
}
\description{
Retrieves a list that describes one or more app block builder associations.

See \url{https://www.paws-r-sdk.com/docs/appstream_describe_app_block_builder_app_block_associations/} for full documentation.
}
\keyword{internal}
