% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_describe_app_block_builders}
\alias{appstream_describe_app_block_builders}
\title{Retrieves a list that describes one or more app block builders}
\usage{
appstream_describe_app_block_builders(
  Names = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Names}{The names of the app block builders.}

\item{NextToken}{The pagination token used to retrieve the next page of results for this
operation.}

\item{MaxResults}{The maximum size of each page of results. The maximum value is 25.}
}
\description{
Retrieves a list that describes one or more app block builders.

See \url{https://www.paws-r-sdk.com/docs/appstream_describe_app_block_builders/} for full documentation.
}
\keyword{internal}
