% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_describe_app_blocks}
\alias{appstream_describe_app_blocks}
\title{Retrieves a list that describes one or more app blocks}
\usage{
appstream_describe_app_blocks(Arns = NULL, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{Arns}{The ARNs of the app blocks.}

\item{NextToken}{The pagination token used to retrieve the next page of results for this
operation.}

\item{MaxResults}{The maximum size of each page of results.}
}
\description{
Retrieves a list that describes one or more app blocks.

See \url{https://www.paws-r-sdk.com/docs/appstream_describe_app_blocks/} for full documentation.
}
\keyword{internal}
