% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_describe_app_license_usage}
\alias{appstream_describe_app_license_usage}
\title{Retrieves license included application usage information}
\usage{
appstream_describe_app_license_usage(
  BillingPeriod,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{BillingPeriod}{[required] Billing period for the usage record.

Specify the value in \emph{yyyy-mm} format. For example, for August 2025, use
\emph{2025-08}.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{Token for pagination of results.}
}
\description{
Retrieves license included application usage information.

See \url{https://www.paws-r-sdk.com/docs/appstream_describe_app_license_usage/} for full documentation.
}
\keyword{internal}
