% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_describe_application_fleet_associations}
\alias{appstream_describe_application_fleet_associations}
\title{Retrieves a list that describes one or more application fleet
associations}
\usage{
appstream_describe_application_fleet_associations(
  FleetName = NULL,
  ApplicationArn = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{FleetName}{The name of the fleet.}

\item{ApplicationArn}{The ARN of the application.}

\item{MaxResults}{The maximum size of each page of results.}

\item{NextToken}{The pagination token used to retrieve the next page of results for this
operation.}
}
\description{
Retrieves a list that describes one or more application fleet associations. Either ApplicationArn or FleetName must be specified.

See \url{https://www.paws-r-sdk.com/docs/appstream_describe_application_fleet_associations/} for full documentation.
}
\keyword{internal}
