% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_describe_applications}
\alias{appstream_describe_applications}
\title{Retrieves a list that describes one or more applications}
\usage{
appstream_describe_applications(
  Arns = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Arns}{The ARNs for the applications.}

\item{NextToken}{The pagination token used to retrieve the next page of results for this
operation.}

\item{MaxResults}{The maximum size of each page of results.}
}
\description{
Retrieves a list that describes one or more applications.

See \url{https://paws-r.github.io/docs/appstream/describe_applications.html} for full documentation.
}
\keyword{internal}
