% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_describe_entitlements}
\alias{appstream_describe_entitlements}
\title{Retrieves a list that describes one of more entitlements}
\usage{
appstream_describe_entitlements(
  Name = NULL,
  StackName,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Name}{The name of the entitlement.}

\item{StackName}{[required] The name of the stack with which the entitlement is associated.}

\item{NextToken}{The pagination token used to retrieve the next page of results for this
operation.}

\item{MaxResults}{The maximum size of each page of results.}
}
\description{
Retrieves a list that describes one of more entitlements.

See \url{https://www.paws-r-sdk.com/docs/appstream_describe_entitlements/} for full documentation.
}
\keyword{internal}
