% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_describe_image_builders}
\alias{appstream_describe_image_builders}
\title{Retrieves a list that describes one or more specified image builders, if
the image builder names are provided}
\usage{
appstream_describe_image_builders(
  Names = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Names}{The names of the image builders to describe.}

\item{MaxResults}{The maximum size of each page of results.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\description{
Retrieves a list that describes one or more specified image builders, if the image builder names are provided. Otherwise, all image builders in the account are described.

See \url{https://www.paws-r-sdk.com/docs/appstream_describe_image_builders/} for full documentation.
}
\keyword{internal}
