% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_describe_software_associations}
\alias{appstream_describe_software_associations}
\title{Retrieves license included application associations for a specified
resource}
\usage{
appstream_describe_software_associations(
  AssociatedResource,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AssociatedResource}{[required] The ARN of the resource to describe software associations. Possible
resources are Image and ImageBuilder.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation.}
}
\description{
Retrieves license included application associations for a specified resource.

See \url{https://www.paws-r-sdk.com/docs/appstream_describe_software_associations/} for full documentation.
}
\keyword{internal}
