% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_describe_theme_for_stack}
\alias{appstream_describe_theme_for_stack}
\title{Retrieves a list that describes the theme for a specified stack}
\usage{
appstream_describe_theme_for_stack(StackName)
}
\arguments{
\item{StackName}{[required] The name of the stack for the theme.}
}
\description{
Retrieves a list that describes the theme for a specified stack. A theme is custom branding that customizes the appearance of the streaming application catalog page.

See \url{https://www.paws-r-sdk.com/docs/appstream_describe_theme_for_stack/} for full documentation.
}
\keyword{internal}
