% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_describe_user_stack_associations}
\alias{appstream_describe_user_stack_associations}
\title{Retrieves a list that describes the UserStackAssociation objects}
\usage{
appstream_describe_user_stack_associations(
  StackName = NULL,
  UserName = NULL,
  AuthenticationType = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{StackName}{The name of the stack that is associated with the user.}

\item{UserName}{The email address of the user who is associated with the stack.

Users' email addresses are case-sensitive.}

\item{AuthenticationType}{The authentication type for the user who is associated with the stack.
You must specify USERPOOL.}

\item{MaxResults}{The maximum size of each page of results.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\description{
Retrieves a list that describes the UserStackAssociation objects. You must specify either or both of the following:

See \url{https://paws-r.github.io/docs/appstream/describe_user_stack_associations.html} for full documentation.
}
\keyword{internal}
