% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_disassociate_application_fleet}
\alias{appstream_disassociate_application_fleet}
\title{Disassociates the specified application from the fleet}
\usage{
appstream_disassociate_application_fleet(FleetName, ApplicationArn)
}
\arguments{
\item{FleetName}{[required] The name of the fleet.}

\item{ApplicationArn}{[required] The ARN of the application.}
}
\description{
Disassociates the specified application from the fleet.

See \url{https://paws-r.github.io/docs/appstream/disassociate_application_fleet.html} for full documentation.
}
\keyword{internal}
