% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_disassociate_software_from_image_builder}
\alias{appstream_disassociate_software_from_image_builder}
\title{Removes license included application(s) association(s) from an image
builder instance}
\usage{
appstream_disassociate_software_from_image_builder(
  ImageBuilderName,
  SoftwareNames
)
}
\arguments{
\item{ImageBuilderName}{[required] The name of the target image builder instance.}

\item{SoftwareNames}{[required] The list of license included applications to disassociate from the image
builder.

Possible values include the following:
\itemize{
\item Microsoft_Office_2021_LTSC_Professional_Plus_32Bit
\item Microsoft_Office_2021_LTSC_Professional_Plus_64Bit
\item Microsoft_Office_2024_LTSC_Professional_Plus_32Bit
\item Microsoft_Office_2024_LTSC_Professional_Plus_64Bit
\item Microsoft_Visio_2021_LTSC_Professional_32Bit
\item Microsoft_Visio_2021_LTSC_Professional_64Bit
\item Microsoft_Visio_2024_LTSC_Professional_32Bit
\item Microsoft_Visio_2024_LTSC_Professional_64Bit
\item Microsoft_Project_2021_Professional_32Bit
\item Microsoft_Project_2021_Professional_64Bit
\item Microsoft_Project_2024_Professional_32Bit
\item Microsoft_Project_2024_Professional_64Bit
\item Microsoft_Office_2021_LTSC_Standard_32Bit
\item Microsoft_Office_2021_LTSC_Standard_64Bit
\item Microsoft_Office_2024_LTSC_Standard_32Bit
\item Microsoft_Office_2024_LTSC_Standard_64Bit
\item Microsoft_Visio_2021_LTSC_Standard_32Bit
\item Microsoft_Visio_2021_LTSC_Standard_64Bit
\item Microsoft_Visio_2024_LTSC_Standard_32Bit
\item Microsoft_Visio_2024_LTSC_Standard_64Bit
\item Microsoft_Project_2021_Standard_32Bit
\item Microsoft_Project_2021_Standard_64Bit
\item Microsoft_Project_2024_Standard_32Bit
\item Microsoft_Project_2024_Standard_64Bit
}}
}
\description{
Removes license included application(s) association(s) from an image builder instance.

See \url{https://www.paws-r-sdk.com/docs/appstream_disassociate_software_from_image_builder/} for full documentation.
}
\keyword{internal}
