% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_drain_session_instance}
\alias{appstream_drain_session_instance}
\title{Drains the instance hosting the specified streaming session}
\usage{
appstream_drain_session_instance(SessionId)
}
\arguments{
\item{SessionId}{[required] The identifier of the streaming session.}
}
\description{
Drains the instance hosting the specified streaming session. The instance stops accepting new sessions while existing sessions continue uninterrupted. Once all sessions end, the instance is reclaimed and replaced. This only applies to multi-session fleets.

See \url{https://www.paws-r-sdk.com/docs/appstream_drain_session_instance/} for full documentation.
}
\keyword{internal}
