% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_enable_user}
\alias{appstream_enable_user}
\title{Enables a user in the user pool}
\usage{
appstream_enable_user(UserName, AuthenticationType)
}
\arguments{
\item{UserName}{[required] The email address of the user.

Users' email addresses are case-sensitive. During login, if they specify
an email address that doesn't use the same capitalization as the email
address specified when their user pool account was created, a "user does
not exist" error message displays.}

\item{AuthenticationType}{[required] The authentication type for the user. You must specify USERPOOL.}
}
\value{
An empty list.
}
\description{
Enables a user in the user pool. After being enabled, users can sign in
to AppStream 2.0 and open applications from the stacks to which they are
assigned.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$enable_user(
  UserName = "string",
  AuthenticationType = "API"|"SAML"|"USERPOOL"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
