% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_expire_session}
\alias{appstream_expire_session}
\title{Immediately stops the specified streaming session}
\usage{
appstream_expire_session(SessionId)
}
\arguments{
\item{SessionId}{[required] The identifier of the streaming session.}
}
\description{
Immediately stops the specified streaming session.

See \url{https://www.paws-r-sdk.com/docs/appstream_expire_session/} for full documentation.
}
\keyword{internal}
