% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_get_export_image_task}
\alias{appstream_get_export_image_task}
\title{Retrieves information about an export image task, including its current
state, progress, and any error details}
\usage{
appstream_get_export_image_task(TaskId = NULL)
}
\arguments{
\item{TaskId}{The unique identifier of the export image task to retrieve information
about.}
}
\description{
Retrieves information about an export image task, including its current state, progress, and any error details.

See \url{https://www.paws-r-sdk.com/docs/appstream_get_export_image_task/} for full documentation.
}
\keyword{internal}
