% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_list_associated_fleets}
\alias{appstream_list_associated_fleets}
\title{Retrieves the name of the fleet that is associated with the specified
stack}
\usage{
appstream_list_associated_fleets(StackName, NextToken)
}
\arguments{
\item{StackName}{[required] The name of the stack.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Names = list(
    "string"
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the name of the fleet that is associated with the specified
stack.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_associated_fleets(
  StackName = "string",
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
