% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_list_associated_stacks}
\alias{appstream_list_associated_stacks}
\title{Retrieves the name of the stack with which the specified fleet is
associated}
\usage{
appstream_list_associated_stacks(FleetName, NextToken)
}
\arguments{
\item{FleetName}{[required] The name of the fleet.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Names = list(
    "string"
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the name of the stack with which the specified fleet is
associated.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_associated_stacks(
  FleetName = "string",
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
