% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_list_entitled_applications}
\alias{appstream_list_entitled_applications}
\title{Retrieves a list of entitled applications}
\usage{
appstream_list_entitled_applications(
  StackName,
  EntitlementName,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{StackName}{[required] The name of the stack with which the entitlement is associated.}

\item{EntitlementName}{[required] The name of the entitlement.}

\item{NextToken}{The pagination token used to retrieve the next page of results for this
operation.}

\item{MaxResults}{The maximum size of each page of results.}
}
\description{
Retrieves a list of entitled applications.

See \url{https://www.paws-r-sdk.com/docs/appstream_list_entitled_applications/} for full documentation.
}
\keyword{internal}
