% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_start_app_block_builder}
\alias{appstream_start_app_block_builder}
\title{Starts an app block builder}
\usage{
appstream_start_app_block_builder(Name)
}
\arguments{
\item{Name}{[required] The name of the app block builder.}
}
\description{
Starts an app block builder.

See \url{https://www.paws-r-sdk.com/docs/appstream_start_app_block_builder/} for full documentation.
}
\keyword{internal}
