% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_start_image_builder}
\alias{appstream_start_image_builder}
\title{Starts the specified image builder}
\usage{
appstream_start_image_builder(Name, AppstreamAgentVersion = NULL)
}
\arguments{
\item{Name}{[required] The name of the image builder.}

\item{AppstreamAgentVersion}{The version of the WorkSpaces Applications agent to use for this image
builder. To use the latest version of the WorkSpaces Applications agent,
specify [LATEST].}
}
\description{
Starts the specified image builder.

See \url{https://www.paws-r-sdk.com/docs/appstream_start_image_builder/} for full documentation.
}
\keyword{internal}
