% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_start_software_deployment_to_image_builder}
\alias{appstream_start_software_deployment_to_image_builder}
\title{Initiates license included applications deployment to an image builder
instance}
\usage{
appstream_start_software_deployment_to_image_builder(
  ImageBuilderName,
  RetryFailedDeployments = NULL
)
}
\arguments{
\item{ImageBuilderName}{[required] The name of the target image builder instance.}

\item{RetryFailedDeployments}{Whether to retry previously failed license included application
deployments.}
}
\description{
Initiates license included applications deployment to an image builder instance.

See \url{https://www.paws-r-sdk.com/docs/appstream_start_software_deployment_to_image_builder/} for full documentation.
}
\keyword{internal}
