% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_stop_image_builder}
\alias{appstream_stop_image_builder}
\title{Stops the specified image builder}
\usage{
appstream_stop_image_builder(Name)
}
\arguments{
\item{Name}{[required] The name of the image builder.}
}
\description{
Stops the specified image builder.

See \url{https://www.paws-r-sdk.com/docs/appstream_stop_image_builder/} for full documentation.
}
\keyword{internal}
