% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_update_application}
\alias{appstream_update_application}
\title{Updates the specified application}
\usage{
appstream_update_application(
  Name,
  DisplayName = NULL,
  Description = NULL,
  IconS3Location = NULL,
  LaunchPath = NULL,
  WorkingDirectory = NULL,
  LaunchParameters = NULL,
  AppBlockArn = NULL,
  AttributesToDelete = NULL
)
}
\arguments{
\item{Name}{[required] The name of the application. This name is visible to users when display
name is not specified.}

\item{DisplayName}{The display name of the application. This name is visible to users in
the application catalog.}

\item{Description}{The description of the application.}

\item{IconS3Location}{The icon S3 location of the application.}

\item{LaunchPath}{The launch path of the application.}

\item{WorkingDirectory}{The working directory of the application.}

\item{LaunchParameters}{The launch parameters of the application.}

\item{AppBlockArn}{The ARN of the app block.}

\item{AttributesToDelete}{The attributes to delete for an application.}
}
\description{
Updates the specified application.

See \url{https://paws-r.github.io/docs/appstream/update_application.html} for full documentation.
}
\keyword{internal}
