% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_update_entitlement}
\alias{appstream_update_entitlement}
\title{Updates the specified entitlement}
\usage{
appstream_update_entitlement(
  Name,
  StackName,
  Description = NULL,
  AppVisibility = NULL,
  Attributes = NULL
)
}
\arguments{
\item{Name}{[required] The name of the entitlement.}

\item{StackName}{[required] The name of the stack with which the entitlement is associated.}

\item{Description}{The description of the entitlement.}

\item{AppVisibility}{Specifies whether all or only selected apps are entitled.}

\item{Attributes}{The attributes of the entitlement.}
}
\description{
Updates the specified entitlement.

See \url{https://www.paws-r-sdk.com/docs/appstream_update_entitlement/} for full documentation.
}
\keyword{internal}
