% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_associate_to_configuration}
\alias{chatbot_associate_to_configuration}
\title{Links a resource (for example, a custom action) to a channel
configuration}
\usage{
chatbot_associate_to_configuration(Resource, ChatConfiguration)
}
\arguments{
\item{Resource}{[required] The resource Amazon Resource Name (ARN) to link.}

\item{ChatConfiguration}{[required] The channel configuration to associate with the resource.}
}
\description{
Links a resource (for example, a custom action) to a channel configuration.

See \url{https://www.paws-r-sdk.com/docs/chatbot_associate_to_configuration/} for full documentation.
}
\keyword{internal}
