% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_delete_chime_webhook_configuration}
\alias{chatbot_delete_chime_webhook_configuration}
\title{Deletes a Amazon Chime webhook configuration for AWS Chatbot}
\usage{
chatbot_delete_chime_webhook_configuration(ChatConfigurationArn)
}
\arguments{
\item{ChatConfigurationArn}{[required] The Amazon Resource Name (ARN) of the ChimeWebhookConfiguration to
delete.}
}
\description{
Deletes a Amazon Chime webhook configuration for AWS Chatbot.

See \url{https://www.paws-r-sdk.com/docs/chatbot_delete_chime_webhook_configuration/} for full documentation.
}
\keyword{internal}
