% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_delete_microsoft_teams_channel_configuration}
\alias{chatbot_delete_microsoft_teams_channel_configuration}
\title{Deletes a Microsoft Teams channel configuration for AWS Chatbot}
\usage{
chatbot_delete_microsoft_teams_channel_configuration(ChatConfigurationArn)
}
\arguments{
\item{ChatConfigurationArn}{[required] The Amazon Resource Name (ARN) of the MicrosoftTeamsChannelConfiguration
associated with the user identity to delete.}
}
\description{
Deletes a Microsoft Teams channel configuration for AWS Chatbot

See \url{https://www.paws-r-sdk.com/docs/chatbot_delete_microsoft_teams_channel_configuration/} for full documentation.
}
\keyword{internal}
