% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_delete_microsoft_teams_user_identity}
\alias{chatbot_delete_microsoft_teams_user_identity}
\title{Identifes a user level permission for a channel configuration}
\usage{
chatbot_delete_microsoft_teams_user_identity(ChatConfigurationArn, UserId)
}
\arguments{
\item{ChatConfigurationArn}{[required] The ARN of the MicrosoftTeamsChannelConfiguration associated with the
user identity to delete.}

\item{UserId}{[required] The Microsoft Teams user ID.}
}
\description{
Identifes a user level permission for a channel configuration.

See \url{https://www.paws-r-sdk.com/docs/chatbot_delete_microsoft_teams_user_identity/} for full documentation.
}
\keyword{internal}
