% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_delete_slack_user_identity}
\alias{chatbot_delete_slack_user_identity}
\title{Deletes a user level permission for a Slack channel configuration}
\usage{
chatbot_delete_slack_user_identity(
  ChatConfigurationArn,
  SlackTeamId,
  SlackUserId
)
}
\arguments{
\item{ChatConfigurationArn}{[required] The ARN of the SlackChannelConfiguration associated with the user
identity to delete.}

\item{SlackTeamId}{[required] The ID of the Slack workspace authorized with AWS Chatbot.}

\item{SlackUserId}{[required] The ID of the user in Slack}
}
\description{
Deletes a user level permission for a Slack channel configuration.

See \url{https://www.paws-r-sdk.com/docs/chatbot_delete_slack_user_identity/} for full documentation.
}
\keyword{internal}
