% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_describe_chime_webhook_configurations}
\alias{chatbot_describe_chime_webhook_configurations}
\title{Lists Amazon Chime webhook configurations optionally filtered by
ChatConfigurationArn}
\usage{
chatbot_describe_chime_webhook_configurations(
  MaxResults = NULL,
  NextToken = NULL,
  ChatConfigurationArn = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified MaxResults value, a token is included
in the response so that the remaining results can be retrieved.}

\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by MaxResults.}

\item{ChatConfigurationArn}{An optional Amazon Resource Name (ARN) of a ChimeWebhookConfiguration to
describe.}
}
\description{
Lists Amazon Chime webhook configurations optionally filtered by ChatConfigurationArn

See \url{https://www.paws-r-sdk.com/docs/chatbot_describe_chime_webhook_configurations/} for full documentation.
}
\keyword{internal}
