% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_describe_slack_workspaces}
\alias{chatbot_describe_slack_workspaces}
\title{List all authorized Slack workspaces connected to the AWS Account
onboarded with AWS Chatbot}
\usage{
chatbot_describe_slack_workspaces(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified MaxResults value, a token is included
in the response so that the remaining results can be retrieved.}

\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by MaxResults.}
}
\description{
List all authorized Slack workspaces connected to the AWS Account onboarded with AWS Chatbot.

See \url{https://www.paws-r-sdk.com/docs/chatbot_describe_slack_workspaces/} for full documentation.
}
\keyword{internal}
