% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_disassociate_from_configuration}
\alias{chatbot_disassociate_from_configuration}
\title{Unlink a resource, for example a custom action, from a channel
configuration}
\usage{
chatbot_disassociate_from_configuration(Resource, ChatConfiguration)
}
\arguments{
\item{Resource}{[required] The resource (for example, a custom action) Amazon Resource Name (ARN)
to unlink.}

\item{ChatConfiguration}{[required] The channel configuration the resource is being disassociated from.}
}
\description{
Unlink a resource, for example a custom action, from a channel configuration.

See \url{https://www.paws-r-sdk.com/docs/chatbot_disassociate_from_configuration/} for full documentation.
}
\keyword{internal}
