% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_get_account_preferences}
\alias{chatbot_get_account_preferences}
\title{Returns AWS Chatbot account preferences}
\usage{
chatbot_get_account_preferences()
}
\description{
Returns AWS Chatbot account preferences.

See \url{https://www.paws-r-sdk.com/docs/chatbot_get_account_preferences/} for full documentation.
}
\keyword{internal}
