% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_get_microsoft_teams_channel_configuration}
\alias{chatbot_get_microsoft_teams_channel_configuration}
\title{Returns a Microsoft Teams channel configuration in an AWS account}
\usage{
chatbot_get_microsoft_teams_channel_configuration(ChatConfigurationArn)
}
\arguments{
\item{ChatConfigurationArn}{[required] The Amazon Resource Name (ARN) of the MicrosoftTeamsChannelConfiguration
to retrieve.}
}
\description{
Returns a Microsoft Teams channel configuration in an AWS account.

See \url{https://www.paws-r-sdk.com/docs/chatbot_get_microsoft_teams_channel_configuration/} for full documentation.
}
\keyword{internal}
