% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_list_associations}
\alias{chatbot_list_associations}
\title{Lists resources associated with a channel configuration}
\usage{
chatbot_list_associations(
  ChatConfiguration,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ChatConfiguration}{[required] The channel configuration to list associations for.}

\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified MaxResults value, a token is included
in the response so that the remaining results can be retrieved.}

\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by MaxResults.}
}
\description{
Lists resources associated with a channel configuration.

See \url{https://www.paws-r-sdk.com/docs/chatbot_list_associations/} for full documentation.
}
\keyword{internal}
