% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_list_microsoft_teams_channel_configurations}
\alias{chatbot_list_microsoft_teams_channel_configurations}
\title{Lists all AWS Chatbot Microsoft Teams channel configurations in an AWS
account}
\usage{
chatbot_list_microsoft_teams_channel_configurations(
  MaxResults = NULL,
  NextToken = NULL,
  TeamId = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified MaxResults value, a token is included
in the response so that the remaining results can be retrieved.}

\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by MaxResults.}

\item{TeamId}{The ID of the Microsoft Teams authorized with AWS Chatbot.

To get the team ID, you must perform the initial authorization flow with
Microsoft Teams in the AWS Chatbot console. Then you can copy and paste
the team ID from the console. For more information, see \href{https://docs.aws.amazon.com/chatbot/latest/adminguide/teams-setup.html#teams-client-setup}{Step 1: Configure a Microsoft Teams client}
in the \emph{AWS Chatbot Administrator Guide}.}
}
\description{
Lists all AWS Chatbot Microsoft Teams channel configurations in an AWS account.

See \url{https://www.paws-r-sdk.com/docs/chatbot_list_microsoft_teams_channel_configurations/} for full documentation.
}
\keyword{internal}
