% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_list_microsoft_teams_configured_teams}
\alias{chatbot_list_microsoft_teams_configured_teams}
\title{Lists all authorized Microsoft Teams for an AWS Account}
\usage{
chatbot_list_microsoft_teams_configured_teams(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified MaxResults value, a token is included
in the response so that the remaining results can be retrieved.}

\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by MaxResults.}
}
\description{
Lists all authorized Microsoft Teams for an AWS Account

See \url{https://www.paws-r-sdk.com/docs/chatbot_list_microsoft_teams_configured_teams/} for full documentation.
}
\keyword{internal}
