% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_list_microsoft_teams_user_identities}
\alias{chatbot_list_microsoft_teams_user_identities}
\title{A list all Microsoft Teams user identities with a mapped role}
\usage{
chatbot_list_microsoft_teams_user_identities(
  ChatConfigurationArn = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ChatConfigurationArn}{The Amazon Resource Name (ARN) of the MicrosoftTeamsChannelConfiguration
associated with the user identities to list.}

\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by MaxResults.}

\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified MaxResults value, a token is included
in the response so that the remaining results can be retrieved.}
}
\description{
A list all Microsoft Teams user identities with a mapped role.

See \url{https://www.paws-r-sdk.com/docs/chatbot_list_microsoft_teams_user_identities/} for full documentation.
}
\keyword{internal}
