% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_list_tags_for_resource}
\alias{chatbot_list_tags_for_resource}
\title{Lists all of the tags associated with the Amazon Resource Name (ARN)
that you specify}
\usage{
chatbot_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the resource to list tags for.}
}
\description{
Lists all of the tags associated with the Amazon Resource Name (ARN) that you specify. The resource can be a user, server, or role.

See \url{https://www.paws-r-sdk.com/docs/chatbot_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
