% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_tag_resource}
\alias{chatbot_tag_resource}
\title{Attaches a key-value pair to a resource, as identified by its Amazon
Resource Name (ARN)}
\usage{
chatbot_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the configuration.}

\item{Tags}{[required] A list of tags to apply to the configuration.}
}
\description{
Attaches a key-value pair to a resource, as identified by its Amazon Resource Name (ARN). Resources are users, servers, roles, and other entities.

See \url{https://www.paws-r-sdk.com/docs/chatbot_tag_resource/} for full documentation.
}
\keyword{internal}
