% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_untag_resource}
\alias{chatbot_untag_resource}
\title{Detaches a key-value pair from a resource, as identified by its Amazon
Resource Name (ARN)}
\usage{
chatbot_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The value of the resource that will have the tag removed. An Amazon
Resource Name (ARN) is an identifier for a specific AWS resource, such
as a server, user, or role.}

\item{TagKeys}{[required] TagKeys are key-value pairs assigned to ARNs that can be used to group
and search for resources by type. This metadata can be attached to
resources for any purpose.}
}
\description{
Detaches a key-value pair from a resource, as identified by its Amazon Resource Name (ARN). Resources are users, servers, roles, and other entities.

See \url{https://www.paws-r-sdk.com/docs/chatbot_untag_resource/} for full documentation.
}
\keyword{internal}
