% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_update_account_preferences}
\alias{chatbot_update_account_preferences}
\title{Updates AWS Chatbot account preferences}
\usage{
chatbot_update_account_preferences(
  UserAuthorizationRequired = NULL,
  TrainingDataCollectionEnabled = NULL
)
}
\arguments{
\item{UserAuthorizationRequired}{Enables use of a user role requirement in your chat configuration.}

\item{TrainingDataCollectionEnabled}{Turns on training data collection.

This helps improve the AWS Chatbot experience by allowing AWS Chatbot to
store and use your customer information, such as AWS Chatbot
configurations, notifications, user inputs, AWS Chatbot generated
responses, and interaction data. This data helps us to continuously
improve and develop Artificial Intelligence (AI) technologies. Your data
is not shared with any third parties and is protected using
sophisticated controls to prevent unauthorized access and misuse. AWS
Chatbot does not store or use interactions in chat channels with Amazon
Q for training AI technologies for AWS Chatbot.}
}
\description{
Updates AWS Chatbot account preferences.

See \url{https://www.paws-r-sdk.com/docs/chatbot_update_account_preferences/} for full documentation.
}
\keyword{internal}
