% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot_operations.R
\name{chatbot_update_custom_action}
\alias{chatbot_update_custom_action}
\title{Updates a custom action}
\usage{
chatbot_update_custom_action(
  CustomActionArn,
  Definition,
  AliasName = NULL,
  Attachments = NULL
)
}
\arguments{
\item{CustomActionArn}{[required] The fully defined Amazon Resource Name (ARN) of the custom action.}

\item{Definition}{[required] The definition of the command to run when invoked as an alias or as an
action button.}

\item{AliasName}{The name used to invoke this action in the chat channel. For example,
\verb{@aws run my-alias}.}

\item{Attachments}{Defines when this custom action button should be attached to a
notification.}
}
\description{
Updates a custom action.

See \url{https://www.paws-r-sdk.com/docs/chatbot_update_custom_action/} for full documentation.
}
\keyword{internal}
