% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivschat_operations.R
\name{ivschat_get_logging_configuration}
\alias{ivschat_get_logging_configuration}
\title{Gets the specified logging configuration}
\usage{
ivschat_get_logging_configuration(identifier)
}
\arguments{
\item{identifier}{[required] Identifier of the logging configuration to be retrieved.}
}
\description{
Gets the specified logging configuration.

See \url{https://paws-r.github.io/docs/ivschat/get_logging_configuration.html} for full documentation.
}
\keyword{internal}
