% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivschat_operations.R
\name{ivschat_get_room}
\alias{ivschat_get_room}
\title{Gets the specified room}
\usage{
ivschat_get_room(identifier)
}
\arguments{
\item{identifier}{[required] Identifier of the room for which the configuration is to be retrieved.
Currently this must be an ARN.}
}
\description{
Gets the specified room.

See \url{https://www.paws-r-sdk.com/docs/ivschat_get_room/} for full documentation.
}
\keyword{internal}
