% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivschat_operations.R
\name{ivschat_list_rooms}
\alias{ivschat_list_rooms}
\title{Gets summary information about all your rooms in the AWS region where
the API request is processed}
\usage{
ivschat_list_rooms(
  loggingConfigurationIdentifier = NULL,
  maxResults = NULL,
  messageReviewHandlerUri = NULL,
  name = NULL,
  nextToken = NULL
)
}
\arguments{
\item{loggingConfigurationIdentifier}{Logging-configuration identifier.}

\item{maxResults}{Maximum number of rooms to return. Default: 50.}

\item{messageReviewHandlerUri}{Filters the list to match the specified message review handler URI.}

\item{name}{Filters the list to match the specified room name.}

\item{nextToken}{The first room to retrieve. This is used for pagination; see the
\code{nextToken} response field.}
}
\description{
Gets summary information about all your rooms in the AWS region where the API request is processed. Results are sorted in descending order of \code{updateTime}.

See \url{https://paws-r.github.io/docs/ivschat/list_rooms.html} for full documentation.
}
\keyword{internal}
