% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivschat_operations.R
\name{ivschat_list_tags_for_resource}
\alias{ivschat_list_tags_for_resource}
\title{Gets information about AWS tags for the specified ARN}
\usage{
ivschat_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to be retrieved. The ARN must be URL-encoded.}
}
\description{
Gets information about AWS tags for the specified ARN.

See \url{https://paws-r.github.io/docs/ivschat/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
