% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivschat_operations.R
\name{ivschat_send_event}
\alias{ivschat_send_event}
\title{Sends an event to a room}
\usage{
ivschat_send_event(attributes = NULL, eventName, roomIdentifier)
}
\arguments{
\item{attributes}{Application-defined metadata to attach to the event sent to clients. The
maximum length of the metadata is 1 KB total.}

\item{eventName}{[required] Application-defined name of the event to send to clients.}

\item{roomIdentifier}{[required] Identifier of the room to which the event will be sent. Currently this
must be an ARN.}
}
\description{
Sends an event to a room. Use this within your application’s business logic to send events to clients of a room; e.g., to notify clients to change the way the chat UI is rendered.

See \url{https://www.paws-r-sdk.com/docs/ivschat_send_event/} for full documentation.
}
\keyword{internal}
