% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivschat_operations.R
\name{ivschat_tag_resource}
\alias{ivschat_tag_resource}
\title{Adds or updates tags for the AWS resource with the specified ARN}
\usage{
ivschat_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to be tagged. The ARN must be URL-encoded.}

\item{tags}{[required] Array of tags to be added or updated.}
}
\description{
Adds or updates tags for the AWS resource with the specified ARN.

See \url{https://paws-r.github.io/docs/ivschat/tag_resource.html} for full documentation.
}
\keyword{internal}
