% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivschat_operations.R
\name{ivschat_untag_resource}
\alias{ivschat_untag_resource}
\title{Removes tags from the resource with the specified ARN}
\usage{
ivschat_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to be untagged. The ARN must be URL-encoded.}

\item{tagKeys}{[required] Array of tags to be removed.}
}
\description{
Removes tags from the resource with the specified ARN.

See \url{https://paws-r.github.io/docs/ivschat/untag_resource.html} for full documentation.
}
\keyword{internal}
