% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivschat_operations.R
\name{ivschat_update_logging_configuration}
\alias{ivschat_update_logging_configuration}
\title{Updates a specified logging configuration}
\usage{
ivschat_update_logging_configuration(
  identifier,
  name = NULL,
  destinationConfiguration = NULL
)
}
\arguments{
\item{identifier}{[required] Identifier of the logging configuration to be updated.}

\item{name}{Logging-configuration name. The value does not need to be unique.}

\item{destinationConfiguration}{A complex type that contains a destination configuration for where chat
content will be logged. There can be only one type of destination
(\code{cloudWatchLogs}, \code{firehose}, or \code{s3}) in a \code{destinationConfiguration}.}
}
\description{
Updates a specified logging configuration.

See \url{https://www.paws-r-sdk.com/docs/ivschat_update_logging_configuration/} for full documentation.
}
\keyword{internal}
