% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_accept_eulas}
\alias{nimblestudio_accept_eulas}
\title{Accept EULAs}
\usage{
nimblestudio_accept_eulas(clientToken = NULL, eulaIds = NULL, studioId)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don’t specify a client token, the
Amazon Web Services SDK automatically generates a client token and uses
it for the request to ensure idempotency.}

\item{eulaIds}{The EULA ID.}

\item{studioId}{[required] The studio ID.}
}
\description{
Accept EULAs.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_accept_eulas/} for full documentation.
}
\keyword{internal}
