% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_create_streaming_image}
\alias{nimblestudio_create_streaming_image}
\title{Creates a streaming image resource in a studio}
\usage{
nimblestudio_create_streaming_image(
  clientToken = NULL,
  description = NULL,
  ec2ImageId,
  name,
  studioId,
  tags = NULL
)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don’t specify a client token, the
Amazon Web Services SDK automatically generates a client token and uses
it for the request to ensure idempotency.}

\item{description}{A human-readable description of the streaming image.}

\item{ec2ImageId}{[required] The ID of an EC2 machine image with which to create this streaming
image.}

\item{name}{[required] A friendly name for a streaming image resource.}

\item{studioId}{[required] The studio ID.}

\item{tags}{A collection of labels, in the form of key-value pairs, that apply to
this resource.}
}
\description{
Creates a streaming image resource in a studio.

See \url{https://paws-r.github.io/docs/nimblestudio/create_streaming_image.html} for full documentation.
}
\keyword{internal}
