% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_delete_launch_profile_member}
\alias{nimblestudio_delete_launch_profile_member}
\title{Delete a user from launch profile membership}
\usage{
nimblestudio_delete_launch_profile_member(
  clientToken = NULL,
  launchProfileId,
  principalId,
  studioId
)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don’t specify a client token, the AWS
SDK automatically generates a client token and uses it for the request
to ensure idempotency.}

\item{launchProfileId}{[required] The Launch Profile ID.}

\item{principalId}{[required] The principal ID. This currently supports a Amazon Web Services SSO
UserId.}

\item{studioId}{[required] The studio ID.}
}
\description{
Delete a user from launch profile membership.

See \url{https://paws-r.github.io/docs/nimblestudio/delete_launch_profile_member.html} for full documentation.
}
\keyword{internal}
