% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_delete_streaming_session}
\alias{nimblestudio_delete_streaming_session}
\title{Deletes streaming session resource}
\usage{
nimblestudio_delete_streaming_session(clientToken = NULL, sessionId, studioId)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don’t specify a client token, the
Amazon Web Services SDK automatically generates a client token and uses
it for the request to ensure idempotency.}

\item{sessionId}{[required] The streaming session ID.}

\item{studioId}{[required] The studio ID.}
}
\description{
Deletes streaming session resource.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_delete_streaming_session/} for full documentation.
}
\keyword{internal}
