% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_delete_studio}
\alias{nimblestudio_delete_studio}
\title{Delete a studio resource}
\usage{
nimblestudio_delete_studio(clientToken = NULL, studioId)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don’t specify a client token, the
Amazon Web Services SDK automatically generates a client token and uses
it for the request to ensure idempotency.}

\item{studioId}{[required] The studio ID.}
}
\description{
Delete a studio resource.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_delete_studio/} for full documentation.
}
\keyword{internal}
