% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_delete_studio_member}
\alias{nimblestudio_delete_studio_member}
\title{Delete a user from studio membership}
\usage{
nimblestudio_delete_studio_member(clientToken = NULL, principalId, studioId)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don’t specify a client token, the
Amazon Web Services SDK automatically generates a client token and uses
it for the request to ensure idempotency.}

\item{principalId}{[required] The principal ID. This currently supports a IAM Identity Center UserId.}

\item{studioId}{[required] The studio ID.}
}
\description{
Delete a user from studio membership.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_delete_studio_member/} for full documentation.
}
\keyword{internal}
