% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_eula}
\alias{nimblestudio_get_eula}
\title{Get EULA}
\usage{
nimblestudio_get_eula(eulaId)
}
\arguments{
\item{eulaId}{[required] The EULA ID.}
}
\description{
Get EULA.

See \url{https://paws-r.github.io/docs/nimblestudio/get_eula.html} for full documentation.
}
\keyword{internal}
