% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_launch_profile}
\alias{nimblestudio_get_launch_profile}
\title{Get a launch profile}
\usage{
nimblestudio_get_launch_profile(launchProfileId, studioId)
}
\arguments{
\item{launchProfileId}{[required] The ID of the launch profile used to control access from the streaming
session.}

\item{studioId}{[required] The studio ID.}
}
\description{
Get a launch profile.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_get_launch_profile/} for full documentation.
}
\keyword{internal}
