% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_launch_profile_details}
\alias{nimblestudio_get_launch_profile_details}
\title{Launch profile details include the launch profile resource and summary
information of resources that are used by, or available to, the launch
profile}
\usage{
nimblestudio_get_launch_profile_details(launchProfileId, studioId)
}
\arguments{
\item{launchProfileId}{[required] The ID of the launch profile used to control access from the streaming
session.}

\item{studioId}{[required] The studio ID.}
}
\description{
Launch profile details include the launch profile resource and summary information of resources that are used by, or available to, the launch profile. This includes the name and description of all studio components used by the launch profiles, and the name and description of streaming images that can be used with this launch profile.

See \url{https://paws-r.github.io/docs/nimblestudio/get_launch_profile_details.html} for full documentation.
}
\keyword{internal}
