% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_launch_profile_member}
\alias{nimblestudio_get_launch_profile_member}
\title{Get a user persona in launch profile membership}
\usage{
nimblestudio_get_launch_profile_member(launchProfileId, principalId, studioId)
}
\arguments{
\item{launchProfileId}{[required] The Launch Profile ID.}

\item{principalId}{[required] The principal ID. This currently supports a Amazon Web Services SSO
UserId.}

\item{studioId}{[required] The studio ID.}
}
\description{
Get a user persona in launch profile membership.

See \url{https://paws-r.github.io/docs/nimblestudio/get_launch_profile_member.html} for full documentation.
}
\keyword{internal}
