% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_streaming_session}
\alias{nimblestudio_get_streaming_session}
\title{Gets StreamingSession resource}
\usage{
nimblestudio_get_streaming_session(sessionId, studioId)
}
\arguments{
\item{sessionId}{[required] The streaming session ID.}

\item{studioId}{[required] The studio ID.}
}
\description{
Gets StreamingSession resource.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_get_streaming_session/} for full documentation.
}
\keyword{internal}
