% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_streaming_session_backup}
\alias{nimblestudio_get_streaming_session_backup}
\title{Gets StreamingSessionBackup resource}
\usage{
nimblestudio_get_streaming_session_backup(backupId, studioId)
}
\arguments{
\item{backupId}{[required] The ID of the backup.}

\item{studioId}{[required] The studio ID.}
}
\description{
Gets \code{StreamingSessionBackup} resource.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_get_streaming_session_backup/} for full documentation.
}
\keyword{internal}
