% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_streaming_session_stream}
\alias{nimblestudio_get_streaming_session_stream}
\title{Gets a StreamingSessionStream for a streaming session}
\usage{
nimblestudio_get_streaming_session_stream(sessionId, streamId, studioId)
}
\arguments{
\item{sessionId}{[required] The streaming session ID.}

\item{streamId}{[required] The streaming session stream ID.}

\item{studioId}{[required] The studio ID.}
}
\description{
Gets a StreamingSessionStream for a streaming session.

See \url{https://paws-r.github.io/docs/nimblestudio/get_streaming_session_stream.html} for full documentation.
}
\keyword{internal}
