% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_studio}
\alias{nimblestudio_get_studio}
\title{Get a Studio resource}
\usage{
nimblestudio_get_studio(studioId)
}
\arguments{
\item{studioId}{[required] The studio ID.}
}
\description{
Get a Studio resource.

See \url{https://paws-r.github.io/docs/nimblestudio/get_studio.html} for full documentation.
}
\keyword{internal}
