% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_studio_component}
\alias{nimblestudio_get_studio_component}
\title{Gets a studio component resource}
\usage{
nimblestudio_get_studio_component(studioComponentId, studioId)
}
\arguments{
\item{studioComponentId}{[required] The studio component ID.}

\item{studioId}{[required] The studio ID.}
}
\description{
Gets a studio component resource.

See \url{https://paws-r.github.io/docs/nimblestudio/get_studio_component.html} for full documentation.
}
\keyword{internal}
