% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_studio_member}
\alias{nimblestudio_get_studio_member}
\title{Get a user's membership in a studio}
\usage{
nimblestudio_get_studio_member(principalId, studioId)
}
\arguments{
\item{principalId}{[required] The principal ID. This currently supports a IAM Identity Center UserId.}

\item{studioId}{[required] The studio ID.}
}
\description{
Get a user's membership in a studio.

See \url{https://paws-r.github.io/docs/nimblestudio/get_studio_member.html} for full documentation.
}
\keyword{internal}
