% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_list_launch_profile_members}
\alias{nimblestudio_list_launch_profile_members}
\title{Get all users in a given launch profile membership}
\usage{
nimblestudio_list_launch_profile_members(
  launchProfileId,
  maxResults = NULL,
  nextToken = NULL,
  studioId
)
}
\arguments{
\item{launchProfileId}{[required] The Launch Profile ID.}

\item{maxResults}{The max number of results to return in the response.}

\item{nextToken}{The token to request the next page of results.}

\item{studioId}{[required] The studio ID.}
}
\description{
Get all users in a given launch profile membership.

See \url{https://paws-r.github.io/docs/nimblestudio/list_launch_profile_members.html} for full documentation.
}
\keyword{internal}
