% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_list_launch_profiles}
\alias{nimblestudio_list_launch_profiles}
\title{List all the launch profiles a studio}
\usage{
nimblestudio_list_launch_profiles(
  maxResults = NULL,
  nextToken = NULL,
  principalId = NULL,
  states = NULL,
  studioId
)
}
\arguments{
\item{maxResults}{The max number of results to return in the response.}

\item{nextToken}{The token to request the next page of results.}

\item{principalId}{The principal ID. This currently supports a Amazon Web Services SSO
UserId.}

\item{states}{Filter this request to launch profiles in any of the given states.}

\item{studioId}{[required] The studio ID.}
}
\description{
List all the launch profiles a studio.

See \url{https://paws-r.github.io/docs/nimblestudio/list_launch_profiles.html} for full documentation.
}
\keyword{internal}
