% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_list_streaming_session_backups}
\alias{nimblestudio_list_streaming_session_backups}
\title{Lists the backups of a streaming session in a studio}
\usage{
nimblestudio_list_streaming_session_backups(
  nextToken = NULL,
  ownedBy = NULL,
  studioId
)
}
\arguments{
\item{nextToken}{The token for the next set of results, or null if there are no more
results.}

\item{ownedBy}{The user ID of the user that owns the streaming session.}

\item{studioId}{[required] The studio ID.}
}
\description{
Lists the backups of a streaming session in a studio.

See \url{https://paws-r.github.io/docs/nimblestudio/list_streaming_session_backups.html} for full documentation.
}
\keyword{internal}
