% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_list_streaming_sessions}
\alias{nimblestudio_list_streaming_sessions}
\title{Lists the streaming sessions in a studio}
\usage{
nimblestudio_list_streaming_sessions(
  createdBy = NULL,
  nextToken = NULL,
  ownedBy = NULL,
  sessionIds = NULL,
  studioId
)
}
\arguments{
\item{createdBy}{Filters the request to streaming sessions created by the given user.}

\item{nextToken}{The token for the next set of results, or null if there are no more
results.}

\item{ownedBy}{Filters the request to streaming session owned by the given user}

\item{sessionIds}{Filters the request to only the provided session IDs.}

\item{studioId}{[required] The studio ID.}
}
\description{
Lists the streaming sessions in a studio.

See \url{https://paws-r.github.io/docs/nimblestudio/list_streaming_sessions.html} for full documentation.
}
\keyword{internal}
