% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_list_studio_components}
\alias{nimblestudio_list_studio_components}
\title{Lists the StudioComponents in a studio}
\usage{
nimblestudio_list_studio_components(
  maxResults = NULL,
  nextToken = NULL,
  states = NULL,
  studioId,
  types = NULL
)
}
\arguments{
\item{maxResults}{The max number of results to return in the response.}

\item{nextToken}{The token for the next set of results, or null if there are no more
results.}

\item{states}{Filters the request to studio components that are in one of the given
states.}

\item{studioId}{[required] The studio ID.}

\item{types}{Filters the request to studio components that are of one of the given
types.}
}
\description{
Lists the \code{StudioComponents} in a studio.

See \url{https://paws-r.github.io/docs/nimblestudio/list_studio_components.html} for full documentation.
}
\keyword{internal}
