% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_list_studio_members}
\alias{nimblestudio_list_studio_members}
\title{Get all users in a given studio membership}
\usage{
nimblestudio_list_studio_members(maxResults = NULL, nextToken = NULL, studioId)
}
\arguments{
\item{maxResults}{The max number of results to return in the response.}

\item{nextToken}{The token for the next set of results, or null if there are no more
results.}

\item{studioId}{[required] The studio ID.}
}
\description{
Get all users in a given studio membership.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_list_studio_members/} for full documentation.
}
\keyword{internal}
