% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_put_studio_members}
\alias{nimblestudio_put_studio_members}
\title{Add/update users with given persona to studio membership}
\usage{
nimblestudio_put_studio_members(
  clientToken = NULL,
  identityStoreId,
  members,
  studioId
)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don’t specify a client token, the
Amazon Web Services SDK automatically generates a client token and uses
it for the request to ensure idempotency.}

\item{identityStoreId}{[required] The ID of the identity store.}

\item{members}{[required] A list of members.}

\item{studioId}{[required] The studio ID.}
}
\description{
Add/update users with given persona to studio membership.

See \url{https://paws-r.github.io/docs/nimblestudio/put_studio_members.html} for full documentation.
}
\keyword{internal}
