% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_tag_resource}
\alias{nimblestudio_tag_resource}
\title{Creates tags for a resource, given its ARN}
\usage{
nimblestudio_tag_resource(resourceArn, tags = NULL)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource you want to add tags to.}

\item{tags}{A collection of labels, in the form of key:value pairs, that apply to
this resource.}
}
\description{
Creates tags for a resource, given its ARN.

See \url{https://paws-r.github.io/docs/nimblestudio/tag_resource.html} for full documentation.
}
\keyword{internal}
