% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_untag_resource}
\alias{nimblestudio_untag_resource}
\title{Deletes the tags for a resource}
\usage{
nimblestudio_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] Identifies the Amazon Resource Name(ARN) key from which you are removing
tags.}

\item{tagKeys}{[required] One or more tag keys. Specify only the tag keys, not the tag values.}
}
\description{
Deletes the tags for a resource.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_untag_resource/} for full documentation.
}
\keyword{internal}
