% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_update_launch_profile}
\alias{nimblestudio_update_launch_profile}
\title{Update a launch profile}
\usage{
nimblestudio_update_launch_profile(
  clientToken = NULL,
  description = NULL,
  launchProfileId,
  launchProfileProtocolVersions = NULL,
  name = NULL,
  streamConfiguration = NULL,
  studioComponentIds = NULL,
  studioId
)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don’t specify a client token, the
Amazon Web Services SDK automatically generates a client token and uses
it for the request to ensure idempotency.}

\item{description}{The description.}

\item{launchProfileId}{[required] The ID of the launch profile used to control access from the streaming
session.}

\item{launchProfileProtocolVersions}{The version number of the protocol that is used by the launch profile.
The only valid version is "2021-03-31".}

\item{name}{The name for the launch profile.}

\item{streamConfiguration}{A configuration for a streaming session.}

\item{studioComponentIds}{Unique identifiers for a collection of studio components that can be
used with this launch profile.}

\item{studioId}{[required] The studio ID.}
}
\description{
Update a launch profile.

See \url{https://paws-r.github.io/docs/nimblestudio/update_launch_profile.html} for full documentation.
}
\keyword{internal}
