% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_update_studio}
\alias{nimblestudio_update_studio}
\title{Update a Studio resource}
\usage{
nimblestudio_update_studio(
  adminRoleArn = NULL,
  clientToken = NULL,
  displayName = NULL,
  studioId,
  userRoleArn = NULL
)
}
\arguments{
\item{adminRoleArn}{The IAM role that Studio Admins will assume when logging in to the
Nimble Studio portal.}

\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don’t specify a client token, the
Amazon Web Services SDK automatically generates a client token and uses
it for the request to ensure idempotency.}

\item{displayName}{A friendly name for the studio.}

\item{studioId}{[required] The studio ID.}

\item{userRoleArn}{The IAM role that Studio Users will assume when logging in to the Nimble
Studio portal.}
}
\description{
Update a Studio resource.

See \url{https://paws-r.github.io/docs/nimblestudio/update_studio.html} for full documentation.
}
\keyword{internal}
