% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_abort_document_version_upload}
\alias{workdocs_abort_document_version_upload}
\title{Aborts the upload of the specified document version that was previously
initiated by InitiateDocumentVersionUpload}
\usage{
workdocs_abort_document_version_upload(AuthenticationToken, DocumentId,
  VersionId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{DocumentId}{[required] The ID of the document.}

\item{VersionId}{[required] The ID of the version.}
}
\value{
An empty list.
}
\description{
Aborts the upload of the specified document version that was previously
initiated by
\code{\link[=workdocs_initiate_document_version_upload]{initiate_document_version_upload}}.
The client should make this call only when it no longer intends to
upload the document version, or fails to do so.
}
\section{Request syntax}{
\preformatted{svc$abort_document_version_upload(
  AuthenticationToken = "string",
  DocumentId = "string",
  VersionId = "string"
)
}
}

\keyword{internal}
