% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_activate_user}
\alias{workdocs_activate_user}
\title{Activates the specified user}
\usage{
workdocs_activate_user(UserId, AuthenticationToken = NULL)
}
\arguments{
\item{UserId}{[required] The ID of the user.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}
}
\description{
Activates the specified user. Only active users can access Amazon WorkDocs.

See \url{https://www.paws-r-sdk.com/docs/workdocs_activate_user/} for full documentation.
}
\keyword{internal}
