% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_create_labels}
\alias{workdocs_create_labels}
\title{Adds the specified list of labels to the given resource (a document or
folder)}
\usage{
workdocs_create_labels(ResourceId, Labels, AuthenticationToken = NULL)
}
\arguments{
\item{ResourceId}{[required] The ID of the resource.}

\item{Labels}{[required] List of labels to add to the resource.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}
}
\description{
Adds the specified list of labels to the given resource (a document or folder)

See \url{https://www.paws-r-sdk.com/docs/workdocs_create_labels/} for full documentation.
}
\keyword{internal}
