% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_create_notification_subscription}
\alias{workdocs_create_notification_subscription}
\title{Configure Amazon WorkDocs to use Amazon SNS notifications}
\usage{
workdocs_create_notification_subscription(
  OrganizationId,
  Endpoint,
  Protocol,
  SubscriptionType
)
}
\arguments{
\item{OrganizationId}{[required] The ID of the organization.}

\item{Endpoint}{[required] The endpoint to receive the notifications. If the protocol is HTTPS, the
endpoint is a URL that begins with \code{https}.}

\item{Protocol}{[required] The protocol to use. The supported value is https, which delivers
JSON-encoded messages using HTTPS POST.}

\item{SubscriptionType}{[required] The notification type.}
}
\description{
Configure Amazon WorkDocs to use Amazon SNS notifications. The endpoint receives a confirmation message, and must confirm the subscription.

See \url{https://paws-r.github.io/docs/workdocs/create_notification_subscription.html} for full documentation.
}
\keyword{internal}
