% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_deactivate_user}
\alias{workdocs_deactivate_user}
\title{Deactivates the specified user, which revokes the user's access to
Amazon WorkDocs}
\usage{
workdocs_deactivate_user(UserId, AuthenticationToken)
}
\arguments{
\item{UserId}{[required] The ID of the user.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}
}
\description{
Deactivates the specified user, which revokes the user\'s access to
Amazon WorkDocs.
}
\section{Request syntax}{
\preformatted{svc$deactivate_user(
  UserId = "string",
  AuthenticationToken = "string"
)
}
}

\keyword{internal}
