% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_delete_comment}
\alias{workdocs_delete_comment}
\title{Deletes the specified comment from the document version}
\usage{
workdocs_delete_comment(
  AuthenticationToken = NULL,
  DocumentId,
  VersionId,
  CommentId
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{DocumentId}{[required] The ID of the document.}

\item{VersionId}{[required] The ID of the document version.}

\item{CommentId}{[required] The ID of the comment.}
}
\description{
Deletes the specified comment from the document version.

See \url{https://www.paws-r-sdk.com/docs/workdocs_delete_comment/} for full documentation.
}
\keyword{internal}
