% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_delete_document}
\alias{workdocs_delete_document}
\title{Permanently deletes the specified document and its associated metadata}
\usage{
workdocs_delete_document(AuthenticationToken, DocumentId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{DocumentId}{[required] The ID of the document.}
}
\value{
An empty list.
}
\description{
Permanently deletes the specified document and its associated metadata.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_document(
  AuthenticationToken = "string",
  DocumentId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
