% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_delete_document_version}
\alias{workdocs_delete_document_version}
\title{Deletes a specific version of a document}
\usage{
workdocs_delete_document_version(
  AuthenticationToken = NULL,
  DocumentId,
  VersionId,
  DeletePriorVersions
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{DocumentId}{[required] The ID of the document associated with the version being deleted.}

\item{VersionId}{[required] The ID of the version being deleted.}

\item{DeletePriorVersions}{[required] Deletes all versions of a document prior to the current version.}
}
\description{
Deletes a specific version of a document.

See \url{https://paws-r.github.io/docs/workdocs/delete_document_version.html} for full documentation.
}
\keyword{internal}
