% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_delete_folder}
\alias{workdocs_delete_folder}
\title{Permanently deletes the specified folder and its contents}
\usage{
workdocs_delete_folder(AuthenticationToken, FolderId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{FolderId}{[required] The ID of the folder.}
}
\value{
An empty list.
}
\description{
Permanently deletes the specified folder and its contents.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_folder(
  AuthenticationToken = "string",
  FolderId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
