% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_delete_labels}
\alias{workdocs_delete_labels}
\title{Deletes the specified list of labels from a resource}
\usage{
workdocs_delete_labels(
  ResourceId,
  AuthenticationToken = NULL,
  Labels = NULL,
  DeleteAll = NULL
)
}
\arguments{
\item{ResourceId}{[required] The ID of the resource.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{Labels}{List of labels to delete from the resource.}

\item{DeleteAll}{Flag to request removal of all labels from the specified resource.}
}
\description{
Deletes the specified list of labels from a resource.

See \url{https://www.paws-r-sdk.com/docs/workdocs_delete_labels/} for full documentation.
}
\keyword{internal}
