% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_describe_folder_contents}
\alias{workdocs_describe_folder_contents}
\title{Describes the contents of the specified folder, including its documents
and subfolders}
\usage{
workdocs_describe_folder_contents(
  AuthenticationToken = NULL,
  FolderId,
  Sort = NULL,
  Order = NULL,
  Limit = NULL,
  Marker = NULL,
  Type = NULL,
  Include = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{FolderId}{[required] The ID of the folder.}

\item{Sort}{The sorting criteria.}

\item{Order}{The order for the contents of the folder.}

\item{Limit}{The maximum number of items to return with this call.}

\item{Marker}{The marker for the next set of results. This marker was received from a
previous call.}

\item{Type}{The type of items.}

\item{Include}{The contents to include. Specify "INITIALIZED" to include initialized
documents.}
}
\description{
Describes the contents of the specified folder, including its documents and subfolders.

See \url{https://paws-r.github.io/docs/workdocs/describe_folder_contents.html} for full documentation.
}
\keyword{internal}
