% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_describe_notification_subscriptions}
\alias{workdocs_describe_notification_subscriptions}
\title{Lists the specified notification subscriptions}
\usage{
workdocs_describe_notification_subscriptions(OrganizationId, Marker,
  Limit)
}
\arguments{
\item{OrganizationId}{[required] The ID of the organization.}

\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call.)}

\item{Limit}{The maximum number of items to return with this call.}
}
\value{
A list with the following syntax:\preformatted{list(
  Subscriptions = list(
    list(
      SubscriptionId = "string",
      EndPoint = "string",
      Protocol = "HTTPS"
    )
  ),
  Marker = "string"
)
}
}
\description{
Lists the specified notification subscriptions.
}
\section{Request syntax}{
\preformatted{svc$describe_notification_subscriptions(
  OrganizationId = "string",
  Marker = "string",
  Limit = 123
)
}
}

\keyword{internal}
