% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_describe_resource_permissions}
\alias{workdocs_describe_resource_permissions}
\title{Describes the permissions of a specified resource}
\usage{
workdocs_describe_resource_permissions(
  AuthenticationToken = NULL,
  ResourceId,
  PrincipalId = NULL,
  Limit = NULL,
  Marker = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{ResourceId}{[required] The ID of the resource.}

\item{PrincipalId}{The ID of the principal to filter permissions by.}

\item{Limit}{The maximum number of items to return with this call.}

\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call)}
}
\description{
Describes the permissions of a specified resource.

See \url{https://www.paws-r-sdk.com/docs/workdocs_describe_resource_permissions/} for full documentation.
}
\keyword{internal}
