% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_describe_users}
\alias{workdocs_describe_users}
\title{Describes the specified users}
\usage{
workdocs_describe_users(
  AuthenticationToken = NULL,
  OrganizationId = NULL,
  UserIds = NULL,
  Query = NULL,
  Include = NULL,
  Order = NULL,
  Sort = NULL,
  Marker = NULL,
  Limit = NULL,
  Fields = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{OrganizationId}{The ID of the organization.}

\item{UserIds}{The IDs of the users.}

\item{Query}{A query to filter users by user name.}

\item{Include}{The state of the users. Specify "ALL" to include inactive users.}

\item{Order}{The order for the results.}

\item{Sort}{The sorting criteria.}

\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call.)}

\item{Limit}{The maximum number of items to return.}

\item{Fields}{A comma-separated list of values. Specify "STORAGE_METADATA" to include
the user storage quota and utilization information.}
}
\description{
Describes the specified users. You can describe all users or filter the results (for example, by status or organization).

See \url{https://paws-r.github.io/docs/workdocs/describe_users.html} for full documentation.
}
\keyword{internal}
