% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_get_current_user}
\alias{workdocs_get_current_user}
\title{Retrieves details of the current user for whom the authentication token
was generated}
\usage{
workdocs_get_current_user(AuthenticationToken)
}
\arguments{
\item{AuthenticationToken}{[required] Amazon WorkDocs authentication token.}
}
\description{
Retrieves details of the current user for whom the authentication token
was generated. This is not a valid action for SigV4 (administrative API)
clients.
}
\details{
This action requires an authentication token. To get an authentication
token, register an application with Amazon WorkDocs. For more
information, see \href{https://docs.aws.amazon.com/workdocs/latest/developerguide/wd-auth-user.html}{Authentication and Access Control for User Applications}
in the \emph{Amazon WorkDocs Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_current_user(
  AuthenticationToken = "string"
)
}
}

\keyword{internal}
