% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_get_document}
\alias{workdocs_get_document}
\title{Retrieves details of a document}
\usage{
workdocs_get_document(
  AuthenticationToken = NULL,
  DocumentId,
  IncludeCustomMetadata = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{DocumentId}{[required] The ID of the document.}

\item{IncludeCustomMetadata}{Set this to \code{TRUE} to include custom metadata in the response.}
}
\description{
Retrieves details of a document.

See \url{https://paws-r.github.io/docs/workdocs/get_document.html} for full documentation.
}
\keyword{internal}
