% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_get_document_version}
\alias{workdocs_get_document_version}
\title{Retrieves version metadata for the specified document}
\usage{
workdocs_get_document_version(
  AuthenticationToken = NULL,
  DocumentId,
  VersionId,
  Fields = NULL,
  IncludeCustomMetadata = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{DocumentId}{[required] The ID of the document.}

\item{VersionId}{[required] The version ID of the document.}

\item{Fields}{A comma-separated list of values. Specify "SOURCE" to include a URL for
the source document.}

\item{IncludeCustomMetadata}{Set this to TRUE to include custom metadata in the response.}
}
\description{
Retrieves version metadata for the specified document.

See \url{https://www.paws-r-sdk.com/docs/workdocs_get_document_version/} for full documentation.
}
\keyword{internal}
