% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_get_folder}
\alias{workdocs_get_folder}
\title{Retrieves the metadata of the specified folder}
\usage{
workdocs_get_folder(
  AuthenticationToken = NULL,
  FolderId,
  IncludeCustomMetadata = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{FolderId}{[required] The ID of the folder.}

\item{IncludeCustomMetadata}{Set to TRUE to include custom metadata in the response.}
}
\description{
Retrieves the metadata of the specified folder.

See \url{https://paws-r.github.io/docs/workdocs/get_folder.html} for full documentation.
}
\keyword{internal}
