% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_get_folder_path}
\alias{workdocs_get_folder_path}
\title{Retrieves the path information (the hierarchy from the root folder) for
the specified folder}
\usage{
workdocs_get_folder_path(
  AuthenticationToken = NULL,
  FolderId,
  Limit = NULL,
  Fields = NULL,
  Marker = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{FolderId}{[required] The ID of the folder.}

\item{Limit}{The maximum number of levels in the hierarchy to return.}

\item{Fields}{A comma-separated list of values. Specify "NAME" to include the names of
the parent folders.}

\item{Marker}{This value is not supported.}
}
\description{
Retrieves the path information (the hierarchy from the root folder) for the specified folder.

See \url{https://paws-r.github.io/docs/workdocs/get_folder_path.html} for full documentation.
}
\keyword{internal}
