% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_initiate_document_version_upload}
\alias{workdocs_initiate_document_version_upload}
\title{Creates a new document object and version object}
\usage{
workdocs_initiate_document_version_upload(
  AuthenticationToken = NULL,
  Id = NULL,
  Name = NULL,
  ContentCreatedTimestamp = NULL,
  ContentModifiedTimestamp = NULL,
  ContentType = NULL,
  DocumentSizeInBytes = NULL,
  ParentFolderId
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{Id}{The ID of the document.}

\item{Name}{The name of the document.}

\item{ContentCreatedTimestamp}{The timestamp when the content of the document was originally created.}

\item{ContentModifiedTimestamp}{The timestamp when the content of the document was modified.}

\item{ContentType}{The content type of the document.}

\item{DocumentSizeInBytes}{The size of the document, in bytes.}

\item{ParentFolderId}{[required] The ID of the parent folder.}
}
\description{
Creates a new document object and version object.

See \url{https://paws-r.github.io/docs/workdocs/initiate_document_version_upload.html} for full documentation.
}
\keyword{internal}
