% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_remove_resource_permission}
\alias{workdocs_remove_resource_permission}
\title{Removes the permission for the specified principal from the specified
resource}
\usage{
workdocs_remove_resource_permission(
  AuthenticationToken = NULL,
  ResourceId,
  PrincipalId,
  PrincipalType = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{ResourceId}{[required] The ID of the resource.}

\item{PrincipalId}{[required] The principal ID of the resource.}

\item{PrincipalType}{The principal type of the resource.}
}
\description{
Removes the permission for the specified principal from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/workdocs_remove_resource_permission/} for full documentation.
}
\keyword{internal}
