% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_restore_document_versions}
\alias{workdocs_restore_document_versions}
\title{Recovers a deleted version of an Amazon WorkDocs document}
\usage{
workdocs_restore_document_versions(AuthenticationToken = NULL, DocumentId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{DocumentId}{[required] The ID of the document.}
}
\description{
Recovers a deleted version of an Amazon WorkDocs document.

See \url{https://www.paws-r-sdk.com/docs/workdocs_restore_document_versions/} for full documentation.
}
\keyword{internal}
