% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_update_document}
\alias{workdocs_update_document}
\title{Updates the specified attributes of a document}
\usage{
workdocs_update_document(
  AuthenticationToken = NULL,
  DocumentId,
  Name = NULL,
  ParentFolderId = NULL,
  ResourceState = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{DocumentId}{[required] The ID of the document.}

\item{Name}{The name of the document.}

\item{ParentFolderId}{The ID of the parent folder.}

\item{ResourceState}{The resource state of the document. Only ACTIVE and RECYCLED are
supported.}
}
\description{
Updates the specified attributes of a document. The user must have access to both the document and its parent folder, if applicable.

See \url{https://paws-r.github.io/docs/workdocs/update_document.html} for full documentation.
}
\keyword{internal}
