% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_update_user}
\alias{workdocs_update_user}
\title{Updates the specified attributes of the specified user, and grants or
revokes administrative privileges to the Amazon WorkDocs site}
\usage{
workdocs_update_user(
  AuthenticationToken = NULL,
  UserId,
  GivenName = NULL,
  Surname = NULL,
  Type = NULL,
  StorageRule = NULL,
  TimeZoneId = NULL,
  Locale = NULL,
  GrantPoweruserPrivileges = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{UserId}{[required] The ID of the user.}

\item{GivenName}{The given name of the user.}

\item{Surname}{The surname of the user.}

\item{Type}{The type of the user.}

\item{StorageRule}{The amount of storage for the user.}

\item{TimeZoneId}{The time zone ID of the user.}

\item{Locale}{The locale of the user.}

\item{GrantPoweruserPrivileges}{Boolean value to determine whether the user is granted Power user
privileges.}
}
\description{
Updates the specified attributes of the specified user, and grants or revokes administrative privileges to the Amazon WorkDocs site.

See \url{https://www.paws-r-sdk.com/docs/workdocs_update_user/} for full documentation.
}
\keyword{internal}
