% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_describe_device_policy_configuration}
\alias{worklink_describe_device_policy_configuration}
\title{Describes the device policy configuration for the specified fleet}
\usage{
worklink_describe_device_policy_configuration(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DeviceCaCertificate = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Describes the device policy configuration for the specified fleet.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_device_policy_configuration(
  FleetArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
