% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_describe_domain}
\alias{worklink_describe_domain}
\title{Provides information about the domain}
\usage{
worklink_describe_domain(FleetArn, DomainName)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{DomainName}{[required] The name of the domain.}
}
\description{
Provides information about the domain.

See \url{https://paws-r.github.io/docs/worklink/describe_domain.html} for full documentation.
}
\keyword{internal}
