% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_describe_identity_provider_configuration}
\alias{worklink_describe_identity_provider_configuration}
\title{Describes the identity provider configuration of the specified fleet}
\usage{
worklink_describe_identity_provider_configuration(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  IdentityProviderType = "SAML",
  ServiceProviderSamlMetadata = "string",
  IdentityProviderSamlMetadata = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Describes the identity provider configuration of the specified fleet.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_identity_provider_configuration(
  FleetArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
