% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_disassociate_website_certificate_authority}
\alias{worklink_disassociate_website_certificate_authority}
\title{Removes a certificate authority (CA)}
\usage{
worklink_disassociate_website_certificate_authority(FleetArn,
  WebsiteCaId)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{WebsiteCaId}{[required] A unique identifier for the CA.}
}
\value{
An empty list.
}
\description{
Removes a certificate authority (CA).
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disassociate_website_certificate_authority(
  FleetArn = "string",
  WebsiteCaId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
