% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_sign_out_user}
\alias{worklink_sign_out_user}
\title{Signs the user out from all of their devices}
\usage{
worklink_sign_out_user(FleetArn, Username)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{Username}{[required] The name of the user.}
}
\description{
Signs the user out from all of their devices. The user can sign in again if they have valid credentials.

See \url{https://www.paws-r-sdk.com/docs/worklink_sign_out_user/} for full documentation.
}
\keyword{internal}
