% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_tag_resource}
\alias{worklink_tag_resource}
\title{Adds or overwrites one or more tags for the specified resource, such as
a fleet}
\usage{
worklink_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the fleet.}

\item{Tags}{[required] The tags to add to the resource. A tag is a key-value pair.}
}
\description{
Adds or overwrites one or more tags for the specified resource, such as a fleet. Each tag consists of a key and an optional value. If a resource already has a tag with the same key, this operation updates its value.

See \url{https://www.paws-r-sdk.com/docs/worklink_tag_resource/} for full documentation.
}
\keyword{internal}
