% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_update_audit_stream_configuration}
\alias{worklink_update_audit_stream_configuration}
\title{Updates the audit stream configuration for the fleet}
\usage{
worklink_update_audit_stream_configuration(FleetArn, AuditStreamArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{AuditStreamArn}{The ARN of the Amazon Kinesis data stream that receives the audit
events.}
}
\value{
An empty list.
}
\description{
Updates the audit stream configuration for the fleet.
}
\section{Request syntax}{
\preformatted{svc$update_audit_stream_configuration(
  FleetArn = "string",
  AuditStreamArn = "string"
)
}
}

\keyword{internal}
