% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_associate_delegate_to_resource}
\alias{workmail_associate_delegate_to_resource}
\title{Adds a member (user or group) to the resource's set of delegates}
\usage{
workmail_associate_delegate_to_resource(OrganizationId, ResourceId, EntityId)
}
\arguments{
\item{OrganizationId}{[required] The organization under which the resource exists.}

\item{ResourceId}{[required] The resource for which members (users or groups) are associated.}

\item{EntityId}{[required] The member (user or group) to associate to the resource.}
}
\description{
Adds a member (user or group) to the resource's set of delegates.

See \url{https://paws-r.github.io/docs/workmail/associate_delegate_to_resource.html} for full documentation.
}
\keyword{internal}
