% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_assume_impersonation_role}
\alias{workmail_assume_impersonation_role}
\title{Assumes an impersonation role for the given WorkMail organization}
\usage{
workmail_assume_impersonation_role(OrganizationId, ImpersonationRoleId)
}
\arguments{
\item{OrganizationId}{[required] The WorkMail organization under which the impersonation role will be
assumed.}

\item{ImpersonationRoleId}{[required] The impersonation role ID to assume.}
}
\description{
Assumes an impersonation role for the given WorkMail organization. This method returns an authentication token you can use to make impersonated calls.

See \url{https://www.paws-r-sdk.com/docs/workmail_assume_impersonation_role/} for full documentation.
}
\keyword{internal}
