% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_cancel_mailbox_export_job}
\alias{workmail_cancel_mailbox_export_job}
\title{Cancels a mailbox export job}
\usage{
workmail_cancel_mailbox_export_job(ClientToken, JobId, OrganizationId)
}
\arguments{
\item{ClientToken}{[required] The idempotency token for the client request.}

\item{JobId}{[required] The job ID.}

\item{OrganizationId}{[required] The organization ID.}
}
\description{
Cancels a mailbox export job.

See \url{https://www.paws-r-sdk.com/docs/workmail_cancel_mailbox_export_job/} for full documentation.
}
\keyword{internal}
