% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_create_alias}
\alias{workmail_create_alias}
\title{Adds an alias to the set of a given member (user or group) of Amazon
WorkMail}
\usage{
workmail_create_alias(OrganizationId, EntityId, Alias)
}
\arguments{
\item{OrganizationId}{[required] The organization under which the member (user or group) exists.}

\item{EntityId}{[required] The member (user or group) to which this alias is added.}

\item{Alias}{[required] The alias to add to the member set.}
}
\description{
Adds an alias to the set of a given member (user or group) of Amazon WorkMail.

See \url{https://paws-r.github.io/docs/workmail/create_alias.html} for full documentation.
}
\keyword{internal}
