% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_create_group}
\alias{workmail_create_group}
\title{Creates a group that can be used in WorkMail by calling the
RegisterToWorkMail operation}
\usage{
workmail_create_group(OrganizationId, Name, HiddenFromGlobalAddressList = NULL)
}
\arguments{
\item{OrganizationId}{[required] The organization under which the group is to be created.}

\item{Name}{[required] The name of the group.}

\item{HiddenFromGlobalAddressList}{If this parameter is enabled, the group will be hidden from the address
book.}
}
\description{
Creates a group that can be used in WorkMail by calling the \code{\link[=workmail_register_to_work_mail]{register_to_work_mail}} operation.

See \url{https://www.paws-r-sdk.com/docs/workmail_create_group/} for full documentation.
}
\keyword{internal}
