% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_create_identity_center_application}
\alias{workmail_create_identity_center_application}
\title{Creates the WorkMail application in IAM Identity Center that can be used
later in the WorkMail - IdC integration}
\usage{
workmail_create_identity_center_application(
  Name,
  InstanceArn,
  ClientToken = NULL
)
}
\arguments{
\item{Name}{[required] The name of the IAM Identity Center application.}

\item{InstanceArn}{[required] The Amazon Resource Name (ARN) of the instance.}

\item{ClientToken}{The idempotency token associated with the request.}
}
\description{
Creates the WorkMail application in IAM Identity Center that can be used later in the WorkMail - IdC integration. For more information, see PutIdentityProviderConfiguration. This action does not affect the authentication settings for any WorkMail organizations.

See \url{https://www.paws-r-sdk.com/docs/workmail_create_identity_center_application/} for full documentation.
}
\keyword{internal}
