% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_access_control_rule}
\alias{workmail_delete_access_control_rule}
\title{Deletes an access control rule for the specified WorkMail organization}
\usage{
workmail_delete_access_control_rule(OrganizationId, Name)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization.}

\item{Name}{[required] The name of the access control rule.}
}
\description{
Deletes an access control rule for the specified WorkMail organization.

See \url{https://www.paws-r-sdk.com/docs/workmail_delete_access_control_rule/} for full documentation.
}
\keyword{internal}
