% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_email_monitoring_configuration}
\alias{workmail_delete_email_monitoring_configuration}
\title{Deletes the email monitoring configuration for a specified organization}
\usage{
workmail_delete_email_monitoring_configuration(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] The ID of the organization from which the email monitoring configuration
is deleted.}
}
\description{
Deletes the email monitoring configuration for a specified organization.

See \url{https://www.paws-r-sdk.com/docs/workmail_delete_email_monitoring_configuration/} for full documentation.
}
\keyword{internal}
