% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_identity_center_application}
\alias{workmail_delete_identity_center_application}
\title{Deletes the IAM Identity Center application from WorkMail}
\usage{
workmail_delete_identity_center_application(ApplicationArn)
}
\arguments{
\item{ApplicationArn}{[required] The Amazon Resource Name (ARN) of the application.}
}
\description{
Deletes the IAM Identity Center application from WorkMail. This action does not affect the authentication settings for any WorkMail organizations.

See \url{https://www.paws-r-sdk.com/docs/workmail_delete_identity_center_application/} for full documentation.
}
\keyword{internal}
