% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_identity_provider_configuration}
\alias{workmail_delete_identity_provider_configuration}
\title{Disables the integration between IdC and WorkMail}
\usage{
workmail_delete_identity_provider_configuration(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] The Organization ID.}
}
\description{
Disables the integration between IdC and WorkMail. Authentication will continue with the directory as it was before the IdC integration. You might have to reset your directory passwords and reconfigure your desktop and mobile email clients.

See \url{https://www.paws-r-sdk.com/docs/workmail_delete_identity_provider_configuration/} for full documentation.
}
\keyword{internal}
