% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_impersonation_role}
\alias{workmail_delete_impersonation_role}
\title{Deletes an impersonation role for the given WorkMail organization}
\usage{
workmail_delete_impersonation_role(OrganizationId, ImpersonationRoleId)
}
\arguments{
\item{OrganizationId}{[required] The WorkMail organization from which to delete the impersonation role.}

\item{ImpersonationRoleId}{[required] The ID of the impersonation role to delete.}
}
\description{
Deletes an impersonation role for the given WorkMail organization.

See \url{https://www.paws-r-sdk.com/docs/workmail_delete_impersonation_role/} for full documentation.
}
\keyword{internal}
