% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_personal_access_token}
\alias{workmail_delete_personal_access_token}
\title{Deletes the Personal Access Token from the provided WorkMail
Organization}
\usage{
workmail_delete_personal_access_token(OrganizationId, PersonalAccessTokenId)
}
\arguments{
\item{OrganizationId}{[required] The Organization ID.}

\item{PersonalAccessTokenId}{[required] The Personal Access Token ID.}
}
\description{
Deletes the Personal Access Token from the provided WorkMail Organization.

See \url{https://www.paws-r-sdk.com/docs/workmail_delete_personal_access_token/} for full documentation.
}
\keyword{internal}
