% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_resource}
\alias{workmail_delete_resource}
\title{Deletes the specified resource}
\usage{
workmail_delete_resource(OrganizationId, ResourceId)
}
\arguments{
\item{OrganizationId}{[required] The identifier associated with the organization from which the resource
is deleted.}

\item{ResourceId}{[required] The identifier of the resource to be deleted.

The identifier can accept \emph{ResourceId}, or \emph{Resourcename}. The following
identity formats are available:
\itemize{
\item Resource ID: r-0123456789a0123456789b0123456789
\item Resource name: resource
}}
}
\description{
Deletes the specified resource.

See \url{https://www.paws-r-sdk.com/docs/workmail_delete_resource/} for full documentation.
}
\keyword{internal}
