% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_retention_policy}
\alias{workmail_delete_retention_policy}
\title{Deletes the specified retention policy from the specified organization}
\usage{
workmail_delete_retention_policy(OrganizationId, Id)
}
\arguments{
\item{OrganizationId}{[required] The organization ID.}

\item{Id}{[required] The retention policy ID.}
}
\description{
Deletes the specified retention policy from the specified organization.

See \url{https://www.paws-r-sdk.com/docs/workmail_delete_retention_policy/} for full documentation.
}
\keyword{internal}
