% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_email_monitoring_configuration}
\alias{workmail_describe_email_monitoring_configuration}
\title{Describes the current email monitoring configuration for a specified
organization}
\usage{
workmail_describe_email_monitoring_configuration(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] The ID of the organization for which the email monitoring configuration
is described.}
}
\description{
Describes the current email monitoring configuration for a specified organization.

See \url{https://www.paws-r-sdk.com/docs/workmail_describe_email_monitoring_configuration/} for full documentation.
}
\keyword{internal}
