% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_entity}
\alias{workmail_describe_entity}
\title{Returns basic details about an entity in WorkMail}
\usage{
workmail_describe_entity(OrganizationId, Email)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the entity exists.}

\item{Email}{[required] The email under which the entity exists.}
}
\description{
Returns basic details about an entity in WorkMail.

See \url{https://www.paws-r-sdk.com/docs/workmail_describe_entity/} for full documentation.
}
\keyword{internal}
