% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_identity_provider_configuration}
\alias{workmail_describe_identity_provider_configuration}
\title{Returns detailed information on the current IdC setup for the WorkMail
organization}
\usage{
workmail_describe_identity_provider_configuration(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] The Organization ID.}
}
\description{
Returns detailed information on the current IdC setup for the WorkMail organization.

See \url{https://www.paws-r-sdk.com/docs/workmail_describe_identity_provider_configuration/} for full documentation.
}
\keyword{internal}
