% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_inbound_dmarc_settings}
\alias{workmail_describe_inbound_dmarc_settings}
\title{Lists the settings in a DMARC policy for a specified organization}
\usage{
workmail_describe_inbound_dmarc_settings(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] Lists the ID of the given organization.}
}
\description{
Lists the settings in a DMARC policy for a specified organization.

See \url{https://paws-r.github.io/docs/workmail/describe_inbound_dmarc_settings.html} for full documentation.
}
\keyword{internal}
