% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_mailbox_export_job}
\alias{workmail_describe_mailbox_export_job}
\title{Describes the current status of a mailbox export job}
\usage{
workmail_describe_mailbox_export_job(JobId, OrganizationId)
}
\arguments{
\item{JobId}{[required] The mailbox export job ID.}

\item{OrganizationId}{[required] The organization ID.}
}
\description{
Describes the current status of a mailbox export job.

See \url{https://www.paws-r-sdk.com/docs/workmail_describe_mailbox_export_job/} for full documentation.
}
\keyword{internal}
