% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_organization}
\alias{workmail_describe_organization}
\title{Provides more information regarding a given organization based on its
identifier}
\usage{
workmail_describe_organization(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization to be described.}
}
\description{
Provides more information regarding a given organization based on its identifier.

See \url{https://www.paws-r-sdk.com/docs/workmail_describe_organization/} for full documentation.
}
\keyword{internal}
