% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_resource}
\alias{workmail_describe_resource}
\title{Returns the data available for the resource}
\usage{
workmail_describe_resource(OrganizationId, ResourceId)
}
\arguments{
\item{OrganizationId}{[required] The identifier associated with the organization for which the resource
is described.}

\item{ResourceId}{[required] The identifier of the resource to be described.}
}
\description{
Returns the data available for the resource.

See \url{https://paws-r.github.io/docs/workmail/describe_resource.html} for full documentation.
}
\keyword{internal}
