% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_get_impersonation_role}
\alias{workmail_get_impersonation_role}
\title{Gets the impersonation role details for the given WorkMail organization}
\usage{
workmail_get_impersonation_role(OrganizationId, ImpersonationRoleId)
}
\arguments{
\item{OrganizationId}{[required] The WorkMail organization from which to retrieve the impersonation role.}

\item{ImpersonationRoleId}{[required] The impersonation role ID to retrieve.}
}
\description{
Gets the impersonation role details for the given WorkMail organization.

See \url{https://paws-r.github.io/docs/workmail/get_impersonation_role.html} for full documentation.
}
\keyword{internal}
