% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_get_impersonation_role_effect}
\alias{workmail_get_impersonation_role_effect}
\title{Tests whether the given impersonation role can impersonate a target user}
\usage{
workmail_get_impersonation_role_effect(
  OrganizationId,
  ImpersonationRoleId,
  TargetUser
)
}
\arguments{
\item{OrganizationId}{[required] The WorkMail organization where the impersonation role is defined.}

\item{ImpersonationRoleId}{[required] The impersonation role ID to test.}

\item{TargetUser}{[required] The WorkMail organization user chosen to test the impersonation role.
The following identity formats are available:
\itemize{
\item User ID: \code{12345678-1234-1234-1234-123456789012} or
\code{S-1-1-12-1234567890-123456789-123456789-1234}
\item Email address: \code{user@domain.tld}
\item User name: \code{user}
}}
}
\description{
Tests whether the given impersonation role can impersonate a target user.

See \url{https://paws-r.github.io/docs/workmail/get_impersonation_role_effect.html} for full documentation.
}
\keyword{internal}
