% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_get_mail_domain}
\alias{workmail_get_mail_domain}
\title{Gets details for a mail domain, including domain records required to
configure your domain with recommended security}
\usage{
workmail_get_mail_domain(OrganizationId, DomainName)
}
\arguments{
\item{OrganizationId}{[required] The WorkMail organization for which the domain is retrieved.}

\item{DomainName}{[required] The domain from which you want to retrieve details.}
}
\description{
Gets details for a mail domain, including domain records required to configure your domain with recommended security.

See \url{https://www.paws-r-sdk.com/docs/workmail_get_mail_domain/} for full documentation.
}
\keyword{internal}
