% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_get_mailbox_details}
\alias{workmail_get_mailbox_details}
\title{Requests a user's mailbox details for a specified organization and user}
\usage{
workmail_get_mailbox_details(OrganizationId, UserId)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization that contains the user whose mailbox
details are being requested.}

\item{UserId}{[required] The identifier for the user whose mailbox details are being requested.

The identifier can be the \emph{UserId}, \emph{Username}, or \emph{email}. The
following identity formats are available:
\itemize{
\item User ID: 12345678-1234-1234-1234-123456789012 or
S-1-1-12-1234567890-123456789-123456789-1234
\item Email address: user@domain.tld
\item User name: user
}}
}
\description{
Requests a user's mailbox details for a specified organization and user.

See \url{https://www.paws-r-sdk.com/docs/workmail_get_mailbox_details/} for full documentation.
}
\keyword{internal}
