% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_get_personal_access_token_metadata}
\alias{workmail_get_personal_access_token_metadata}
\title{Requests details of a specific Personal Access Token within the WorkMail
organization}
\usage{
workmail_get_personal_access_token_metadata(
  OrganizationId,
  PersonalAccessTokenId
)
}
\arguments{
\item{OrganizationId}{[required] The Organization ID.}

\item{PersonalAccessTokenId}{[required] The Personal Access Token ID.}
}
\description{
Requests details of a specific Personal Access Token within the WorkMail organization.

See \url{https://www.paws-r-sdk.com/docs/workmail_get_personal_access_token_metadata/} for full documentation.
}
\keyword{internal}
