% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_access_control_rules}
\alias{workmail_list_access_control_rules}
\title{Lists the access control rules for the specified organization}
\usage{
workmail_list_access_control_rules(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization.}
}
\description{
Lists the access control rules for the specified organization.

See \url{https://www.paws-r-sdk.com/docs/workmail_list_access_control_rules/} for full documentation.
}
\keyword{internal}
