% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_availability_configurations}
\alias{workmail_list_availability_configurations}
\title{List all the AvailabilityConfiguration's for the given WorkMail
organization}
\usage{
workmail_list_availability_configurations(
  OrganizationId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{OrganizationId}{[required] The WorkMail organization for which the \code{AvailabilityConfiguration}'s
will be listed.}

\item{MaxResults}{The maximum number of results to return in a single call.}

\item{NextToken}{The token to use to retrieve the next page of results. The first call
does not require a token.}
}
\description{
List all the \code{AvailabilityConfiguration}'s for the given WorkMail organization.

See \url{https://www.paws-r-sdk.com/docs/workmail_list_availability_configurations/} for full documentation.
}
\keyword{internal}
