% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_impersonation_roles}
\alias{workmail_list_impersonation_roles}
\title{Lists all the impersonation roles for the given WorkMail organization}
\usage{
workmail_list_impersonation_roles(
  OrganizationId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{OrganizationId}{[required] The WorkMail organization to which the listed impersonation roles
belong.}

\item{NextToken}{The token used to retrieve the next page of results. The first call
doesn't require a token.}

\item{MaxResults}{The maximum number of results returned in a single call.}
}
\description{
Lists all the impersonation roles for the given WorkMail organization.

See \url{https://paws-r.github.io/docs/workmail/list_impersonation_roles.html} for full documentation.
}
\keyword{internal}
