% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_mailbox_export_jobs}
\alias{workmail_list_mailbox_export_jobs}
\title{Lists the mailbox export jobs started for the specified organization
within the last seven days}
\usage{
workmail_list_mailbox_export_jobs(
  OrganizationId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{OrganizationId}{[required] The organization ID.}

\item{NextToken}{The token to use to retrieve the next page of results.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
Lists the mailbox export jobs started for the specified organization within the last seven days.

See \url{https://paws-r.github.io/docs/workmail/list_mailbox_export_jobs.html} for full documentation.
}
\keyword{internal}
