% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_mobile_device_access_rules}
\alias{workmail_list_mobile_device_access_rules}
\title{Lists the mobile device access rules for the specified WorkMail
organization}
\usage{
workmail_list_mobile_device_access_rules(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] The WorkMail organization for which to list the rules.}
}
\description{
Lists the mobile device access rules for the specified WorkMail organization.

See \url{https://paws-r.github.io/docs/workmail/list_mobile_device_access_rules.html} for full documentation.
}
\keyword{internal}
