% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_organizations}
\alias{workmail_list_organizations}
\title{Returns summaries of the customer's organizations}
\usage{
workmail_list_organizations(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The token to use to retrieve the next page of results. The first call
does not contain any tokens.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
Returns summaries of the customer's organizations.

See \url{https://www.paws-r-sdk.com/docs/workmail_list_organizations/} for full documentation.
}
\keyword{internal}
