% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_personal_access_tokens}
\alias{workmail_list_personal_access_tokens}
\title{Returns a summary of your Personal Access Tokens}
\usage{
workmail_list_personal_access_tokens(
  OrganizationId,
  UserId = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{OrganizationId}{[required] The Organization ID.}

\item{UserId}{The WorkMail User ID.}

\item{NextToken}{The token from the previous response to query the next page.}

\item{MaxResults}{The maximum amount of items that should be returned in a response.}
}
\description{
Returns a summary of your Personal Access Tokens.

See \url{https://www.paws-r-sdk.com/docs/workmail_list_personal_access_tokens/} for full documentation.
}
\keyword{internal}
