% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_resource_delegates}
\alias{workmail_list_resource_delegates}
\title{Lists the delegates associated with a resource}
\usage{
workmail_list_resource_delegates(
  OrganizationId,
  ResourceId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization that contains the resource for which
delegates are listed.}

\item{ResourceId}{[required] The identifier for the resource whose delegates are listed.}

\item{NextToken}{The token used to paginate through the delegates associated with a
resource.}

\item{MaxResults}{The number of maximum results in a page.}
}
\description{
Lists the delegates associated with a resource. Users and groups can be resource delegates and answer requests on behalf of the resource.

See \url{https://paws-r.github.io/docs/workmail/list_resource_delegates.html} for full documentation.
}
\keyword{internal}
