% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_resources}
\alias{workmail_list_resources}
\title{Returns summaries of the organization's resources}
\usage{
workmail_list_resources(
  OrganizationId,
  NextToken = NULL,
  MaxResults = NULL,
  Filters = NULL
)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the resources exist.}

\item{NextToken}{The token to use to retrieve the next page of results. The first call
does not contain any tokens.}

\item{MaxResults}{The maximum number of results to return in a single call.}

\item{Filters}{Limit the resource search results based on the filter criteria. You can
only use one filter per request.}
}
\description{
Returns summaries of the organization's resources.

See \url{https://www.paws-r-sdk.com/docs/workmail_list_resources/} for full documentation.
}
\keyword{internal}
