% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_tags_for_resource}
\alias{workmail_list_tags_for_resource}
\title{Lists the tags applied to an WorkMail organization resource}
\usage{
workmail_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The resource ARN.}
}
\description{
Lists the tags applied to an WorkMail organization resource.

See \url{https://paws-r.github.io/docs/workmail/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
