% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_put_email_monitoring_configuration}
\alias{workmail_put_email_monitoring_configuration}
\title{Creates or updates the email monitoring configuration for a specified
organization}
\usage{
workmail_put_email_monitoring_configuration(
  OrganizationId,
  RoleArn = NULL,
  LogGroupArn
)
}
\arguments{
\item{OrganizationId}{[required] The ID of the organization for which the email monitoring configuration
is set.}

\item{RoleArn}{The Amazon Resource Name (ARN) of the IAM Role associated with the email
monitoring configuration. If absent, the IAM Role Arn of
AWSServiceRoleForAmazonWorkMailEvents will be used.}

\item{LogGroupArn}{[required] The Amazon Resource Name (ARN) of the CloudWatch Log group associated
with the email monitoring configuration.}
}
\description{
Creates or updates the email monitoring configuration for a specified organization.

See \url{https://www.paws-r-sdk.com/docs/workmail_put_email_monitoring_configuration/} for full documentation.
}
\keyword{internal}
