% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_put_inbound_dmarc_settings}
\alias{workmail_put_inbound_dmarc_settings}
\title{Enables or disables a DMARC policy for a given organization}
\usage{
workmail_put_inbound_dmarc_settings(OrganizationId, Enforced)
}
\arguments{
\item{OrganizationId}{[required] The ID of the organization that you are applying the DMARC policy to.}

\item{Enforced}{[required] Enforces or suspends a policy after it's applied.}
}
\description{
Enables or disables a DMARC policy for a given organization.

See \url{https://www.paws-r-sdk.com/docs/workmail_put_inbound_dmarc_settings/} for full documentation.
}
\keyword{internal}
