% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_put_retention_policy}
\alias{workmail_put_retention_policy}
\title{Puts a retention policy to the specified organization}
\usage{
workmail_put_retention_policy(
  OrganizationId,
  Id = NULL,
  Name,
  Description = NULL,
  FolderConfigurations
)
}
\arguments{
\item{OrganizationId}{[required] The organization ID.}

\item{Id}{The retention policy ID.}

\item{Name}{[required] The retention policy name.}

\item{Description}{The retention policy description.}

\item{FolderConfigurations}{[required] The retention policy folder configurations.}
}
\description{
Puts a retention policy to the specified organization.

See \url{https://paws-r.github.io/docs/workmail/put_retention_policy.html} for full documentation.
}
\keyword{internal}
