% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_reset_password}
\alias{workmail_reset_password}
\title{Allows the administrator to reset the password for a user}
\usage{
workmail_reset_password(OrganizationId, UserId, Password)
}
\arguments{
\item{OrganizationId}{[required] The identifier of the organization that contains the user for which the
password is reset.}

\item{UserId}{[required] The identifier of the user for whom the password is reset.}

\item{Password}{[required] The new password for the user.}
}
\description{
Allows the administrator to reset the password for a user.

See \url{https://www.paws-r-sdk.com/docs/workmail_reset_password/} for full documentation.
}
\keyword{internal}
