% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_tag_resource}
\alias{workmail_tag_resource}
\title{Applies the specified tags to the specified WorkMailorganization
resource}
\usage{
workmail_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The resource ARN.}

\item{Tags}{[required] The tag key-value pairs.}
}
\description{
Applies the specified tags to the specified WorkMailorganization resource.

See \url{https://www.paws-r-sdk.com/docs/workmail_tag_resource/} for full documentation.
}
\keyword{internal}
