% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_untag_resource}
\alias{workmail_untag_resource}
\title{Untags the specified tags from the specified WorkMail organization
resource}
\usage{
workmail_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The resource ARN.}

\item{TagKeys}{[required] The tag keys.}
}
\description{
Untags the specified tags from the specified WorkMail organization resource.

See \url{https://paws-r.github.io/docs/workmail/untag_resource.html} for full documentation.
}
\keyword{internal}
