% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_update_default_mail_domain}
\alias{workmail_update_default_mail_domain}
\title{Updates the default mail domain for an organization}
\usage{
workmail_update_default_mail_domain(OrganizationId, DomainName)
}
\arguments{
\item{OrganizationId}{[required] The WorkMail organization for which to list domains.}

\item{DomainName}{[required] The domain name that will become the default domain.}
}
\description{
Updates the default mail domain for an organization. The default mail domain is used by the WorkMail AWS Console to suggest an email address when enabling a mail user. You can only have one default domain.

See \url{https://paws-r.github.io/docs/workmail/update_default_mail_domain.html} for full documentation.
}
\keyword{internal}
