% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_update_group}
\alias{workmail_update_group}
\title{Updates attributes in a group}
\usage{
workmail_update_group(
  OrganizationId,
  GroupId,
  HiddenFromGlobalAddressList = NULL
)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the group exists.}

\item{GroupId}{[required] The identifier for the group to be updated.

The identifier can accept \emph{GroupId}, \emph{Groupname}, or \emph{email}. The
following identity formats are available:
\itemize{
\item Group ID: 12345678-1234-1234-1234-123456789012 or
S-1-1-12-1234567890-123456789-123456789-1234
\item Email address: group@domain.tld
\item Group name: group
}}

\item{HiddenFromGlobalAddressList}{If enabled, the group is hidden from the global address list.}
}
\description{
Updates attributes in a group.

See \url{https://www.paws-r-sdk.com/docs/workmail_update_group/} for full documentation.
}
\keyword{internal}
