% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_update_impersonation_role}
\alias{workmail_update_impersonation_role}
\title{Updates an impersonation role for the given WorkMail organization}
\usage{
workmail_update_impersonation_role(
  OrganizationId,
  ImpersonationRoleId,
  Name,
  Type,
  Description = NULL,
  Rules
)
}
\arguments{
\item{OrganizationId}{[required] The WorkMail organization that contains the impersonation role to
update.}

\item{ImpersonationRoleId}{[required] The ID of the impersonation role to update.}

\item{Name}{[required] The updated impersonation role name.}

\item{Type}{[required] The updated impersonation role type.}

\item{Description}{The updated impersonation role description.}

\item{Rules}{[required] The updated list of rules.}
}
\description{
Updates an impersonation role for the given WorkMail organization.

See \url{https://paws-r.github.io/docs/workmail/update_impersonation_role.html} for full documentation.
}
\keyword{internal}
