% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_update_primary_email_address}
\alias{workmail_update_primary_email_address}
\title{Updates the primary email for a user, group, or resource}
\usage{
workmail_update_primary_email_address(OrganizationId, EntityId, Email)
}
\arguments{
\item{OrganizationId}{[required] The organization that contains the user, group, or resource to update.}

\item{EntityId}{[required] The user, group, or resource to update.

The identifier can accept \emph{UseriD, ResourceId, or GroupId}, \emph{Username,
Resourcename, or Groupname}, or \emph{email}. The following identity formats
are available:
\itemize{
\item Entity ID: 12345678-1234-1234-1234-123456789012,
r-0123456789a0123456789b0123456789, or
S-1-1-12-1234567890-123456789-123456789-1234
\item Email address: entity@domain.tld
\item Entity name: entity
}}

\item{Email}{[required] The value of the email to be updated as primary.}
}
\description{
Updates the primary email for a user, group, or resource. The current email is moved into the list of aliases (or swapped between an existing alias and the current primary email), and the email provided in the input is promoted as the primary.

See \url{https://www.paws-r-sdk.com/docs/workmail_update_primary_email_address/} for full documentation.
}
\keyword{internal}
