% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmailmessageflow_operations.R
\name{workmailmessageflow_get_raw_message_content}
\alias{workmailmessageflow_get_raw_message_content}
\title{Retrieves the raw content of an in-transit email message, in MIME format}
\usage{
workmailmessageflow_get_raw_message_content(messageId)
}
\arguments{
\item{messageId}{[required] The identifier of the email message to retrieve.}
}
\description{
Retrieves the raw content of an in-transit email message, in MIME format.

See \url{https://www.paws-r-sdk.com/docs/workmailmessageflow_get_raw_message_content/} for full documentation.
}
\keyword{internal}
