% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmailmessageflow_operations.R
\name{workmailmessageflow_put_raw_message_content}
\alias{workmailmessageflow_put_raw_message_content}
\title{Updates the raw content of an in-transit email message, in MIME format}
\usage{
workmailmessageflow_put_raw_message_content(messageId, content)
}
\arguments{
\item{messageId}{[required] The identifier of the email message being updated.}

\item{content}{[required] Describes the raw message content of the updated email message.}
}
\description{
Updates the raw content of an in-transit email message, in MIME format.

See \url{https://www.paws-r-sdk.com/docs/workmailmessageflow_put_raw_message_content/} for full documentation.
}
\keyword{internal}
