% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_associate_ip_groups}
\alias{workspaces_associate_ip_groups}
\title{Associates the specified IP access control group with the specified
directory}
\usage{
workspaces_associate_ip_groups(DirectoryId, GroupIds)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{GroupIds}{[required] The identifiers of one or more IP access control groups.}
}
\description{
Associates the specified IP access control group with the specified directory.

See \url{https://paws-r.github.io/docs/workspaces/associate_ip_groups.html} for full documentation.
}
\keyword{internal}
