% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_associate_workspace_application}
\alias{workspaces_associate_workspace_application}
\title{Associates the specified application to the specified WorkSpace}
\usage{
workspaces_associate_workspace_application(WorkspaceId, ApplicationId)
}
\arguments{
\item{WorkspaceId}{[required] The identifier of the WorkSpace.}

\item{ApplicationId}{[required] The identifier of the application.}
}
\description{
Associates the specified application to the specified WorkSpace.

See \url{https://www.paws-r-sdk.com/docs/workspaces_associate_workspace_application/} for full documentation.
}
\keyword{internal}
