% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_authorize_ip_rules}
\alias{workspaces_authorize_ip_rules}
\title{Adds one or more rules to the specified IP access control group}
\usage{
workspaces_authorize_ip_rules(GroupId, UserRules)
}
\arguments{
\item{GroupId}{[required] The identifier of the group.}

\item{UserRules}{[required] The rules to add to the group.}
}
\description{
Adds one or more rules to the specified IP access control group.

See \url{https://paws-r.github.io/docs/workspaces/authorize_ip_rules.html} for full documentation.
}
\keyword{internal}
