% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_account_link_invitation}
\alias{workspaces_create_account_link_invitation}
\title{Creates the account link invitation}
\usage{
workspaces_create_account_link_invitation(TargetAccountId, ClientToken = NULL)
}
\arguments{
\item{TargetAccountId}{[required] The identifier of the target account.}

\item{ClientToken}{A string of up to 64 ASCII characters that Amazon WorkSpaces uses to
ensure idempotent creation.}
}
\description{
Creates the account link invitation.

See \url{https://www.paws-r-sdk.com/docs/workspaces_create_account_link_invitation/} for full documentation.
}
\keyword{internal}
