% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_ip_group}
\alias{workspaces_create_ip_group}
\title{Creates an IP access control group}
\usage{
workspaces_create_ip_group(
  GroupName,
  GroupDesc = NULL,
  UserRules = NULL,
  Tags = NULL
)
}
\arguments{
\item{GroupName}{[required] The name of the group.}

\item{GroupDesc}{The description of the group.}

\item{UserRules}{The rules to add to the group.}

\item{Tags}{The tags. Each WorkSpaces resource can have a maximum of 50 tags.}
}
\description{
Creates an IP access control group.

See \url{https://www.paws-r-sdk.com/docs/workspaces_create_ip_group/} for full documentation.
}
\keyword{internal}
