% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_standby_workspaces}
\alias{workspaces_create_standby_workspaces}
\title{Creates a standby WorkSpace in a secondary Region}
\usage{
workspaces_create_standby_workspaces(PrimaryRegion, StandbyWorkspaces)
}
\arguments{
\item{PrimaryRegion}{[required] The Region of the primary WorkSpace.}

\item{StandbyWorkspaces}{[required] Information about the standby WorkSpace to be created.}
}
\description{
Creates a standby WorkSpace in a secondary Region.

See \url{https://paws-r.github.io/docs/workspaces/create_standby_workspaces.html} for full documentation.
}
\keyword{internal}
