% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_tags}
\alias{workspaces_create_tags}
\title{Creates the specified tags for the specified WorkSpaces resource}
\usage{
workspaces_create_tags(ResourceId, Tags)
}
\arguments{
\item{ResourceId}{[required] The identifier of the WorkSpaces resource. The supported resource types
are WorkSpaces, registered directories, images, custom bundles, IP
access control groups, and connection aliases.}

\item{Tags}{[required] The tags. Each WorkSpaces resource can have a maximum of 50 tags.}
}
\description{
Creates the specified tags for the specified WorkSpaces resource.

See \url{https://paws-r.github.io/docs/workspaces/create_tags.html} for full documentation.
}
\keyword{internal}
