% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_workspace_bundle}
\alias{workspaces_create_workspace_bundle}
\title{Creates the specified WorkSpace bundle}
\usage{
workspaces_create_workspace_bundle(BundleName, BundleDescription,
  ImageId, ComputeType, UserStorage, RootStorage, Tags)
}
\arguments{
\item{BundleName}{[required] The name of the bundle.}

\item{BundleDescription}{[required] The description of the bundle.}

\item{ImageId}{[required] The identifier of the image that is used to create the bundle.}

\item{ComputeType}{[required]}

\item{UserStorage}{[required]}

\item{RootStorage}{}

\item{Tags}{The tags associated with the bundle.

To add tags at the same time that you're creating the bundle, you must
create an IAM policy that grants your IAM user permissions to use
\code{workspaces:CreateTags}.}
}
\value{
A list with the following syntax:\preformatted{list(
  WorkspaceBundle = list(
    BundleId = "string",
    Name = "string",
    Owner = "string",
    Description = "string",
    ImageId = "string",
    RootStorage = list(
      Capacity = "string"
    ),
    UserStorage = list(
      Capacity = "string"
    ),
    ComputeType = list(
      Name = "VALUE"|"STANDARD"|"PERFORMANCE"|"POWER"|"GRAPHICS"|"POWERPRO"|"GRAPHICSPRO"
    ),
    LastUpdatedTime = as.POSIXct(
      "2015-01-01"
    ),
    CreationTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Creates the specified WorkSpace bundle. For more information about
creating WorkSpace bundles, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/create-custom-bundle.html}{Create a Custom WorkSpaces Image and Bundle}.
}
\section{Request syntax}{
\preformatted{svc$create_workspace_bundle(
  BundleName = "string",
  BundleDescription = "string",
  ImageId = "string",
  ComputeType = list(
    Name = "VALUE"|"STANDARD"|"PERFORMANCE"|"POWER"|"GRAPHICS"|"POWERPRO"|"GRAPHICSPRO"
  ),
  UserStorage = list(
    Capacity = "string"
  ),
  RootStorage = list(
    Capacity = "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
