% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_workspace_image}
\alias{workspaces_create_workspace_image}
\title{Creates a new WorkSpace image from an existing WorkSpace}
\usage{
workspaces_create_workspace_image(Name, Description, WorkspaceId, Tags = NULL)
}
\arguments{
\item{Name}{[required] The name of the new WorkSpace image.}

\item{Description}{[required] The description of the new WorkSpace image.}

\item{WorkspaceId}{[required] The identifier of the source WorkSpace}

\item{Tags}{The tags that you want to add to the new WorkSpace image. To add tags
when you're creating the image, you must create an IAM policy that
grants your IAM user permission to use \code{workspaces:CreateTags}.}
}
\description{
Creates a new WorkSpace image from an existing WorkSpace.

See \url{https://www.paws-r-sdk.com/docs/workspaces_create_workspace_image/} for full documentation.
}
\keyword{internal}
