% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_workspaces_pool}
\alias{workspaces_create_workspaces_pool}
\title{End of support notice: On December 31, 2027, Amazon Web Services will
end support for Amazon WorkSpaces Pools}
\usage{
workspaces_create_workspaces_pool(
  PoolName,
  Description,
  BundleId,
  DirectoryId,
  Capacity,
  Tags = NULL,
  ApplicationSettings = NULL,
  TimeoutSettings = NULL,
  RunningMode = NULL
)
}
\arguments{
\item{PoolName}{[required] The name of the pool.}

\item{Description}{[required] The pool description.}

\item{BundleId}{[required] The identifier of the bundle for the pool.}

\item{DirectoryId}{[required] The identifier of the directory for the pool.}

\item{Capacity}{[required] The user capacity of the pool.}

\item{Tags}{The tags for the pool.}

\item{ApplicationSettings}{Indicates the application settings of the pool.}

\item{TimeoutSettings}{Indicates the timeout settings of the pool.}

\item{RunningMode}{The running mode for the pool.}
}
\description{
End of support notice: On December 31, 2027, Amazon Web Services will end support for Amazon WorkSpaces Pools. After December 31, 2027, you will no longer be able to access the Amazon WorkSpaces Pools console or Amazon WorkSpaces Pools resources. For more information, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/wsp-pools-end-of-support.html}{Amazon WorkSpaces Pools end of support}.

See \url{https://www.paws-r-sdk.com/docs/workspaces_create_workspaces_pool/} for full documentation.
}
\keyword{internal}
