% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_delete_account_link_invitation}
\alias{workspaces_delete_account_link_invitation}
\title{Deletes the account link invitation}
\usage{
workspaces_delete_account_link_invitation(LinkId, ClientToken = NULL)
}
\arguments{
\item{LinkId}{[required] The identifier of the account link.}

\item{ClientToken}{A string of up to 64 ASCII characters that Amazon WorkSpaces uses to
ensure idempotent creation.}
}
\description{
Deletes the account link invitation.

See \url{https://www.paws-r-sdk.com/docs/workspaces_delete_account_link_invitation/} for full documentation.
}
\keyword{internal}
