% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_delete_client_branding}
\alias{workspaces_delete_client_branding}
\title{Deletes customized client branding}
\usage{
workspaces_delete_client_branding(ResourceId, Platforms)
}
\arguments{
\item{ResourceId}{[required] The directory identifier of the WorkSpace for which you want to delete
client branding.}

\item{Platforms}{[required] The device type for which you want to delete client branding.}
}
\description{
Deletes customized client branding. Client branding allows you to customize your WorkSpace's client login portal. You can tailor your login portal company logo, the support email address, support link, link to reset password, and a custom message for users trying to sign in.

See \url{https://www.paws-r-sdk.com/docs/workspaces_delete_client_branding/} for full documentation.
}
\keyword{internal}
