% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_delete_connection_alias}
\alias{workspaces_delete_connection_alias}
\title{Deletes the specified connection alias}
\usage{
workspaces_delete_connection_alias(AliasId)
}
\arguments{
\item{AliasId}{[required] The identifier of the connection alias to delete.}
}
\description{
Deletes the specified connection alias. For more information, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/cross-region-redirection.html}{Cross-Region Redirection for Amazon WorkSpaces}.

See \url{https://www.paws-r-sdk.com/docs/workspaces_delete_connection_alias/} for full documentation.
}
\keyword{internal}
