% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_delete_tags}
\alias{workspaces_delete_tags}
\title{Deletes the specified tags from the specified WorkSpaces resource}
\usage{
workspaces_delete_tags(ResourceId, TagKeys)
}
\arguments{
\item{ResourceId}{[required] The identifier of the WorkSpaces resource. The supported resource types
are WorkSpaces, registered directories, images, custom bundles, IP
access control groups, and connection aliases.}

\item{TagKeys}{[required] The tag keys.}
}
\description{
Deletes the specified tags from the specified WorkSpaces resource.

See \url{https://www.paws-r-sdk.com/docs/workspaces_delete_tags/} for full documentation.
}
\keyword{internal}
