% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_delete_workspace_image}
\alias{workspaces_delete_workspace_image}
\title{Deletes the specified image from your account}
\usage{
workspaces_delete_workspace_image(ImageId)
}
\arguments{
\item{ImageId}{[required] The identifier of the image.}
}
\value{
An empty list.
}
\description{
Deletes the specified image from your account. To delete an image, you
must first delete any bundles that are associated with the image and
unshare the image if it is shared with other accounts.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_workspace_image(
  ImageId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
