% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_application_associations}
\alias{workspaces_describe_application_associations}
\title{Describes the associations between the application and the specified
associated resources}
\usage{
workspaces_describe_application_associations(
  MaxResults = NULL,
  NextToken = NULL,
  ApplicationId,
  AssociatedResourceTypes
)
}
\arguments{
\item{MaxResults}{The maximum number of associations to return.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}

\item{ApplicationId}{[required] The identifier of the specified application.}

\item{AssociatedResourceTypes}{[required] The resource type of the associated resources.}
}
\description{
Describes the associations between the application and the specified associated resources.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_application_associations/} for full documentation.
}
\keyword{internal}
