% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_applications}
\alias{workspaces_describe_applications}
\title{Describes the specified applications by filtering based on their compute
types, license availability, operating systems, and owners}
\usage{
workspaces_describe_applications(
  ApplicationIds = NULL,
  ComputeTypeNames = NULL,
  LicenseType = NULL,
  OperatingSystemNames = NULL,
  Owner = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ApplicationIds}{The identifiers of one or more applications.}

\item{ComputeTypeNames}{The compute types supported by the applications.}

\item{LicenseType}{The license availability for the applications.}

\item{OperatingSystemNames}{The operating systems supported by the applications.}

\item{Owner}{The owner of the applications.}

\item{MaxResults}{The maximum number of applications to return.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}
}
\description{
Describes the specified applications by filtering based on their compute types, license availability, operating systems, and owners.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_applications/} for full documentation.
}
\keyword{internal}
