% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_bundle_associations}
\alias{workspaces_describe_bundle_associations}
\title{Describes the associations between the applications and the specified
bundle}
\usage{
workspaces_describe_bundle_associations(BundleId, AssociatedResourceTypes)
}
\arguments{
\item{BundleId}{[required] The identifier of the bundle.}

\item{AssociatedResourceTypes}{[required] The resource types of the associated resource.}
}
\description{
Describes the associations between the applications and the specified bundle.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_bundle_associations/} for full documentation.
}
\keyword{internal}
