% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_client_branding}
\alias{workspaces_describe_client_branding}
\title{Describes the specified client branding}
\usage{
workspaces_describe_client_branding(ResourceId)
}
\arguments{
\item{ResourceId}{[required] The directory identifier of the WorkSpace for which you want to view
client branding information.}
}
\description{
Describes the specified client branding. Client branding allows you to customize the log in page of various device types for your users. You can add your company logo, the support email address, support link, link to reset password, and a custom message for users trying to sign in.

See \url{https://paws-r.github.io/docs/workspaces/describe_client_branding.html} for full documentation.
}
\keyword{internal}
