% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_client_properties}
\alias{workspaces_describe_client_properties}
\title{Retrieves a list that describes one or more specified Amazon WorkSpaces
clients}
\usage{
workspaces_describe_client_properties(ResourceIds)
}
\arguments{
\item{ResourceIds}{[required] The resource identifier, in the form of directory IDs.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ClientPropertiesList = list(
    list(
      ResourceId = "string",
      ClientProperties = list(
        ReconnectEnabled = "ENABLED"|"DISABLED"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves a list that describes one or more specified Amazon WorkSpaces
clients.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_client_properties(
  ResourceIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
