% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_connection_aliases}
\alias{workspaces_describe_connection_aliases}
\title{Retrieves a list that describes the connection aliases used for
cross-Region redirection}
\usage{
workspaces_describe_connection_aliases(
  AliasIds = NULL,
  ResourceId = NULL,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AliasIds}{The identifiers of the connection aliases to describe.}

\item{ResourceId}{The identifier of the directory associated with the connection alias.}

\item{Limit}{The maximum number of connection aliases to return.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}
}
\description{
Retrieves a list that describes the connection aliases used for cross-Region redirection. For more information, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/cross-region-redirection.html}{Cross-Region Redirection for Amazon WorkSpaces}.

See \url{https://paws-r.github.io/docs/workspaces/describe_connection_aliases.html} for full documentation.
}
\keyword{internal}
