% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_custom_workspace_image_import}
\alias{workspaces_describe_custom_workspace_image_import}
\title{Retrieves information about a WorkSpace BYOL image being imported via
ImportCustomWorkspaceImage}
\usage{
workspaces_describe_custom_workspace_image_import(ImageId)
}
\arguments{
\item{ImageId}{[required] The identifier of the WorkSpace image.}
}
\description{
Retrieves information about a WorkSpace BYOL image being imported via ImportCustomWorkspaceImage.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_custom_workspace_image_import/} for full documentation.
}
\keyword{internal}
