% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_image_associations}
\alias{workspaces_describe_image_associations}
\title{Describes the associations between the applications and the specified
image}
\usage{
workspaces_describe_image_associations(ImageId, AssociatedResourceTypes)
}
\arguments{
\item{ImageId}{[required] The identifier of the image.}

\item{AssociatedResourceTypes}{[required] The resource types of the associated resource.}
}
\description{
Describes the associations between the applications and the specified image.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_image_associations/} for full documentation.
}
\keyword{internal}
