% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_tags}
\alias{workspaces_describe_tags}
\title{Describes the specified tags for the specified WorkSpaces resource}
\usage{
workspaces_describe_tags(ResourceId)
}
\arguments{
\item{ResourceId}{[required] The identifier of the WorkSpaces resource. The supported resource types
are WorkSpaces, registered directories, images, custom bundles, IP
access control groups, and connection aliases.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TagList = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the specified tags for the specified WorkSpaces resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_tags(
  ResourceId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
