% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_workspace_associations}
\alias{workspaces_describe_workspace_associations}
\title{Describes the associations betweens applications and the specified
WorkSpace}
\usage{
workspaces_describe_workspace_associations(
  WorkspaceId,
  AssociatedResourceTypes
)
}
\arguments{
\item{WorkspaceId}{[required] The identifier of the WorkSpace.}

\item{AssociatedResourceTypes}{[required] The resource types of the associated resources.}
}
\description{
Describes the associations betweens applications and the specified WorkSpace.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_workspace_associations/} for full documentation.
}
\keyword{internal}
