% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_workspace_bundles}
\alias{workspaces_describe_workspace_bundles}
\title{Retrieves a list that describes the available WorkSpace bundles}
\usage{
workspaces_describe_workspace_bundles(
  BundleIds = NULL,
  Owner = NULL,
  NextToken = NULL
)
}
\arguments{
\item{BundleIds}{The identifiers of the bundles. You cannot combine this parameter with
any other filter.}

\item{Owner}{The owner of the bundles. You cannot combine this parameter with any
other filter.

To describe the bundles provided by Amazon Web Services, specify
\code{AMAZON}. To describe the bundles that belong to your account, don't
specify a value.}

\item{NextToken}{The token for the next set of results. (You received this token from a
previous call.)}
}
\description{
Retrieves a list that describes the available WorkSpace bundles.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_workspace_bundles/} for full documentation.
}
\keyword{internal}
