% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_workspace_directories}
\alias{workspaces_describe_workspace_directories}
\title{Describes the available AWS Directory Service directories that are
registered with Amazon WorkSpaces}
\usage{
workspaces_describe_workspace_directories(DirectoryIds, NextToken)
}
\arguments{
\item{DirectoryIds}{The identifiers of the directories. If the value is null, all
directories are retrieved.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}
}
\description{
Describes the available AWS Directory Service directories that are
registered with Amazon WorkSpaces.
}
\section{Request syntax}{
\preformatted{svc$describe_workspace_directories(
  DirectoryIds = list(
    "string"
  ),
  NextToken = "string"
)
}
}

\keyword{internal}
