% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_workspace_images}
\alias{workspaces_describe_workspace_images}
\title{Retrieves a list that describes one or more specified images, if the
image identifiers are provided}
\usage{
workspaces_describe_workspace_images(
  ImageIds = NULL,
  ImageType = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ImageIds}{The identifier of the image.}

\item{ImageType}{The type (owned or shared) of the image.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}

\item{MaxResults}{The maximum number of items to return.}
}
\description{
Retrieves a list that describes one or more specified images, if the image identifiers are provided. Otherwise, all images in the account are described.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_workspace_images/} for full documentation.
}
\keyword{internal}
