% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_workspace_snapshots}
\alias{workspaces_describe_workspace_snapshots}
\title{Describes the snapshots for the specified WorkSpace}
\usage{
workspaces_describe_workspace_snapshots(WorkspaceId)
}
\arguments{
\item{WorkspaceId}{[required] The identifier of the WorkSpace.}
}
\description{
Describes the snapshots for the specified WorkSpace.
}
\section{Request syntax}{
\preformatted{svc$describe_workspace_snapshots(
  WorkspaceId = "string"
)
}
}

\keyword{internal}
