% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_workspaces_connection_status}
\alias{workspaces_describe_workspaces_connection_status}
\title{Describes the connection status of the specified WorkSpaces}
\usage{
workspaces_describe_workspaces_connection_status(
  WorkspaceIds = NULL,
  NextToken = NULL
)
}
\arguments{
\item{WorkspaceIds}{The identifiers of the WorkSpaces. You can specify up to 25 WorkSpaces.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}
}
\description{
Describes the connection status of the specified WorkSpaces.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_workspaces_connection_status/} for full documentation.
}
\keyword{internal}
