% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_workspaces_pools}
\alias{workspaces_describe_workspaces_pools}
\title{End of support notice: On December 31, 2027, Amazon Web Services will
end support for Amazon WorkSpaces Pools}
\usage{
workspaces_describe_workspaces_pools(
  PoolIds = NULL,
  Filters = NULL,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{PoolIds}{The identifier of the WorkSpaces Pools.}

\item{Filters}{The filter conditions for the WorkSpaces Pool to return.}

\item{Limit}{The maximum number of items to return.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated, provide this token to receive the next set of results.}
}
\description{
End of support notice: On December 31, 2027, Amazon Web Services will end support for Amazon WorkSpaces Pools. After December 31, 2027, you will no longer be able to access the Amazon WorkSpaces Pools console or Amazon WorkSpaces Pools resources. For more information, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/wsp-pools-end-of-support.html}{Amazon WorkSpaces Pools end of support}.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_workspaces_pools/} for full documentation.
}
\keyword{internal}
