% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_disassociate_connection_alias}
\alias{workspaces_disassociate_connection_alias}
\title{Disassociates a connection alias from a directory}
\usage{
workspaces_disassociate_connection_alias(AliasId)
}
\arguments{
\item{AliasId}{[required] The identifier of the connection alias to disassociate.}
}
\description{
Disassociates a connection alias from a directory. Disassociating a connection alias disables cross-Region redirection between two directories in different Regions. For more information, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/cross-region-redirection.html}{Cross-Region Redirection for Amazon WorkSpaces}.

See \url{https://www.paws-r-sdk.com/docs/workspaces_disassociate_connection_alias/} for full documentation.
}
\keyword{internal}
