% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_disassociate_ip_groups}
\alias{workspaces_disassociate_ip_groups}
\title{Disassociates the specified IP access control group from the specified
directory}
\usage{
workspaces_disassociate_ip_groups(DirectoryId, GroupIds)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{GroupIds}{[required] The identifiers of one or more IP access control groups.}
}
\value{
An empty list.
}
\description{
Disassociates the specified IP access control group from the specified
directory.
}
\section{Request syntax}{
\preformatted{svc$disassociate_ip_groups(
  DirectoryId = "string",
  GroupIds = list(
    "string"
  )
)
}
}

\keyword{internal}
