% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_get_account_link}
\alias{workspaces_get_account_link}
\title{Retrieves account link information}
\usage{
workspaces_get_account_link(LinkId = NULL, LinkedAccountId = NULL)
}
\arguments{
\item{LinkId}{The identifier of the account to link.}

\item{LinkedAccountId}{The identifier of the account link}
}
\description{
Retrieves account link information.

See \url{https://www.paws-r-sdk.com/docs/workspaces_get_account_link/} for full documentation.
}
\keyword{internal}
