% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_import_client_branding}
\alias{workspaces_import_client_branding}
\title{Imports client branding}
\usage{
workspaces_import_client_branding(
  ResourceId,
  DeviceTypeWindows = NULL,
  DeviceTypeOsx = NULL,
  DeviceTypeAndroid = NULL,
  DeviceTypeIos = NULL,
  DeviceTypeLinux = NULL,
  DeviceTypeWeb = NULL
)
}
\arguments{
\item{ResourceId}{[required] The directory identifier of the WorkSpace for which you want to import
client branding.}

\item{DeviceTypeWindows}{The branding information to import for Windows devices.}

\item{DeviceTypeOsx}{The branding information to import for macOS devices.}

\item{DeviceTypeAndroid}{The branding information to import for Android devices.}

\item{DeviceTypeIos}{The branding information to import for iOS devices.}

\item{DeviceTypeLinux}{The branding information to import for Linux devices.}

\item{DeviceTypeWeb}{The branding information to import for web access.}
}
\description{
Imports client branding. Client branding allows you to customize your WorkSpace's client login portal. You can tailor your login portal company logo, the support email address, support link, link to reset password, and a custom message for users trying to sign in.

See \url{https://paws-r.github.io/docs/workspaces/import_client_branding.html} for full documentation.
}
\keyword{internal}
