% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_import_custom_workspace_image}
\alias{workspaces_import_custom_workspace_image}
\title{Imports the specified Windows 10 or 11 Bring Your Own License (BYOL)
image into Amazon WorkSpaces using EC2 Image Builder}
\usage{
workspaces_import_custom_workspace_image(
  ImageName,
  ImageDescription,
  ComputeType,
  Protocol,
  ImageSource,
  InfrastructureConfigurationArn,
  Platform,
  OsVersion,
  Tags = NULL
)
}
\arguments{
\item{ImageName}{[required] The name of the WorkSpace image.}

\item{ImageDescription}{[required] The description of the WorkSpace image.}

\item{ComputeType}{[required] The supported compute type for the WorkSpace image.}

\item{Protocol}{[required] The supported protocol for the WorkSpace image. Windows 11 does not
support PCOIP protocol.}

\item{ImageSource}{[required] The options for image import source.}

\item{InfrastructureConfigurationArn}{[required] The infrastructure configuration ARN that specifies how the WorkSpace
image is built.}

\item{Platform}{[required] The platform for the WorkSpace image source.}

\item{OsVersion}{[required] The OS version for the WorkSpace image source.}

\item{Tags}{The resource tags. Each WorkSpaces resource can have a maximum of 50
tags.}
}
\description{
Imports the specified Windows 10 or 11 Bring Your Own License (BYOL) image into Amazon WorkSpaces using EC2 Image Builder. The image must be an already licensed image that is in your Amazon Web Services account, and you must own the image. For more information about creating BYOL images, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/byol-windows-images.html}{Bring Your Own Windows Desktop Licenses}.

See \url{https://www.paws-r-sdk.com/docs/workspaces_import_custom_workspace_image/} for full documentation.
}
\keyword{internal}
