% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_list_available_management_cidr_ranges}
\alias{workspaces_list_available_management_cidr_ranges}
\title{Retrieves a list of IP address ranges, specified as IPv4 CIDR blocks,
that you can use for the network management interface when you enable
Bring Your Own License (BYOL)}
\usage{
workspaces_list_available_management_cidr_ranges(
  ManagementCidrRangeConstraint,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ManagementCidrRangeConstraint}{[required] The IP address range to search. Specify an IP address range that is
compatible with your network and in CIDR notation (that is, specify the
range as an IPv4 CIDR block).}

\item{MaxResults}{The maximum number of items to return.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}
}
\description{
Retrieves a list of IP address ranges, specified as IPv4 CIDR blocks, that you can use for the network management interface when you enable Bring Your Own License (BYOL).

See \url{https://www.paws-r-sdk.com/docs/workspaces_list_available_management_cidr_ranges/} for full documentation.
}
\keyword{internal}
