% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_migrate_workspace}
\alias{workspaces_migrate_workspace}
\title{Migrates a WorkSpace from one operating system or bundle type to
another, while retaining the data on the user volume}
\usage{
workspaces_migrate_workspace(SourceWorkspaceId, BundleId)
}
\arguments{
\item{SourceWorkspaceId}{[required] The identifier of the WorkSpace to migrate from.}

\item{BundleId}{[required] The identifier of the target bundle type to migrate the WorkSpace to.}
}
\description{
Migrates a WorkSpace from one operating system or bundle type to another, while retaining the data on the user volume.

See \url{https://www.paws-r-sdk.com/docs/workspaces_migrate_workspace/} for full documentation.
}
\keyword{internal}
