% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_modify_certificate_based_auth_properties}
\alias{workspaces_modify_certificate_based_auth_properties}
\title{Modifies the properties of the certificate-based authentication you want
to use with your WorkSpaces}
\usage{
workspaces_modify_certificate_based_auth_properties(
  ResourceId,
  CertificateBasedAuthProperties = NULL,
  PropertiesToDelete = NULL
)
}
\arguments{
\item{ResourceId}{[required] The resource identifiers, in the form of directory IDs.}

\item{CertificateBasedAuthProperties}{The properties of the certificate-based authentication.}

\item{PropertiesToDelete}{The properties of the certificate-based authentication you want to
delete.}
}
\description{
Modifies the properties of the certificate-based authentication you want to use with your WorkSpaces.

See \url{https://www.paws-r-sdk.com/docs/workspaces_modify_certificate_based_auth_properties/} for full documentation.
}
\keyword{internal}
