% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_modify_endpoint_encryption_mode}
\alias{workspaces_modify_endpoint_encryption_mode}
\title{Modifies the endpoint encryption mode that allows you to configure the
specified directory between Standard TLS and FIPS 140-2 validated mode}
\usage{
workspaces_modify_endpoint_encryption_mode(DirectoryId, EndpointEncryptionMode)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{EndpointEncryptionMode}{[required] The encryption mode used for endpoint connections when streaming to
WorkSpaces Personal or WorkSpace Pools.}
}
\description{
Modifies the endpoint encryption mode that allows you to configure the specified directory between Standard TLS and FIPS 140-2 validated mode.

See \url{https://www.paws-r-sdk.com/docs/workspaces_modify_endpoint_encryption_mode/} for full documentation.
}
\keyword{internal}
