% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_modify_streaming_properties}
\alias{workspaces_modify_streaming_properties}
\title{Modifies the specified streaming properties}
\usage{
workspaces_modify_streaming_properties(ResourceId, StreamingProperties = NULL)
}
\arguments{
\item{ResourceId}{[required] The identifier of the resource.}

\item{StreamingProperties}{The streaming properties to configure.}
}
\description{
Modifies the specified streaming properties.

See \url{https://www.paws-r-sdk.com/docs/workspaces_modify_streaming_properties/} for full documentation.
}
\keyword{internal}
