% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_modify_workspace_access_properties}
\alias{workspaces_modify_workspace_access_properties}
\title{Specifies which devices and operating systems users can use to access
their WorkSpaces}
\usage{
workspaces_modify_workspace_access_properties(
  ResourceId,
  WorkspaceAccessProperties
)
}
\arguments{
\item{ResourceId}{[required] The identifier of the directory.}

\item{WorkspaceAccessProperties}{[required] The device types and operating systems to enable or disable for access.}
}
\description{
Specifies which devices and operating systems users can use to access their WorkSpaces. For more information, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/update-directory-details.html#control-device-access}{Control Device Access}.

See \url{https://paws-r.github.io/docs/workspaces/modify_workspace_access_properties.html} for full documentation.
}
\keyword{internal}
