% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_modify_workspace_state}
\alias{workspaces_modify_workspace_state}
\title{Sets the state of the specified WorkSpace}
\usage{
workspaces_modify_workspace_state(WorkspaceId, WorkspaceState)
}
\arguments{
\item{WorkspaceId}{[required] The identifier of the WorkSpace.}

\item{WorkspaceState}{[required] The WorkSpace state.}
}
\description{
Sets the state of the specified WorkSpace.

See \url{https://www.paws-r-sdk.com/docs/workspaces_modify_workspace_state/} for full documentation.
}
\keyword{internal}
