% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_reboot_workspaces}
\alias{workspaces_reboot_workspaces}
\title{Reboots the specified WorkSpaces}
\usage{
workspaces_reboot_workspaces(RebootWorkspaceRequests)
}
\arguments{
\item{RebootWorkspaceRequests}{[required] The WorkSpaces to reboot. You can specify up to 25 WorkSpaces.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FailedRequests = list(
    list(
      WorkspaceId = "string",
      ErrorCode = "string",
      ErrorMessage = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Reboots the specified WorkSpaces.

You cannot reboot a WorkSpace unless its state is \code{AVAILABLE} or
\code{UNHEALTHY}.

This operation is asynchronous and returns before the WorkSpaces have
rebooted.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$reboot_workspaces(
  RebootWorkspaceRequests = list(
    list(
      WorkspaceId = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
