% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_register_workspace_directory}
\alias{workspaces_register_workspace_directory}
\title{Registers the specified directory}
\usage{
workspaces_register_workspace_directory(
  DirectoryId = NULL,
  SubnetIds = NULL,
  EnableSelfService = NULL,
  Tenancy = NULL,
  Tags = NULL,
  WorkspaceDirectoryName = NULL,
  WorkspaceDirectoryDescription = NULL,
  UserIdentityType = NULL,
  IdcInstanceArn = NULL,
  MicrosoftEntraConfig = NULL,
  WorkspaceType = NULL,
  ActiveDirectoryConfig = NULL
)
}
\arguments{
\item{DirectoryId}{The identifier of the directory. You cannot register a directory if it
does not have a status of Active. If the directory does not have a
status of Active, you will receive an InvalidResourceStateException
error. If you have already registered the maximum number of directories
that you can register with Amazon WorkSpaces, you will receive a
ResourceLimitExceededException error. Deregister directories that you
are not using for WorkSpaces, and try again.}

\item{SubnetIds}{The identifiers of the subnets for your virtual private cloud (VPC).
Make sure that the subnets are in supported Availability Zones. The
subnets must also be in separate Availability Zones. If these conditions
are not met, you will receive an OperationNotSupportedException error.}

\item{EnableSelfService}{Indicates whether self-service capabilities are enabled or disabled.}

\item{Tenancy}{Indicates whether your WorkSpace directory is dedicated or shared. To
use Bring Your Own License (BYOL) images, this value must be set to
\code{DEDICATED} and your Amazon Web Services account must be enabled for
BYOL. If your account has not been enabled for BYOL, you will receive an
InvalidParameterValuesException error. For more information about BYOL
images, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/byol-windows-images.html}{Bring Your Own Windows Desktop Images}.}

\item{Tags}{The tags associated with the directory.}

\item{WorkspaceDirectoryName}{The name of the directory to register.}

\item{WorkspaceDirectoryDescription}{Description of the directory to register.}

\item{UserIdentityType}{The type of identity management the user is using.}

\item{IdcInstanceArn}{The Amazon Resource Name (ARN) of the identity center instance.}

\item{MicrosoftEntraConfig}{The details about Microsoft Entra config.}

\item{WorkspaceType}{Indicates whether the directory's WorkSpace type is personal or pools.}

\item{ActiveDirectoryConfig}{The active directory config of the directory.}
}
\description{
Registers the specified directory. This operation is asynchronous and returns before the WorkSpace directory is registered. If this is the first time you are registering a directory, you will need to create the workspaces_DefaultRole role before you can register a directory. For more information, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/workspaces-access-control.html#create-default-role}{Creating the workspaces_DefaultRole Role}.

See \url{https://www.paws-r-sdk.com/docs/workspaces_register_workspace_directory/} for full documentation.
}
\keyword{internal}
