% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_reject_account_link_invitation}
\alias{workspaces_reject_account_link_invitation}
\title{Rejects the account link invitation}
\usage{
workspaces_reject_account_link_invitation(LinkId, ClientToken = NULL)
}
\arguments{
\item{LinkId}{[required] The identifier of the account link}

\item{ClientToken}{The client token of the account link invitation to reject.}
}
\description{
Rejects the account link invitation.

See \url{https://www.paws-r-sdk.com/docs/workspaces_reject_account_link_invitation/} for full documentation.
}
\keyword{internal}
