% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_restore_workspace}
\alias{workspaces_restore_workspace}
\title{Restores the specified WorkSpace to its last known healthy state}
\usage{
workspaces_restore_workspace(WorkspaceId)
}
\arguments{
\item{WorkspaceId}{[required] The identifier of the WorkSpace.}
}
\description{
Restores the specified WorkSpace to its last known healthy state.

See \url{https://paws-r.github.io/docs/workspaces/restore_workspace.html} for full documentation.
}
\keyword{internal}
