% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_revoke_ip_rules}
\alias{workspaces_revoke_ip_rules}
\title{Removes one or more rules from the specified IP access control group}
\usage{
workspaces_revoke_ip_rules(GroupId, UserRules)
}
\arguments{
\item{GroupId}{[required] The identifier of the group.}

\item{UserRules}{[required] The rules to remove from the group.}
}
\value{
An empty list.
}
\description{
Removes one or more rules from the specified IP access control group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$revoke_ip_rules(
  GroupId = "string",
  UserRules = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
