% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_start_workspaces}
\alias{workspaces_start_workspaces}
\title{Starts the specified WorkSpaces}
\usage{
workspaces_start_workspaces(StartWorkspaceRequests)
}
\arguments{
\item{StartWorkspaceRequests}{[required] The WorkSpaces to start. You can specify up to 25 WorkSpaces.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FailedRequests = list(
    list(
      WorkspaceId = "string",
      ErrorCode = "string",
      ErrorMessage = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Starts the specified WorkSpaces.

You cannot start a WorkSpace unless it has a running mode of \code{AutoStop}
and a state of \code{STOPPED}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_workspaces(
  StartWorkspaceRequests = list(
    list(
      WorkspaceId = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
