% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_stop_workspaces}
\alias{workspaces_stop_workspaces}
\title{Stops the specified WorkSpaces}
\usage{
workspaces_stop_workspaces(StopWorkspaceRequests)
}
\arguments{
\item{StopWorkspaceRequests}{[required] The WorkSpaces to stop. You can specify up to 25 WorkSpaces.}
}
\description{
Stops the specified WorkSpaces.

See \url{https://www.paws-r-sdk.com/docs/workspaces_stop_workspaces/} for full documentation.
}
\keyword{internal}
