% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_stop_workspaces_pool}
\alias{workspaces_stop_workspaces_pool}
\title{End of support notice: On December 31, 2027, Amazon Web Services will
end support for Amazon WorkSpaces Pools}
\usage{
workspaces_stop_workspaces_pool(PoolId)
}
\arguments{
\item{PoolId}{[required] The identifier of the pool.}
}
\description{
End of support notice: On December 31, 2027, Amazon Web Services will end support for Amazon WorkSpaces Pools. After December 31, 2027, you will no longer be able to access the Amazon WorkSpaces Pools console or Amazon WorkSpaces Pools resources. For more information, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/wsp-pools-end-of-support.html}{Amazon WorkSpaces Pools end of support}.

See \url{https://www.paws-r-sdk.com/docs/workspaces_stop_workspaces_pool/} for full documentation.
}
\keyword{internal}
