% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_terminate_workspaces}
\alias{workspaces_terminate_workspaces}
\title{Terminates the specified WorkSpaces}
\usage{
workspaces_terminate_workspaces(TerminateWorkspaceRequests)
}
\arguments{
\item{TerminateWorkspaceRequests}{[required] The WorkSpaces to terminate. You can specify up to 25 WorkSpaces.}
}
\description{
Terminates the specified WorkSpaces.

See \url{https://paws-r.github.io/docs/workspaces/terminate_workspaces.html} for full documentation.
}
\keyword{internal}
