% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_update_connect_client_add_in}
\alias{workspaces_update_connect_client_add_in}
\title{Updates a Connect Customer client add-in}
\usage{
workspaces_update_connect_client_add_in(
  AddInId,
  ResourceId,
  Name = NULL,
  URL = NULL
)
}
\arguments{
\item{AddInId}{[required] The identifier of the client add-in to update.}

\item{ResourceId}{[required] The directory identifier for which the client add-in is configured.}

\item{Name}{The name of the client add-in.}

\item{URL}{The endpoint URL of the Connect Customer client add-in.}
}
\description{
Updates a Connect Customer client add-in. Use this action to update the name and endpoint URL of a Connect Customer client add-in.

See \url{https://www.paws-r-sdk.com/docs/workspaces_update_connect_client_add_in/} for full documentation.
}
\keyword{internal}
