% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_update_workspace_bundle}
\alias{workspaces_update_workspace_bundle}
\title{Updates a WorkSpace bundle with a new image}
\usage{
workspaces_update_workspace_bundle(BundleId, ImageId)
}
\arguments{
\item{BundleId}{The identifier of the bundle.}

\item{ImageId}{The identifier of the image.}
}
\value{
An empty list.
}
\description{
Updates a WorkSpace bundle with a new image. For more information about
updating WorkSpace bundles, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/update-custom-bundle.html}{Update a Custom WorkSpaces Bundle}.

Existing WorkSpaces aren't automatically updated when you update the
bundle that they're based on. To update existing WorkSpaces that are
based on a bundle that you've updated, you must either rebuild the
WorkSpaces or delete and recreate them.
}
\section{Request syntax}{
\preformatted{svc$update_workspace_bundle(
  BundleId = "string",
  ImageId = "string"
)
}
}

\keyword{internal}
