% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_update_workspaces_pool}
\alias{workspaces_update_workspaces_pool}
\title{End of support notice: On December 31, 2027, Amazon Web Services will
end support for Amazon WorkSpaces Pools}
\usage{
workspaces_update_workspaces_pool(
  PoolId,
  Description = NULL,
  BundleId = NULL,
  DirectoryId = NULL,
  Capacity = NULL,
  ApplicationSettings = NULL,
  TimeoutSettings = NULL,
  RunningMode = NULL
)
}
\arguments{
\item{PoolId}{[required] The identifier of the specified pool to update.}

\item{Description}{Describes the specified pool to update.}

\item{BundleId}{The identifier of the bundle.}

\item{DirectoryId}{The identifier of the directory.}

\item{Capacity}{The desired capacity for the pool.}

\item{ApplicationSettings}{The persistent application settings for users in the pool.}

\item{TimeoutSettings}{Indicates the timeout settings of the specified pool.}

\item{RunningMode}{The desired running mode for the pool. The running mode can only be updated when the pool is in a stopped state.}
}
\description{
End of support notice: On December 31, 2027, Amazon Web Services will end support for Amazon WorkSpaces Pools. After December 31, 2027, you will no longer be able to access the Amazon WorkSpaces Pools console or Amazon WorkSpaces Pools resources. For more information, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/wsp-pools-end-of-support.html}{Amazon WorkSpaces Pools end of support}.

See \url{https://www.paws-r-sdk.com/docs/workspaces_update_workspaces_pool/} for full documentation.
}
\keyword{internal}
