% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_associate_browser_settings}
\alias{workspacesweb_associate_browser_settings}
\title{Associates a browser settings resource with a web portal}
\usage{
workspacesweb_associate_browser_settings(portalArn, browserSettingsArn)
}
\arguments{
\item{portalArn}{[required] The ARN of the web portal.}

\item{browserSettingsArn}{[required] The ARN of the browser settings.}
}
\description{
Associates a browser settings resource with a web portal.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_associate_browser_settings/} for full documentation.
}
\keyword{internal}
