% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_associate_data_protection_settings}
\alias{workspacesweb_associate_data_protection_settings}
\title{Associates a data protection settings resource with a web portal}
\usage{
workspacesweb_associate_data_protection_settings(
  portalArn,
  dataProtectionSettingsArn
)
}
\arguments{
\item{portalArn}{[required] The ARN of the web portal.}

\item{dataProtectionSettingsArn}{[required] The ARN of the data protection settings.}
}
\description{
Associates a data protection settings resource with a web portal.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_associate_data_protection_settings/} for full documentation.
}
\keyword{internal}
