% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_associate_ip_access_settings}
\alias{workspacesweb_associate_ip_access_settings}
\title{Associates an IP access settings resource with a web portal}
\usage{
workspacesweb_associate_ip_access_settings(portalArn, ipAccessSettingsArn)
}
\arguments{
\item{portalArn}{[required] The ARN of the web portal.}

\item{ipAccessSettingsArn}{[required] The ARN of the IP access settings.}
}
\description{
Associates an IP access settings resource with a web portal.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_associate_ip_access_settings/} for full documentation.
}
\keyword{internal}
