% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_associate_network_settings}
\alias{workspacesweb_associate_network_settings}
\title{Associates a network settings resource with a web portal}
\usage{
workspacesweb_associate_network_settings(networkSettingsArn, portalArn)
}
\arguments{
\item{networkSettingsArn}{[required] The ARN of the network settings.}

\item{portalArn}{[required] The ARN of the web portal.}
}
\description{
Associates a network settings resource with a web portal.

See \url{https://paws-r.github.io/docs/workspacesweb/associate_network_settings.html} for full documentation.
}
\keyword{internal}
