% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_associate_session_logger}
\alias{workspacesweb_associate_session_logger}
\title{Associates a session logger with a portal}
\usage{
workspacesweb_associate_session_logger(portalArn, sessionLoggerArn)
}
\arguments{
\item{portalArn}{[required] The ARN of the portal to associate to the session logger ARN.}

\item{sessionLoggerArn}{[required] The ARN of the session logger to associate to the portal ARN.}
}
\description{
Associates a session logger with a portal.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_associate_session_logger/} for full documentation.
}
\keyword{internal}
