% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_associate_trust_store}
\alias{workspacesweb_associate_trust_store}
\title{Associates a trust store with a web portal}
\usage{
workspacesweb_associate_trust_store(portalArn, trustStoreArn)
}
\arguments{
\item{portalArn}{[required] The ARN of the web portal.}

\item{trustStoreArn}{[required] The ARN of the trust store.}
}
\description{
Associates a trust store with a web portal.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_associate_trust_store/} for full documentation.
}
\keyword{internal}
