% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_associate_user_access_logging_settings}
\alias{workspacesweb_associate_user_access_logging_settings}
\title{Associates a user access logging settings resource with a web portal}
\usage{
workspacesweb_associate_user_access_logging_settings(
  portalArn,
  userAccessLoggingSettingsArn
)
}
\arguments{
\item{portalArn}{[required] The ARN of the web portal.}

\item{userAccessLoggingSettingsArn}{[required] The ARN of the user access logging settings.}
}
\description{
Associates a user access logging settings resource with a web portal.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_associate_user_access_logging_settings/} for full documentation.
}
\keyword{internal}
