% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_associate_user_settings}
\alias{workspacesweb_associate_user_settings}
\title{Associates a user settings resource with a web portal}
\usage{
workspacesweb_associate_user_settings(portalArn, userSettingsArn)
}
\arguments{
\item{portalArn}{[required] The ARN of the web portal.}

\item{userSettingsArn}{[required] The ARN of the user settings.}
}
\description{
Associates a user settings resource with a web portal.

See \url{https://paws-r.github.io/docs/workspacesweb/associate_user_settings.html} for full documentation.
}
\keyword{internal}
