% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_create_portal}
\alias{workspacesweb_create_portal}
\title{Creates a web portal}
\usage{
workspacesweb_create_portal(
  additionalEncryptionContext = NULL,
  clientToken = NULL,
  customerManagedKey = NULL,
  displayName = NULL,
  tags = NULL
)
}
\arguments{
\item{additionalEncryptionContext}{The additional encryption context of the portal.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Idempotency ensures that an API request
completes only once. With an idempotent request, if the original request
completes successfully, subsequent retries with the same client token
returns the result from the original successful request.

If you do not specify a client token, one is automatically generated by
the AWS SDK.}

\item{customerManagedKey}{The customer managed key of the web portal.}

\item{displayName}{The name of the web portal. This is not visible to users who log into
the web portal.}

\item{tags}{The tags to add to the web portal. A tag is a key-value pair.}
}
\description{
Creates a web portal.

See \url{https://paws-r.github.io/docs/workspacesweb/create_portal.html} for full documentation.
}
\keyword{internal}
