% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_create_session_logger}
\alias{workspacesweb_create_session_logger}
\title{Creates a session logger}
\usage{
workspacesweb_create_session_logger(
  eventFilter,
  logConfiguration,
  displayName = NULL,
  customerManagedKey = NULL,
  additionalEncryptionContext = NULL,
  tags = NULL,
  clientToken = NULL
)
}
\arguments{
\item{eventFilter}{[required] The filter that specifies the events to monitor.}

\item{logConfiguration}{[required] The configuration that specifies where logs are delivered.}

\item{displayName}{The human-readable display name for the session logger resource.}

\item{customerManagedKey}{The custom managed key of the session logger.}

\item{additionalEncryptionContext}{The additional encryption context of the session logger.}

\item{tags}{The tags to add to the session logger.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Idempotency ensures that an API request
completes only once. With an idempotent request, if the original request
completes successfully, subsequent retries with the same client token
returns the result from the original successful request. If you do not
specify a client token, one is automatically generated by the AWS SDK.}
}
\description{
Creates a session logger.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_create_session_logger/} for full documentation.
}
\keyword{internal}
