% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_create_user_access_logging_settings}
\alias{workspacesweb_create_user_access_logging_settings}
\title{Creates a user access logging settings resource that can be associated
with a web portal}
\usage{
workspacesweb_create_user_access_logging_settings(
  clientToken = NULL,
  kinesisStreamArn,
  tags = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Idempotency ensures that an API request
completes only once. With an idempotent request, if the original request
completes successfully, subsequent retries with the same client token
returns the result from the original successful request.

If you do not specify a client token, one is automatically generated by
the AWS SDK.}

\item{kinesisStreamArn}{[required] The ARN of the Kinesis stream.}

\item{tags}{The tags to add to the user settings resource. A tag is a key-value
pair.}
}
\description{
Creates a user access logging settings resource that can be associated with a web portal.

See \url{https://paws-r.github.io/docs/workspacesweb/create_user_access_logging_settings.html} for full documentation.
}
\keyword{internal}
