% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_delete_browser_settings}
\alias{workspacesweb_delete_browser_settings}
\title{Deletes browser settings}
\usage{
workspacesweb_delete_browser_settings(browserSettingsArn)
}
\arguments{
\item{browserSettingsArn}{[required] The ARN of the browser settings.}
}
\description{
Deletes browser settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_delete_browser_settings/} for full documentation.
}
\keyword{internal}
