% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_delete_identity_provider}
\alias{workspacesweb_delete_identity_provider}
\title{Deletes the identity provider}
\usage{
workspacesweb_delete_identity_provider(identityProviderArn)
}
\arguments{
\item{identityProviderArn}{[required] The ARN of the identity provider.}
}
\description{
Deletes the identity provider.

See \url{https://paws-r.github.io/docs/workspacesweb/delete_identity_provider.html} for full documentation.
}
\keyword{internal}
