% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_delete_network_settings}
\alias{workspacesweb_delete_network_settings}
\title{Deletes network settings}
\usage{
workspacesweb_delete_network_settings(networkSettingsArn)
}
\arguments{
\item{networkSettingsArn}{[required] The ARN of the network settings.}
}
\description{
Deletes network settings.

See \url{https://paws-r.github.io/docs/workspacesweb/delete_network_settings.html} for full documentation.
}
\keyword{internal}
