% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_delete_portal}
\alias{workspacesweb_delete_portal}
\title{Deletes a web portal}
\usage{
workspacesweb_delete_portal(portalArn)
}
\arguments{
\item{portalArn}{[required] The ARN of the web portal.}
}
\description{
Deletes a web portal.

See \url{https://paws-r.github.io/docs/workspacesweb/delete_portal.html} for full documentation.
}
\keyword{internal}
