% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_delete_trust_store}
\alias{workspacesweb_delete_trust_store}
\title{Deletes the trust store}
\usage{
workspacesweb_delete_trust_store(trustStoreArn)
}
\arguments{
\item{trustStoreArn}{[required] The ARN of the trust store.}
}
\description{
Deletes the trust store.

See \url{https://paws-r.github.io/docs/workspacesweb/delete_trust_store.html} for full documentation.
}
\keyword{internal}
