% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_disassociate_session_logger}
\alias{workspacesweb_disassociate_session_logger}
\title{Disassociates a session logger from a portal}
\usage{
workspacesweb_disassociate_session_logger(portalArn)
}
\arguments{
\item{portalArn}{[required] The ARN of the portal to disassociate from the a session logger.}
}
\description{
Disassociates a session logger from a portal.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_disassociate_session_logger/} for full documentation.
}
\keyword{internal}
