% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_disassociate_user_access_logging_settings}
\alias{workspacesweb_disassociate_user_access_logging_settings}
\title{Disassociates user access logging settings from a web portal}
\usage{
workspacesweb_disassociate_user_access_logging_settings(portalArn)
}
\arguments{
\item{portalArn}{[required] The ARN of the web portal.}
}
\description{
Disassociates user access logging settings from a web portal.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_disassociate_user_access_logging_settings/} for full documentation.
}
\keyword{internal}
