% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_expire_session}
\alias{workspacesweb_expire_session}
\title{Expires an active secure browser session}
\usage{
workspacesweb_expire_session(portalId, sessionId)
}
\arguments{
\item{portalId}{[required] The ID of the web portal for the session.}

\item{sessionId}{[required] The ID of the session to expire.}
}
\description{
Expires an active secure browser session.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_expire_session/} for full documentation.
}
\keyword{internal}
