% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_data_protection_settings}
\alias{workspacesweb_get_data_protection_settings}
\title{Gets the data protection settings}
\usage{
workspacesweb_get_data_protection_settings(dataProtectionSettingsArn)
}
\arguments{
\item{dataProtectionSettingsArn}{[required] The ARN of the data protection settings.}
}
\description{
Gets the data protection settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_data_protection_settings/} for full documentation.
}
\keyword{internal}
