% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_ip_access_settings}
\alias{workspacesweb_get_ip_access_settings}
\title{Gets the IP access settings}
\usage{
workspacesweb_get_ip_access_settings(ipAccessSettingsArn)
}
\arguments{
\item{ipAccessSettingsArn}{[required] The ARN of the IP access settings.}
}
\description{
Gets the IP access settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_ip_access_settings/} for full documentation.
}
\keyword{internal}
