% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_network_settings}
\alias{workspacesweb_get_network_settings}
\title{Gets the network settings}
\usage{
workspacesweb_get_network_settings(networkSettingsArn)
}
\arguments{
\item{networkSettingsArn}{[required] The ARN of the network settings.}
}
\description{
Gets the network settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_network_settings/} for full documentation.
}
\keyword{internal}
