% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_portal}
\alias{workspacesweb_get_portal}
\title{Gets the web portal}
\usage{
workspacesweb_get_portal(portalArn)
}
\arguments{
\item{portalArn}{[required] The ARN of the web portal.}
}
\description{
Gets the web portal.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_portal/} for full documentation.
}
\keyword{internal}
