% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_portal_service_provider_metadata}
\alias{workspacesweb_get_portal_service_provider_metadata}
\title{Gets the service provider metadata}
\usage{
workspacesweb_get_portal_service_provider_metadata(portalArn)
}
\arguments{
\item{portalArn}{[required] The ARN of the web portal.}
}
\description{
Gets the service provider metadata.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_portal_service_provider_metadata/} for full documentation.
}
\keyword{internal}
