% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_session}
\alias{workspacesweb_get_session}
\title{Gets information for a secure browser session}
\usage{
workspacesweb_get_session(portalId, sessionId)
}
\arguments{
\item{portalId}{[required] The ID of the web portal for the session.}

\item{sessionId}{[required] The ID of the session.}
}
\description{
Gets information for a secure browser session.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_session/} for full documentation.
}
\keyword{internal}
