% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_session_logger}
\alias{workspacesweb_get_session_logger}
\title{Gets details about a specific session logger resource}
\usage{
workspacesweb_get_session_logger(sessionLoggerArn)
}
\arguments{
\item{sessionLoggerArn}{[required] The ARN of the session logger.}
}
\description{
Gets details about a specific session logger resource.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_session_logger/} for full documentation.
}
\keyword{internal}
