% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_trust_store_certificate}
\alias{workspacesweb_get_trust_store_certificate}
\title{Gets the trust store certificate}
\usage{
workspacesweb_get_trust_store_certificate(thumbprint, trustStoreArn)
}
\arguments{
\item{thumbprint}{[required] The thumbprint of the trust store certificate.}

\item{trustStoreArn}{[required] The ARN of the trust store certificate.}
}
\description{
Gets the trust store certificate.

See \url{https://paws-r.github.io/docs/workspacesweb/get_trust_store_certificate.html} for full documentation.
}
\keyword{internal}
