% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_list_data_protection_settings}
\alias{workspacesweb_list_data_protection_settings}
\title{Retrieves a list of data protection settings}
\usage{
workspacesweb_list_data_protection_settings(
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to be included in the next page.}

\item{nextToken}{The pagination token used to retrieve the next page of results for this
operation.}
}
\description{
Retrieves a list of data protection settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_list_data_protection_settings/} for full documentation.
}
\keyword{internal}
