% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_list_session_loggers}
\alias{workspacesweb_list_session_loggers}
\title{Lists all available session logger resources}
\usage{
workspacesweb_list_session_loggers(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The pagination token used to retrieve the next page of results for this
operation.}

\item{maxResults}{The maximum number of results to be included in the next page.}
}
\description{
Lists all available session logger resources.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_list_session_loggers/} for full documentation.
}
\keyword{internal}
