% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_list_sessions}
\alias{workspacesweb_list_sessions}
\title{Lists information for multiple secure browser sessions from a specific
portal}
\usage{
workspacesweb_list_sessions(
  maxResults = NULL,
  nextToken = NULL,
  portalId,
  sessionId = NULL,
  sortBy = NULL,
  status = NULL,
  username = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to be included in the next page.}

\item{nextToken}{The pagination token used to retrieve the next page of results for this
operation.}

\item{portalId}{[required] The ID of the web portal for the sessions.}

\item{sessionId}{The ID of the session.}

\item{sortBy}{The method in which the returned sessions should be sorted.}

\item{status}{The status of the session.}

\item{username}{The username of the session.}
}
\description{
Lists information for multiple secure browser sessions from a specific portal.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_list_sessions/} for full documentation.
}
\keyword{internal}
