% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_list_tags_for_resource}
\alias{workspacesweb_list_tags_for_resource}
\title{Retrieves a list of tags for a resource}
\usage{
workspacesweb_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}
}
\description{
Retrieves a list of tags for a resource.

See \url{https://paws-r.github.io/docs/workspacesweb/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
