% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_list_trust_stores}
\alias{workspacesweb_list_trust_stores}
\title{Retrieves a list of trust stores}
\usage{
workspacesweb_list_trust_stores(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to be included in the next page.}

\item{nextToken}{The pagination token used to retrieve the next page of results for this
operation.}
}
\description{
Retrieves a list of trust stores.

See \url{https://paws-r.github.io/docs/workspacesweb/list_trust_stores.html} for full documentation.
}
\keyword{internal}
