% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_list_user_access_logging_settings}
\alias{workspacesweb_list_user_access_logging_settings}
\title{Retrieves a list of user access logging settings}
\usage{
workspacesweb_list_user_access_logging_settings(
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to be included in the next page.}

\item{nextToken}{The pagination token used to retrieve the next page of results for this
operation.}
}
\description{
Retrieves a list of user access logging settings.

See \url{https://paws-r.github.io/docs/workspacesweb/list_user_access_logging_settings.html} for full documentation.
}
\keyword{internal}
