% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_list_user_settings}
\alias{workspacesweb_list_user_settings}
\title{Retrieves a list of user settings}
\usage{
workspacesweb_list_user_settings(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The pagination token used to retrieve the next page of results for this
operation.}

\item{maxResults}{The maximum number of results to be included in the next page.}
}
\description{
Retrieves a list of user settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_list_user_settings/} for full documentation.
}
\keyword{internal}
