% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_tag_resource}
\alias{workspacesweb_tag_resource}
\title{Adds or overwrites one or more tags for the specified resource}
\usage{
workspacesweb_tag_resource(resourceArn, tags, clientToken = NULL)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tags}{[required] The tags of the resource.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Idempotency ensures that an API request
completes only once. With an idempotent request, if the original request
completes successfully, subsequent retries with the same client token
returns the result from the original successful request.

If you do not specify a client token, one is automatically generated by
the Amazon Web Services SDK.}
}
\description{
Adds or overwrites one or more tags for the specified resource.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_tag_resource/} for full documentation.
}
\keyword{internal}
