% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_untag_resource}
\alias{workspacesweb_untag_resource}
\title{Removes one or more tags from the specified resource}
\usage{
workspacesweb_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tagKeys}{[required] The list of tag keys to remove from the resource.}
}
\description{
Removes one or more tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_untag_resource/} for full documentation.
}
\keyword{internal}
