% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_update_browser_settings}
\alias{workspacesweb_update_browser_settings}
\title{Updates browser settings}
\usage{
workspacesweb_update_browser_settings(
  browserSettingsArn,
  browserPolicy = NULL,
  clientToken = NULL,
  webContentFilteringPolicy = NULL
)
}
\arguments{
\item{browserSettingsArn}{[required] The ARN of the browser settings.}

\item{browserPolicy}{A JSON string containing Chrome Enterprise policies that will be applied
to all streaming sessions.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Idempotency ensures that an API request
completes only once. With an idempotent request, if the original request
completes successfully, subsequent retries with the same client token
return the result from the original successful request.

If you do not specify a client token, one is automatically generated by
the Amazon Web Services SDK.}

\item{webContentFilteringPolicy}{The policy that specifies which URLs end users are allowed to access or
which URLs or domain categories they are restricted from accessing for
enhanced security.}
}
\description{
Updates browser settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_update_browser_settings/} for full documentation.
}
\keyword{internal}
