% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_update_data_protection_settings}
\alias{workspacesweb_update_data_protection_settings}
\title{Updates data protection settings}
\usage{
workspacesweb_update_data_protection_settings(
  dataProtectionSettingsArn,
  inlineRedactionConfiguration = NULL,
  displayName = NULL,
  description = NULL,
  clientToken = NULL
)
}
\arguments{
\item{dataProtectionSettingsArn}{[required] The ARN of the data protection settings.}

\item{inlineRedactionConfiguration}{The inline redaction configuration of the data protection settings that
will be applied to all sessions.}

\item{displayName}{The display name of the data protection settings.}

\item{description}{The description of the data protection settings.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Idempotency ensures that an API request
completes only once. With an idempotent request, if the original request
completes successfully, subsequent retries with the same client token
return the result from the original successful request.

If you do not specify a client token, one is automatically generated by
the Amazon Web Services SDK.}
}
\description{
Updates data protection settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_update_data_protection_settings/} for full documentation.
}
\keyword{internal}
