% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_update_ip_access_settings}
\alias{workspacesweb_update_ip_access_settings}
\title{Updates IP access settings}
\usage{
workspacesweb_update_ip_access_settings(
  ipAccessSettingsArn,
  displayName = NULL,
  description = NULL,
  ipRules = NULL,
  clientToken = NULL
)
}
\arguments{
\item{ipAccessSettingsArn}{[required] The ARN of the IP access settings.}

\item{displayName}{The display name of the IP access settings.}

\item{description}{The description of the IP access settings.}

\item{ipRules}{The updated IP rules of the IP access settings.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Idempotency ensures that an API request
completes only once. With an idempotent request, if the original request
completes successfully, subsequent retries with the same client token
return the result from the original successful request.

If you do not specify a client token, one is automatically generated by
the Amazon Web Services SDK.}
}
\description{
Updates IP access settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_update_ip_access_settings/} for full documentation.
}
\keyword{internal}
