% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_update_portal}
\alias{workspacesweb_update_portal}
\title{Updates a web portal}
\usage{
workspacesweb_update_portal(displayName = NULL, portalArn)
}
\arguments{
\item{displayName}{The name of the web portal. This is not visible to users who log into
the web portal.}

\item{portalArn}{[required] The ARN of the web portal.}
}
\description{
Updates a web portal.

See \url{https://paws-r.github.io/docs/workspacesweb/update_portal.html} for full documentation.
}
\keyword{internal}
