% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_update_session_logger}
\alias{workspacesweb_update_session_logger}
\title{Updates the details of a session logger}
\usage{
workspacesweb_update_session_logger(
  sessionLoggerArn,
  eventFilter = NULL,
  logConfiguration = NULL,
  displayName = NULL
)
}
\arguments{
\item{sessionLoggerArn}{[required] The ARN of the session logger to update.}

\item{eventFilter}{The updated eventFilter.}

\item{logConfiguration}{The updated logConfiguration.}

\item{displayName}{The updated display name.}
}
\description{
Updates the details of a session logger.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_update_session_logger/} for full documentation.
}
\keyword{internal}
