% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_update_trust_store}
\alias{workspacesweb_update_trust_store}
\title{Updates the trust store}
\usage{
workspacesweb_update_trust_store(
  trustStoreArn,
  certificatesToAdd = NULL,
  certificatesToDelete = NULL,
  clientToken = NULL
)
}
\arguments{
\item{trustStoreArn}{[required] The ARN of the trust store.}

\item{certificatesToAdd}{A list of CA certificates to add to the trust store.}

\item{certificatesToDelete}{A list of CA certificates to delete from a trust store.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Idempotency ensures that an API request
completes only once. With an idempotent request, if the original request
completes successfully, subsequent retries with the same client token
return the result from the original successful request.

If you do not specify a client token, one is automatically generated by
the Amazon Web Services SDK.}
}
\description{
Updates the trust store.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_update_trust_store/} for full documentation.
}
\keyword{internal}
