% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_associate_application_fleet}
\alias{appstream_associate_application_fleet}
\title{Associates the specified application with the specified fleet}
\usage{
appstream_associate_application_fleet(FleetName, ApplicationArn)
}
\arguments{
\item{FleetName}{[required] The name of the fleet.}

\item{ApplicationArn}{[required] The ARN of the application.}
}
\description{
Associates the specified application with the specified fleet. This is only supported for Elastic fleets.

See \url{https://www.paws-r-sdk.com/docs/appstream_associate_application_fleet/} for full documentation.
}
\keyword{internal}
