% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_associate_application_to_entitlement}
\alias{appstream_associate_application_to_entitlement}
\title{Associates an application to entitle}
\usage{
appstream_associate_application_to_entitlement(
  StackName,
  EntitlementName,
  ApplicationIdentifier
)
}
\arguments{
\item{StackName}{[required] The name of the stack.}

\item{EntitlementName}{[required] The name of the entitlement.}

\item{ApplicationIdentifier}{[required] The identifier of the application.}
}
\description{
Associates an application to entitle.

See \url{https://www.paws-r-sdk.com/docs/appstream_associate_application_to_entitlement/} for full documentation.
}
\keyword{internal}
