% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_batch_disassociate_user_stack}
\alias{appstream_batch_disassociate_user_stack}
\title{Disassociates the specified users from the specified stacks}
\usage{
appstream_batch_disassociate_user_stack(UserStackAssociations)
}
\arguments{
\item{UserStackAssociations}{[required] The list of UserStackAssociation objects.}
}
\description{
Disassociates the specified users from the specified stacks.

See \url{https://paws-r.github.io/docs/appstream/batch_disassociate_user_stack.html} for full documentation.
}
\keyword{internal}
