% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_directory_config}
\alias{appstream_delete_directory_config}
\title{Deletes the specified Directory Config object from AppStream 2}
\usage{
appstream_delete_directory_config(DirectoryName)
}
\arguments{
\item{DirectoryName}{[required] The name of the directory configuration.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes the specified Directory Config object from AppStream 2.0. This
object includes the information required to join streaming instances to
an Active Directory domain.
}
\section{Request syntax}{
\preformatted{svc$delete_directory_config(
  DirectoryName = "string"
)
}
}

\keyword{internal}
