% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_directory_config}
\alias{appstream_delete_directory_config}
\title{Deletes the specified Directory Config object from WorkSpaces
Applications}
\usage{
appstream_delete_directory_config(DirectoryName)
}
\arguments{
\item{DirectoryName}{[required] The name of the directory configuration.}
}
\description{
Deletes the specified Directory Config object from WorkSpaces Applications. This object includes the information required to join streaming instances to an Active Directory domain.

See \url{https://www.paws-r-sdk.com/docs/appstream_delete_directory_config/} for full documentation.
}
\keyword{internal}
