% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_usage_report_subscription}
\alias{appstream_delete_usage_report_subscription}
\title{Disables usage report generation}
\usage{
appstream_delete_usage_report_subscription()
}
\value{
An empty list.
}
\description{
Disables usage report generation.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_usage_report_subscription()
}\if{html}{\out{</div>}}
}

\keyword{internal}
