% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_disable_user}
\alias{appstream_disable_user}
\title{Disables the specified user in the user pool}
\usage{
appstream_disable_user(UserName, AuthenticationType)
}
\arguments{
\item{UserName}{[required] The email address of the user.

Users' email addresses are case-sensitive.}

\item{AuthenticationType}{[required] The authentication type for the user. You must specify USERPOOL.}
}
\value{
An empty list.
}
\description{
Disables the specified user in the user pool. Users can't sign in to
AppStream 2.0 until they are re-enabled. This action does not delete the
user.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disable_user(
  UserName = "string",
  AuthenticationType = "API"|"SAML"|"USERPOOL"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
