% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_disable_user}
\alias{appstream_disable_user}
\title{Disables the specified user in the user pool}
\usage{
appstream_disable_user(UserName, AuthenticationType)
}
\arguments{
\item{UserName}{[required] The email address of the user.

Users' email addresses are case-sensitive.}

\item{AuthenticationType}{[required] The authentication type for the user. You must specify USERPOOL.}
}
\description{
Disables the specified user in the user pool. Users can't sign in to WorkSpaces Applications until they are re-enabled. This action does not delete the user.

See \url{https://www.paws-r-sdk.com/docs/appstream_disable_user/} for full documentation.
}
\keyword{internal}
