% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_list_associated_stacks}
\alias{appstream_list_associated_stacks}
\title{Retrieves the name of the stack with which the specified fleet is
associated}
\usage{
appstream_list_associated_stacks(FleetName, NextToken = NULL)
}
\arguments{
\item{FleetName}{[required] The name of the fleet.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\description{
Retrieves the name of the stack with which the specified fleet is associated.

See \url{https://paws-r.github.io/docs/appstream/list_associated_stacks.html} for full documentation.
}
\keyword{internal}
