% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_list_tags_for_resource}
\alias{appstream_list_tags_for_resource}
\title{Retrieves a list of all tags for the specified WorkSpaces Applications
resource}
\usage{
appstream_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Retrieves a list of all tags for the specified WorkSpaces Applications resource. You can tag WorkSpaces Applications image builders, images, fleets, and stacks.

See \url{https://www.paws-r-sdk.com/docs/appstream_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
