% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_start_fleet}
\alias{appstream_start_fleet}
\title{Starts the specified fleet}
\usage{
appstream_start_fleet(Name)
}
\arguments{
\item{Name}{[required] The name of the fleet.}
}
\description{
Starts the specified fleet.

See \url{https://paws-r.github.io/docs/appstream/start_fleet.html} for full documentation.
}
\keyword{internal}
