% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_start_fleet}
\alias{appstream_start_fleet}
\title{Starts the specified fleet}
\usage{
appstream_start_fleet(Name)
}
\arguments{
\item{Name}{[required] The name of the fleet.}
}
\value{
An empty list.
}
\description{
Starts the specified fleet.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_fleet(
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
