% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_stop_fleet}
\alias{appstream_stop_fleet}
\title{Stops the specified fleet}
\usage{
appstream_stop_fleet(Name)
}
\arguments{
\item{Name}{[required] The name of the fleet.}
}
\value{
An empty list.
}
\description{
Stops the specified fleet.
}
\section{Request syntax}{
\preformatted{svc$stop_fleet(
  Name = "string"
)
}
}

\keyword{internal}
