% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_tag_resource}
\alias{appstream_tag_resource}
\title{Adds or overwrites one or more tags for the specified WorkSpaces
Applications resource}
\usage{
appstream_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{Tags}{[required] The tags to associate. A tag is a key-value pair, and the value is
optional. For example, Environment=Test. If you do not specify a value,
Environment=.

If you do not specify a value, the value is set to an empty string.

Generally allowed characters are: letters, numbers, and spaces
representable in UTF-8, and the following special characters:

_ . : / = + \\ - @}
}
\description{
Adds or overwrites one or more tags for the specified WorkSpaces Applications resource. You can tag WorkSpaces Applications image builders, images, fleets, and stacks.

See \url{https://www.paws-r-sdk.com/docs/appstream_tag_resource/} for full documentation.
}
\keyword{internal}
