% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivschat_operations.R
\name{ivschat_delete_logging_configuration}
\alias{ivschat_delete_logging_configuration}
\title{Deletes the specified logging configuration}
\usage{
ivschat_delete_logging_configuration(identifier)
}
\arguments{
\item{identifier}{[required] Identifier of the logging configuration to be deleted.}
}
\description{
Deletes the specified logging configuration.

See \url{https://www.paws-r-sdk.com/docs/ivschat_delete_logging_configuration/} for full documentation.
}
\keyword{internal}
