% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivschat_operations.R
\name{ivschat_delete_room}
\alias{ivschat_delete_room}
\title{Deletes the specified room}
\usage{
ivschat_delete_room(identifier)
}
\arguments{
\item{identifier}{[required] Identifier of the room to be deleted. Currently this must be an ARN.}
}
\description{
Deletes the specified room.

See \url{https://paws-r.github.io/docs/ivschat/delete_room.html} for full documentation.
}
\keyword{internal}
