% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivschat_operations.R
\name{ivschat_list_logging_configurations}
\alias{ivschat_list_logging_configurations}
\title{Gets summary information about all your logging configurations in the
AWS region where the API request is processed}
\usage{
ivschat_list_logging_configurations(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{Maximum number of logging configurations to return. Default: 50.}

\item{nextToken}{The first logging configurations to retrieve. This is used for
pagination; see the \code{nextToken} response field.}
}
\description{
Gets summary information about all your logging configurations in the AWS region where the API request is processed.

See \url{https://www.paws-r-sdk.com/docs/ivschat_list_logging_configurations/} for full documentation.
}
\keyword{internal}
