% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivschat_operations.R
\name{ivschat_tag_resource}
\alias{ivschat_tag_resource}
\title{Adds or updates tags for the AWS resource with the specified ARN}
\usage{
ivschat_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to be tagged. The ARN must be URL-encoded.}

\item{tags}{[required] Array of tags to be added or updated. Array of maps, each of the form
\code{string:string (key:value)}. See \href{https://docs.aws.amazon.com/tag-editor/latest/userguide/best-practices-and-strats.html}{Best practices and strategies}
in \emph{Tagging Amazon Web Services Resources and Tag Editor} for details,
including restrictions that apply to tags and "Tag naming limits and
requirements"; Amazon IVS Chat has no constraints beyond what is
documented there.}
}
\description{
Adds or updates tags for the AWS resource with the specified ARN.

See \url{https://www.paws-r-sdk.com/docs/ivschat_tag_resource/} for full documentation.
}
\keyword{internal}
