% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_delete_launch_profile}
\alias{nimblestudio_delete_launch_profile}
\title{Permanently delete a launch profile}
\usage{
nimblestudio_delete_launch_profile(
  clientToken = NULL,
  launchProfileId,
  studioId
)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don’t specify a client token, the
Amazon Web Services SDK automatically generates a client token and uses
it for the request to ensure idempotency.}

\item{launchProfileId}{[required] The ID of the launch profile used to control access from the streaming
session.}

\item{studioId}{[required] The studio ID.}
}
\description{
Permanently delete a launch profile.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_delete_launch_profile/} for full documentation.
}
\keyword{internal}
