% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_launch_profile}
\alias{nimblestudio_get_launch_profile}
\title{Get a launch profile}
\usage{
nimblestudio_get_launch_profile(launchProfileId, studioId)
}
\arguments{
\item{launchProfileId}{[required] The Launch Profile ID.}

\item{studioId}{[required] The studio ID.}
}
\description{
Get a launch profile.

See \url{https://paws-r.github.io/docs/nimblestudio/get_launch_profile.html} for full documentation.
}
\keyword{internal}
