% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_launch_profile_initialization}
\alias{nimblestudio_get_launch_profile_initialization}
\title{Get a launch profile initialization}
\usage{
nimblestudio_get_launch_profile_initialization(
  launchProfileId,
  launchProfileProtocolVersions,
  launchPurpose,
  platform,
  studioId
)
}
\arguments{
\item{launchProfileId}{[required] The ID of the launch profile used to control access from the streaming
session.}

\item{launchProfileProtocolVersions}{[required] The launch profile protocol versions supported by the client.}

\item{launchPurpose}{[required] The launch purpose.}

\item{platform}{[required] The platform where this Launch Profile will be used, either Windows or
Linux.}

\item{studioId}{[required] The studio ID.}
}
\description{
Get a launch profile initialization.

See \url{https://paws-r.github.io/docs/nimblestudio/get_launch_profile_initialization.html} for full documentation.
}
\keyword{internal}
