% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_streaming_image}
\alias{nimblestudio_get_streaming_image}
\title{Get streaming image}
\usage{
nimblestudio_get_streaming_image(streamingImageId, studioId)
}
\arguments{
\item{streamingImageId}{[required] The streaming image ID.}

\item{studioId}{[required] The studio ID.}
}
\description{
Get streaming image.

See \url{https://paws-r.github.io/docs/nimblestudio/get_streaming_image.html} for full documentation.
}
\keyword{internal}
