% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_get_studio}
\alias{nimblestudio_get_studio}
\title{Get a studio resource}
\usage{
nimblestudio_get_studio(studioId)
}
\arguments{
\item{studioId}{[required] The studio ID.}
}
\description{
Get a studio resource.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_get_studio/} for full documentation.
}
\keyword{internal}
