% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_list_eula_acceptances}
\alias{nimblestudio_list_eula_acceptances}
\title{List EULA acceptances}
\usage{
nimblestudio_list_eula_acceptances(eulaIds = NULL, nextToken = NULL, studioId)
}
\arguments{
\item{eulaIds}{The list of EULA IDs that have been previously accepted.}

\item{nextToken}{The token for the next set of results, or null if there are no more
results.}

\item{studioId}{[required] The studio ID.}
}
\description{
List EULA acceptances.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_list_eula_acceptances/} for full documentation.
}
\keyword{internal}
