% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_list_streaming_images}
\alias{nimblestudio_list_streaming_images}
\title{List the streaming image resources available to this studio}
\usage{
nimblestudio_list_streaming_images(nextToken = NULL, owner = NULL, studioId)
}
\arguments{
\item{nextToken}{The token for the next set of results, or null if there are no more
results.}

\item{owner}{Filter this request to streaming images with the given owner}

\item{studioId}{[required] The studio ID.}
}
\description{
List the streaming image resources available to this studio.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_list_streaming_images/} for full documentation.
}
\keyword{internal}
