% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_list_studios}
\alias{nimblestudio_list_studios}
\title{List studios in your Amazon Web Services account in the requested Amazon
Web Services Region}
\usage{
nimblestudio_list_studios(nextToken = NULL)
}
\arguments{
\item{nextToken}{The token to request the next page of results.}
}
\description{
List studios in your Amazon Web Services account in the requested Amazon Web Services Region.

See \url{https://paws-r.github.io/docs/nimblestudio/list_studios.html} for full documentation.
}
\keyword{internal}
