% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_list_tags_for_resource}
\alias{nimblestudio_list_tags_for_resource}
\title{Gets the tags for a resource, given its Amazon Resource Names (ARN)}
\usage{
nimblestudio_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which you want to
list tags.}
}
\description{
Gets the tags for a resource, given its Amazon Resource Names (ARN).

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
