% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimblestudio_operations.R
\name{nimblestudio_update_launch_profile_member}
\alias{nimblestudio_update_launch_profile_member}
\title{Update a user persona in launch profile membership}
\usage{
nimblestudio_update_launch_profile_member(
  clientToken = NULL,
  launchProfileId,
  persona,
  principalId,
  studioId
)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don’t specify a client token, the
Amazon Web Services SDK automatically generates a client token and uses
it for the request to ensure idempotency.}

\item{launchProfileId}{[required] The ID of the launch profile used to control access from the streaming
session.}

\item{persona}{[required] The persona.}

\item{principalId}{[required] The principal ID. This currently supports a IAM Identity Center UserId.}

\item{studioId}{[required] The studio ID.}
}
\description{
Update a user persona in launch profile membership.

See \url{https://www.paws-r-sdk.com/docs/nimblestudio_update_launch_profile_member/} for full documentation.
}
\keyword{internal}
