% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_add_resource_permissions}
\alias{workdocs_add_resource_permissions}
\title{Creates a set of permissions for the specified folder or document}
\usage{
workdocs_add_resource_permissions(
  AuthenticationToken = NULL,
  ResourceId,
  Principals,
  NotificationOptions = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{ResourceId}{[required] The ID of the resource.}

\item{Principals}{[required] The users, groups, or organization being granted permission.}

\item{NotificationOptions}{The notification options.}
}
\description{
Creates a set of permissions for the specified folder or document. The resource permissions are overwritten if the principals already have different permissions.

See \url{https://www.paws-r-sdk.com/docs/workdocs_add_resource_permissions/} for full documentation.
}
\keyword{internal}
