% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_create_folder}
\alias{workdocs_create_folder}
\title{Creates a folder with the specified name and parent folder}
\usage{
workdocs_create_folder(AuthenticationToken = NULL, Name = NULL, ParentFolderId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{Name}{The name of the new folder.}

\item{ParentFolderId}{[required] The ID of the parent folder.}
}
\description{
Creates a folder with the specified name and parent folder.

See \url{https://paws-r.github.io/docs/workdocs/create_folder.html} for full documentation.
}
\keyword{internal}
