% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_deactivate_user}
\alias{workdocs_deactivate_user}
\title{Deactivates the specified user, which revokes the user's access to
Amazon WorkDocs}
\usage{
workdocs_deactivate_user(UserId, AuthenticationToken = NULL)
}
\arguments{
\item{UserId}{[required] The ID of the user.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}
}
\description{
Deactivates the specified user, which revokes the user's access to Amazon WorkDocs.

See \url{https://paws-r.github.io/docs/workdocs/deactivate_user.html} for full documentation.
}
\keyword{internal}
