% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_delete_comment}
\alias{workdocs_delete_comment}
\title{Deletes the specified comment from the document version}
\usage{
workdocs_delete_comment(AuthenticationToken, DocumentId, VersionId,
  CommentId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{DocumentId}{[required] The ID of the document.}

\item{VersionId}{[required] The ID of the document version.}

\item{CommentId}{[required] The ID of the comment.}
}
\value{
An empty list.
}
\description{
Deletes the specified comment from the document version.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_comment(
  AuthenticationToken = "string",
  DocumentId = "string",
  VersionId = "string",
  CommentId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
