% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_delete_document}
\alias{workdocs_delete_document}
\title{Permanently deletes the specified document and its associated metadata}
\usage{
workdocs_delete_document(AuthenticationToken = NULL, DocumentId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{DocumentId}{[required] The ID of the document.}
}
\description{
Permanently deletes the specified document and its associated metadata.

See \url{https://www.paws-r-sdk.com/docs/workdocs_delete_document/} for full documentation.
}
\keyword{internal}
