% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_delete_folder}
\alias{workdocs_delete_folder}
\title{Permanently deletes the specified folder and its contents}
\usage{
workdocs_delete_folder(AuthenticationToken = NULL, FolderId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{FolderId}{[required] The ID of the folder.}
}
\description{
Permanently deletes the specified folder and its contents.

See \url{https://paws-r.github.io/docs/workdocs/delete_folder.html} for full documentation.
}
\keyword{internal}
