% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_delete_folder}
\alias{workdocs_delete_folder}
\title{Permanently deletes the specified folder and its contents}
\usage{
workdocs_delete_folder(AuthenticationToken = NULL, FolderId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{FolderId}{[required] The ID of the folder.}
}
\description{
Permanently deletes the specified folder and its contents.

See \url{https://www.paws-r-sdk.com/docs/workdocs_delete_folder/} for full documentation.
}
\keyword{internal}
