% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_delete_folder_contents}
\alias{workdocs_delete_folder_contents}
\title{Deletes the contents of the specified folder}
\usage{
workdocs_delete_folder_contents(AuthenticationToken = NULL, FolderId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{FolderId}{[required] The ID of the folder.}
}
\description{
Deletes the contents of the specified folder.

See \url{https://paws-r.github.io/docs/workdocs/delete_folder_contents.html} for full documentation.
}
\keyword{internal}
