% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_delete_labels}
\alias{workdocs_delete_labels}
\title{Deletes the specified list of labels from a resource}
\usage{
workdocs_delete_labels(ResourceId, AuthenticationToken, Labels,
  DeleteAll)
}
\arguments{
\item{ResourceId}{[required] The ID of the resource.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{Labels}{List of labels to delete from the resource.}

\item{DeleteAll}{Flag to request removal of all labels from the specified resource.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes the specified list of labels from a resource.
}
\section{Request syntax}{
\preformatted{svc$delete_labels(
  ResourceId = "string",
  AuthenticationToken = "string",
  Labels = list(
    "string"
  ),
  DeleteAll = TRUE|FALSE
)
}
}

\keyword{internal}
