% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_describe_comments}
\alias{workdocs_describe_comments}
\title{List all the comments for the specified document version}
\usage{
workdocs_describe_comments(
  AuthenticationToken = NULL,
  DocumentId,
  VersionId,
  Limit = NULL,
  Marker = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{DocumentId}{[required] The ID of the document.}

\item{VersionId}{[required] The ID of the document version.}

\item{Limit}{The maximum number of items to return.}

\item{Marker}{The marker for the next set of results. This marker was received from a
previous call.}
}
\description{
List all the comments for the specified document version.

See \url{https://paws-r.github.io/docs/workdocs/describe_comments.html} for full documentation.
}
\keyword{internal}
