% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_get_current_user}
\alias{workdocs_get_current_user}
\title{Retrieves details of the current user for whom the authentication token
was generated}
\usage{
workdocs_get_current_user(AuthenticationToken)
}
\arguments{
\item{AuthenticationToken}{[required] Amazon WorkDocs authentication token.}
}
\description{
Retrieves details of the current user for whom the authentication token was generated. This is not a valid action for SigV4 (administrative API) clients.

See \url{https://www.paws-r-sdk.com/docs/workdocs_get_current_user/} for full documentation.
}
\keyword{internal}
