% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_remove_all_resource_permissions}
\alias{workdocs_remove_all_resource_permissions}
\title{Removes all the permissions from the specified resource}
\usage{
workdocs_remove_all_resource_permissions(
  AuthenticationToken = NULL,
  ResourceId
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{ResourceId}{[required] The ID of the resource.}
}
\description{
Removes all the permissions from the specified resource.

See \url{https://paws-r.github.io/docs/workdocs/remove_all_resource_permissions.html} for full documentation.
}
\keyword{internal}
