% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_remove_resource_permission}
\alias{workdocs_remove_resource_permission}
\title{Removes the permission for the specified principal from the specified
resource}
\usage{
workdocs_remove_resource_permission(AuthenticationToken, ResourceId,
  PrincipalId, PrincipalType)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using AWS
administrator credentials to access the API.}

\item{ResourceId}{[required] The ID of the resource.}

\item{PrincipalId}{[required] The principal ID of the resource.}

\item{PrincipalType}{The principal type of the resource.}
}
\value{
An empty list.
}
\description{
Removes the permission for the specified principal from the specified
resource.
}
\section{Request syntax}{
\preformatted{svc$remove_resource_permission(
  AuthenticationToken = "string",
  ResourceId = "string",
  PrincipalId = "string",
  PrincipalType = "USER"|"GROUP"|"INVITE"|"ANONYMOUS"|"ORGANIZATION"
)
}
}

\keyword{internal}
