% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_search_resources}
\alias{workdocs_search_resources}
\title{Searches metadata and the content of folders, documents, document
versions, and comments}
\usage{
workdocs_search_resources(
  AuthenticationToken = NULL,
  QueryText = NULL,
  QueryScopes = NULL,
  OrganizationId = NULL,
  AdditionalResponseFields = NULL,
  Filters = NULL,
  OrderBy = NULL,
  Limit = NULL,
  Marker = NULL
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Not required when using Amazon Web
Services administrator credentials to access the API.}

\item{QueryText}{The String to search for. Searches across different text fields based on
request parameters. Use double quotes around the query string for exact
phrase matches.}

\item{QueryScopes}{Filter based on the text field type. A Folder has only a name and no
content. A Comment has only content and no name. A Document or Document
Version has a name and content}

\item{OrganizationId}{Filters based on the resource owner OrgId. This is a mandatory parameter
when using Admin SigV4 credentials.}

\item{AdditionalResponseFields}{A list of attributes to include in the response. Used to request fields
that are not normally returned in a standard response.}

\item{Filters}{Filters results based on entity metadata.}

\item{OrderBy}{Order by results in one or more categories.}

\item{Limit}{Max results count per page.}

\item{Marker}{The marker for the next set of results.}
}
\description{
Searches metadata and the content of folders, documents, document versions, and comments.

See \url{https://paws-r.github.io/docs/workdocs/search_resources.html} for full documentation.
}
\keyword{internal}
