% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_associate_domain}
\alias{worklink_associate_domain}
\title{Specifies a domain to be associated to Amazon WorkLink}
\usage{
worklink_associate_domain(
  FleetArn,
  DomainName,
  DisplayName = NULL,
  AcmCertificateArn
)
}
\arguments{
\item{FleetArn}{[required] The Amazon Resource Name (ARN) of the fleet.}

\item{DomainName}{[required] The fully qualified domain name (FQDN).}

\item{DisplayName}{The name to display.}

\item{AcmCertificateArn}{[required] The ARN of an issued ACM certificate that is valid for the domain being
associated.}
}
\description{
Specifies a domain to be associated to Amazon WorkLink.

See \url{https://www.paws-r-sdk.com/docs/worklink_associate_domain/} for full documentation.
}
\keyword{internal}
