% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_describe_company_network_configuration}
\alias{worklink_describe_company_network_configuration}
\title{Describes the networking configuration to access the internal websites
associated with the specified fleet}
\usage{
worklink_describe_company_network_configuration(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  VpcId = "string",
  SubnetIds = list(
    "string"
  ),
  SecurityGroupIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the networking configuration to access the internal websites
associated with the specified fleet.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_company_network_configuration(
  FleetArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
