% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_describe_device_policy_configuration}
\alias{worklink_describe_device_policy_configuration}
\title{Describes the device policy configuration for the specified fleet}
\usage{
worklink_describe_device_policy_configuration(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\description{
Describes the device policy configuration for the specified fleet.

See \url{https://www.paws-r-sdk.com/docs/worklink_describe_device_policy_configuration/} for full documentation.
}
\keyword{internal}
