% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_describe_fleet_metadata}
\alias{worklink_describe_fleet_metadata}
\title{Provides basic information for the specified fleet, excluding identity
provider, networking, and device configuration details}
\usage{
worklink_describe_fleet_metadata(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The Amazon Resource Name (ARN) of the fleet.}
}
\description{
Provides basic information for the specified fleet, excluding identity provider, networking, and device configuration details.

See \url{https://paws-r.github.io/docs/worklink/describe_fleet_metadata.html} for full documentation.
}
\keyword{internal}
