% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_describe_identity_provider_configuration}
\alias{worklink_describe_identity_provider_configuration}
\title{Describes the identity provider configuration of the specified fleet}
\usage{
worklink_describe_identity_provider_configuration(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\description{
Describes the identity provider configuration of the specified fleet.

See \url{https://paws-r.github.io/docs/worklink/describe_identity_provider_configuration.html} for full documentation.
}
\keyword{internal}
