% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_describe_identity_provider_configuration}
\alias{worklink_describe_identity_provider_configuration}
\title{Describes the identity provider configuration of the specified fleet}
\usage{
worklink_describe_identity_provider_configuration(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\value{
A list with the following syntax:\preformatted{list(
  IdentityProviderType = "SAML",
  ServiceProviderSamlMetadata = "string",
  IdentityProviderSamlMetadata = "string"
)
}
}
\description{
Describes the identity provider configuration of the specified fleet.
}
\section{Request syntax}{
\preformatted{svc$describe_identity_provider_configuration(
  FleetArn = "string"
)
}
}

\keyword{internal}
