% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_disassociate_website_certificate_authority}
\alias{worklink_disassociate_website_certificate_authority}
\title{Removes a certificate authority (CA)}
\usage{
worklink_disassociate_website_certificate_authority(FleetArn, WebsiteCaId)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{WebsiteCaId}{[required] A unique identifier for the CA.}
}
\description{
Removes a certificate authority (CA).

See \url{https://www.paws-r-sdk.com/docs/worklink_disassociate_website_certificate_authority/} for full documentation.
}
\keyword{internal}
