% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_list_tags_for_resource}
\alias{worklink_list_tags_for_resource}
\title{Retrieves a list of tags for the specified resource}
\usage{
worklink_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the fleet.}
}
\description{
Retrieves a list of tags for the specified resource.

See \url{https://www.paws-r-sdk.com/docs/worklink_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
