% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_restore_domain_access}
\alias{worklink_restore_domain_access}
\title{Moves a domain to ACTIVE status if it was in the INACTIVE status}
\usage{
worklink_restore_domain_access(FleetArn, DomainName)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{DomainName}{[required] The name of the domain.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Moves a domain to ACTIVE status if it was in the INACTIVE status.
}
\section{Request syntax}{
\preformatted{svc$restore_domain_access(
  FleetArn = "string",
  DomainName = "string"
)
}
}

\keyword{internal}
