% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklink_operations.R
\name{worklink_tag_resource}
\alias{worklink_tag_resource}
\title{Adds or overwrites one or more tags for the specified resource, such as
a fleet}
\usage{
worklink_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the fleet.}

\item{Tags}{[required] The tags to add to the resource. A tag is a key-value pair.}
}
\value{
An empty list.
}
\description{
Adds or overwrites one or more tags for the specified resource, such as
a fleet. Each tag consists of a key and an optional value. If a resource
already has a tag with the same key, this operation updates its value.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
