% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_create_resource}
\alias{workmail_create_resource}
\title{Creates a new WorkMail resource}
\usage{
workmail_create_resource(OrganizationId, Name, Type)
}
\arguments{
\item{OrganizationId}{[required] The identifier associated with the organization for which the resource
is created.}

\item{Name}{[required] The name of the new resource.}

\item{Type}{[required] The type of the new resource. The available types are \code{equipment} and
\code{room}.}
}
\description{
Creates a new WorkMail resource.

See \url{https://paws-r.github.io/docs/workmail/create_resource.html} for full documentation.
}
\keyword{internal}
