% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_create_user}
\alias{workmail_create_user}
\title{Creates a user who can be used in WorkMail by calling the
RegisterToWorkMail operation}
\usage{
workmail_create_user(OrganizationId, Name, DisplayName, Password)
}
\arguments{
\item{OrganizationId}{[required] The identifier of the organization for which the user is created.}

\item{Name}{[required] The name for the new user. WorkMail directory user names have a maximum
length of 64. All others have a maximum length of 20.}

\item{DisplayName}{[required] The display name for the new user.}

\item{Password}{[required] The password for the new user.}
}
\description{
Creates a user who can be used in WorkMail by calling the \code{\link[=workmail_register_to_work_mail]{register_to_work_mail}} operation.

See \url{https://paws-r.github.io/docs/workmail/create_user.html} for full documentation.
}
\keyword{internal}
