% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_availability_configuration}
\alias{workmail_delete_availability_configuration}
\title{Deletes the AvailabilityConfiguration for the given WorkMail
organization and domain}
\usage{
workmail_delete_availability_configuration(OrganizationId, DomainName)
}
\arguments{
\item{OrganizationId}{[required] The WorkMail organization for which the \code{AvailabilityConfiguration} will
be deleted.}

\item{DomainName}{[required] The domain for which the \code{AvailabilityConfiguration} will be deleted.}
}
\description{
Deletes the \code{AvailabilityConfiguration} for the given WorkMail organization and domain.

See \url{https://www.paws-r-sdk.com/docs/workmail_delete_availability_configuration/} for full documentation.
}
\keyword{internal}
