% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_group}
\alias{workmail_delete_group}
\title{Deletes a group from WorkMail}
\usage{
workmail_delete_group(OrganizationId, GroupId)
}
\arguments{
\item{OrganizationId}{[required] The organization that contains the group.}

\item{GroupId}{[required] The identifier of the group to be deleted.}
}
\description{
Deletes a group from WorkMail.

See \url{https://www.paws-r-sdk.com/docs/workmail_delete_group/} for full documentation.
}
\keyword{internal}
