% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_resource}
\alias{workmail_describe_resource}
\title{Returns the data available for the resource}
\usage{
workmail_describe_resource(OrganizationId, ResourceId)
}
\arguments{
\item{OrganizationId}{[required] The identifier associated with the organization for which the resource
is described.}

\item{ResourceId}{[required] The identifier of the resource to be described.

The identifier can accept \emph{ResourceId}, \emph{Resourcename}, or \emph{email}. The
following identity formats are available:
\itemize{
\item Resource ID: r-0123456789a0123456789b0123456789
\item Email address: resource@domain.tld
\item Resource name: resource
}}
}
\description{
Returns the data available for the resource.

See \url{https://www.paws-r-sdk.com/docs/workmail_describe_resource/} for full documentation.
}
\keyword{internal}
