% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_groups}
\alias{workmail_list_groups}
\title{Returns summaries of the organization's groups}
\usage{
workmail_list_groups(OrganizationId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the groups exist.}

\item{NextToken}{The token to use to retrieve the next page of results. The first call
does not contain any tokens.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
Returns summaries of the organization's groups.

See \url{https://paws-r.github.io/docs/workmail/list_groups.html} for full documentation.
}
\keyword{internal}
