% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_mailbox_permissions}
\alias{workmail_list_mailbox_permissions}
\title{Lists the mailbox permissions associated with a user, group, or resource
mailbox}
\usage{
workmail_list_mailbox_permissions(
  OrganizationId,
  EntityId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{OrganizationId}{[required] The identifier of the organization under which the user, group, or
resource exists.}

\item{EntityId}{[required] The identifier of the user, group, or resource for which to list mailbox
permissions.}

\item{NextToken}{The token to use to retrieve the next page of results. The first call
does not contain any tokens.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
Lists the mailbox permissions associated with a user, group, or resource mailbox.

See \url{https://paws-r.github.io/docs/workmail/list_mailbox_permissions.html} for full documentation.
}
\keyword{internal}
