% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_update_mailbox_quota}
\alias{workmail_update_mailbox_quota}
\title{Updates a user's current mailbox quota for a specified organization and
user}
\usage{
workmail_update_mailbox_quota(OrganizationId, UserId, MailboxQuota)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization that contains the user for whom to
update the mailbox quota.}

\item{UserId}{[required] The identifer for the user for whom to update the mailbox quota.}

\item{MailboxQuota}{[required] The updated mailbox quota, in MB, for the specified user.}
}
\description{
Updates a user's current mailbox quota for a specified organization and user.

See \url{https://www.paws-r-sdk.com/docs/workmail_update_mailbox_quota/} for full documentation.
}
\keyword{internal}
