% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_copy_workspace_image}
\alias{workspaces_copy_workspace_image}
\title{Copies the specified image from the specified Region to the current
Region}
\usage{
workspaces_copy_workspace_image(
  Name,
  Description = NULL,
  SourceImageId,
  SourceRegion,
  Tags = NULL
)
}
\arguments{
\item{Name}{[required] The name of the image.}

\item{Description}{A description of the image.}

\item{SourceImageId}{[required] The identifier of the source image.}

\item{SourceRegion}{[required] The identifier of the source Region.}

\item{Tags}{The tags for the image.}
}
\description{
Copies the specified image from the specified Region to the current Region. For more information about copying images, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/copy-custom-image.html}{Copy a Custom WorkSpaces Image}.

See \url{https://www.paws-r-sdk.com/docs/workspaces_copy_workspace_image/} for full documentation.
}
\keyword{internal}
