% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_connect_client_add_in}
\alias{workspaces_create_connect_client_add_in}
\title{Creates a client-add-in for Amazon Connect within a directory}
\usage{
workspaces_create_connect_client_add_in(ResourceId, Name, URL)
}
\arguments{
\item{ResourceId}{[required] The directory identifier for which to configure the client add-in.}

\item{Name}{[required] The name of the client add-in.}

\item{URL}{[required] The endpoint URL of the Amazon Connect client add-in.}
}
\description{
Creates a client-add-in for Amazon Connect within a directory. You can create only one Amazon Connect client add-in within a directory.

See \url{https://paws-r.github.io/docs/workspaces/create_connect_client_add_in.html} for full documentation.
}
\keyword{internal}
