% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_workspace_bundle}
\alias{workspaces_create_workspace_bundle}
\title{Creates the specified WorkSpace bundle}
\usage{
workspaces_create_workspace_bundle(
  BundleName,
  BundleDescription,
  ImageId,
  ComputeType,
  UserStorage,
  RootStorage = NULL,
  Tags = NULL
)
}
\arguments{
\item{BundleName}{[required] The name of the bundle.}

\item{BundleDescription}{[required] The description of the bundle.}

\item{ImageId}{[required] The identifier of the image that is used to create the bundle.}

\item{ComputeType}{[required]}

\item{UserStorage}{[required]}

\item{RootStorage}{}

\item{Tags}{The tags associated with the bundle.

To add tags at the same time when you're creating the bundle, you must
create an IAM policy that grants your IAM user permissions to use
\code{workspaces:CreateTags}.}
}
\description{
Creates the specified WorkSpace bundle. For more information about creating WorkSpace bundles, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/create-custom-bundle.html}{Create a Custom WorkSpaces Image and Bundle}.

See \url{https://www.paws-r-sdk.com/docs/workspaces_create_workspace_bundle/} for full documentation.
}
\keyword{internal}
