% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_workspaces}
\alias{workspaces_create_workspaces}
\title{Creates one or more WorkSpaces}
\usage{
workspaces_create_workspaces(Workspaces)
}
\arguments{
\item{Workspaces}{[required] The WorkSpaces to create. You can specify up to 25 WorkSpaces.}
}
\description{
Creates one or more WorkSpaces.

See \url{https://www.paws-r-sdk.com/docs/workspaces_create_workspaces/} for full documentation.
}
\keyword{internal}
