% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_create_workspaces_pool}
\alias{workspaces_create_workspaces_pool}
\title{Creates a pool of WorkSpaces}
\usage{
workspaces_create_workspaces_pool(
  PoolName,
  Description,
  BundleId,
  DirectoryId,
  Capacity,
  Tags = NULL,
  ApplicationSettings = NULL,
  TimeoutSettings = NULL,
  RunningMode = NULL
)
}
\arguments{
\item{PoolName}{[required] The name of the pool.}

\item{Description}{[required] The pool description.}

\item{BundleId}{[required] The identifier of the bundle for the pool.}

\item{DirectoryId}{[required] The identifier of the directory for the pool.}

\item{Capacity}{[required] The user capacity of the pool.}

\item{Tags}{The tags for the pool.}

\item{ApplicationSettings}{Indicates the application settings of the pool.}

\item{TimeoutSettings}{Indicates the timeout settings of the pool.}

\item{RunningMode}{The running mode for the pool.}
}
\description{
Creates a pool of WorkSpaces.

See \url{https://www.paws-r-sdk.com/docs/workspaces_create_workspaces_pool/} for full documentation.
}
\keyword{internal}
