% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_delete_connect_client_add_in}
\alias{workspaces_delete_connect_client_add_in}
\title{Deletes a client-add-in for Amazon Connect that is configured within a
directory}
\usage{
workspaces_delete_connect_client_add_in(AddInId, ResourceId)
}
\arguments{
\item{AddInId}{[required] The identifier of the client add-in to delete.}

\item{ResourceId}{[required] The directory identifier for which the client add-in is configured.}
}
\description{
Deletes a client-add-in for Amazon Connect that is configured within a directory.

See \url{https://paws-r.github.io/docs/workspaces/delete_connect_client_add_in.html} for full documentation.
}
\keyword{internal}
