% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_delete_tags}
\alias{workspaces_delete_tags}
\title{Deletes the specified tags from the specified WorkSpaces resource}
\usage{
workspaces_delete_tags(ResourceId, TagKeys)
}
\arguments{
\item{ResourceId}{[required] The identifier of the WorkSpaces resource. The supported resource types
are WorkSpaces, registered directories, images, custom bundles, IP
access control groups, and connection aliases.}

\item{TagKeys}{[required] The tag keys.}
}
\value{
An empty list.
}
\description{
Deletes the specified tags from the specified WorkSpaces resource.
}
\section{Request syntax}{
\preformatted{svc$delete_tags(
  ResourceId = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
