% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_delete_workspace_bundle}
\alias{workspaces_delete_workspace_bundle}
\title{Deletes the specified WorkSpace bundle}
\usage{
workspaces_delete_workspace_bundle(BundleId)
}
\arguments{
\item{BundleId}{The identifier of the bundle.}
}
\value{
An empty list.
}
\description{
Deletes the specified WorkSpace bundle. For more information about
deleting WorkSpace bundles, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/delete_bundle.html}{Delete a Custom WorkSpaces Bundle or Image}.
}
\section{Request syntax}{
\preformatted{svc$delete_workspace_bundle(
  BundleId = "string"
)
}
}

\keyword{internal}
