% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_deregister_workspace_directory}
\alias{workspaces_deregister_workspace_directory}
\title{Deregisters the specified directory}
\usage{
workspaces_deregister_workspace_directory(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory. If any WorkSpaces are registered to
this directory, you must remove them before you deregister the
directory, or you will receive an OperationNotSupportedException error.}
}
\description{
Deregisters the specified directory. This operation is asynchronous and returns before the WorkSpace directory is deregistered. If any WorkSpaces are registered to this directory, you must remove them before you can deregister the directory.

See \url{https://paws-r.github.io/docs/workspaces/deregister_workspace_directory.html} for full documentation.
}
\keyword{internal}
