% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_client_properties}
\alias{workspaces_describe_client_properties}
\title{Retrieves a list that describes one or more specified Amazon WorkSpaces
clients}
\usage{
workspaces_describe_client_properties(ResourceIds)
}
\arguments{
\item{ResourceIds}{[required] The resource identifier, in the form of directory IDs.}
}
\description{
Retrieves a list that describes one or more specified Amazon WorkSpaces clients.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_client_properties/} for full documentation.
}
\keyword{internal}
