% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_connect_client_add_ins}
\alias{workspaces_describe_connect_client_add_ins}
\title{Retrieves a list of Amazon Connect client add-ins that have been created}
\usage{
workspaces_describe_connect_client_add_ins(
  ResourceId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ResourceId}{[required] The directory identifier for which the client add-in is configured.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}

\item{MaxResults}{The maximum number of items to return.}
}
\description{
Retrieves a list of Amazon Connect client add-ins that have been created.

See \url{https://paws-r.github.io/docs/workspaces/describe_connect_client_add_ins.html} for full documentation.
}
\keyword{internal}
