% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_connection_alias_permissions}
\alias{workspaces_describe_connection_alias_permissions}
\title{Describes the permissions that the owner of a connection alias has
granted to another Amazon Web Services account for the specified
connection alias}
\usage{
workspaces_describe_connection_alias_permissions(
  AliasId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AliasId}{[required] The identifier of the connection alias.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Describes the permissions that the owner of a connection alias has granted to another Amazon Web Services account for the specified connection alias. For more information, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/cross-region-redirection.html}{Cross-Region Redirection for Amazon WorkSpaces}.

See \url{https://paws-r.github.io/docs/workspaces/describe_connection_alias_permissions.html} for full documentation.
}
\keyword{internal}
