% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_tags}
\alias{workspaces_describe_tags}
\title{Describes the specified tags for the specified WorkSpaces resource}
\usage{
workspaces_describe_tags(ResourceId)
}
\arguments{
\item{ResourceId}{[required] The identifier of the WorkSpaces resource. The supported resource types
are WorkSpaces, registered directories, images, custom bundles, IP
access control groups, and connection aliases.}
}
\description{
Describes the specified tags for the specified WorkSpaces resource.

See \url{https://paws-r.github.io/docs/workspaces/describe_tags.html} for full documentation.
}
\keyword{internal}
