% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_workspace_directories}
\alias{workspaces_describe_workspace_directories}
\title{Describes the available directories that are registered with Amazon
WorkSpaces}
\usage{
workspaces_describe_workspace_directories(
  DirectoryIds = NULL,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DirectoryIds}{The identifiers of the directories. If the value is null, all
directories are retrieved.}

\item{Limit}{The maximum number of directories to return.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}
}
\description{
Describes the available directories that are registered with Amazon WorkSpaces.

See \url{https://paws-r.github.io/docs/workspaces/describe_workspace_directories.html} for full documentation.
}
\keyword{internal}
