% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_workspace_image_permissions}
\alias{workspaces_describe_workspace_image_permissions}
\title{Describes the permissions that the owner of an image has granted to
other Amazon Web Services accounts for an image}
\usage{
workspaces_describe_workspace_image_permissions(
  ImageId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ImageId}{[required] The identifier of the image.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}

\item{MaxResults}{The maximum number of items to return.}
}
\description{
Describes the permissions that the owner of an image has granted to other Amazon Web Services accounts for an image.

See \url{https://paws-r.github.io/docs/workspaces/describe_workspace_image_permissions.html} for full documentation.
}
\keyword{internal}
