% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_workspace_snapshots}
\alias{workspaces_describe_workspace_snapshots}
\title{Describes the snapshots for the specified WorkSpace}
\usage{
workspaces_describe_workspace_snapshots(WorkspaceId)
}
\arguments{
\item{WorkspaceId}{[required] The identifier of the WorkSpace.}
}
\value{
A list with the following syntax:\preformatted{list(
  RebuildSnapshots = list(
    list(
      SnapshotTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  RestoreSnapshots = list(
    list(
      SnapshotTime = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}
\description{
Describes the snapshots for the specified WorkSpace.
}
\section{Request syntax}{
\preformatted{svc$describe_workspace_snapshots(
  WorkspaceId = "string"
)
}
}

\keyword{internal}
