% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_workspaces_pool_sessions}
\alias{workspaces_describe_workspaces_pool_sessions}
\title{Retrieves a list that describes the streaming sessions for a specified
pool}
\usage{
workspaces_describe_workspaces_pool_sessions(
  PoolId,
  UserId = NULL,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{PoolId}{[required] The identifier of the pool.}

\item{UserId}{The identifier of the user.}

\item{Limit}{The maximum number of items to return.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}
}
\description{
Retrieves a list that describes the streaming sessions for a specified pool.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_workspaces_pool_sessions/} for full documentation.
}
\keyword{internal}
