% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_workspaces_pools}
\alias{workspaces_describe_workspaces_pools}
\title{Describes the specified WorkSpaces Pools}
\usage{
workspaces_describe_workspaces_pools(
  PoolIds = NULL,
  Filters = NULL,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{PoolIds}{The identifier of the WorkSpaces Pools.}

\item{Filters}{The filter conditions for the WorkSpaces Pool to return.}

\item{Limit}{The maximum number of items to return.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated,
provide this token to receive the next set of results.}
}
\description{
Describes the specified WorkSpaces Pools.

See \url{https://www.paws-r-sdk.com/docs/workspaces_describe_workspaces_pools/} for full documentation.
}
\keyword{internal}
