% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_disassociate_connection_alias}
\alias{workspaces_disassociate_connection_alias}
\title{Disassociates a connection alias from a directory}
\usage{
workspaces_disassociate_connection_alias(AliasId)
}
\arguments{
\item{AliasId}{[required] The identifier of the connection alias to disassociate.}
}
\value{
An empty list.
}
\description{
Disassociates a connection alias from a directory. Disassociating a
connection alias disables cross-Region redirection between two
directories in different AWS Regions. For more information, see
\href{https://docs.aws.amazon.com/workspaces/latest/adminguide/cross-region-redirection.html}{Cross-Region Redirection for Amazon WorkSpaces}.

Before performing this operation, call
\href{https://docs.aws.amazon.com/workspaces/latest/api/API_DescribeConnectionAliases.html}{DescribeConnectionAliases}
to make sure that the current state of the connection alias is
\code{CREATED}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disassociate_connection_alias(
  AliasId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
