% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_modify_client_properties}
\alias{workspaces_modify_client_properties}
\title{Modifies the properties of the specified Amazon WorkSpaces clients}
\usage{
workspaces_modify_client_properties(ResourceId, ClientProperties)
}
\arguments{
\item{ResourceId}{[required] The resource identifiers, in the form of directory IDs.}

\item{ClientProperties}{[required] Information about the Amazon WorkSpaces client.}
}
\description{
Modifies the properties of the specified Amazon WorkSpaces clients.

See \url{https://paws-r.github.io/docs/workspaces/modify_client_properties.html} for full documentation.
}
\keyword{internal}
