% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_modify_workspace_creation_properties}
\alias{workspaces_modify_workspace_creation_properties}
\title{Modify the default properties used to create WorkSpaces}
\usage{
workspaces_modify_workspace_creation_properties(
  ResourceId,
  WorkspaceCreationProperties
)
}
\arguments{
\item{ResourceId}{[required] The identifier of the directory.}

\item{WorkspaceCreationProperties}{[required] The default properties for creating WorkSpaces.}
}
\description{
Modify the default properties used to create WorkSpaces.

See \url{https://paws-r.github.io/docs/workspaces/modify_workspace_creation_properties.html} for full documentation.
}
\keyword{internal}
