% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_modify_workspace_properties}
\alias{workspaces_modify_workspace_properties}
\title{Modifies the specified WorkSpace properties}
\usage{
workspaces_modify_workspace_properties(
  WorkspaceId,
  WorkspaceProperties = NULL,
  DataReplication = NULL
)
}
\arguments{
\item{WorkspaceId}{[required] The identifier of the WorkSpace.}

\item{WorkspaceProperties}{The properties of the WorkSpace.}

\item{DataReplication}{Indicates the data replication status.}
}
\description{
Modifies the specified WorkSpace properties. For important information about how to modify the size of the root and user volumes, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/modify-workspaces.html}{Modify a WorkSpace}.

See \url{https://www.paws-r-sdk.com/docs/workspaces_modify_workspace_properties/} for full documentation.
}
\keyword{internal}
