% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_reboot_workspaces}
\alias{workspaces_reboot_workspaces}
\title{Reboots the specified WorkSpaces}
\usage{
workspaces_reboot_workspaces(RebootWorkspaceRequests)
}
\arguments{
\item{RebootWorkspaceRequests}{[required] The WorkSpaces to reboot. You can specify up to 25 WorkSpaces.}
}
\description{
Reboots the specified WorkSpaces.

See \url{https://paws-r.github.io/docs/workspaces/reboot_workspaces.html} for full documentation.
}
\keyword{internal}
