% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_rebuild_workspaces}
\alias{workspaces_rebuild_workspaces}
\title{Rebuilds the specified WorkSpace}
\usage{
workspaces_rebuild_workspaces(RebuildWorkspaceRequests)
}
\arguments{
\item{RebuildWorkspaceRequests}{[required] The WorkSpace to rebuild. You can specify a single WorkSpace.}
}
\description{
Rebuilds the specified WorkSpace.

See \url{https://paws-r.github.io/docs/workspaces/rebuild_workspaces.html} for full documentation.
}
\keyword{internal}
