% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_revoke_ip_rules}
\alias{workspaces_revoke_ip_rules}
\title{Removes one or more rules from the specified IP access control group}
\usage{
workspaces_revoke_ip_rules(GroupId, UserRules)
}
\arguments{
\item{GroupId}{[required] The identifier of the group.}

\item{UserRules}{[required] The rules to remove from the group.}
}
\description{
Removes one or more rules from the specified IP access control group.

See \url{https://www.paws-r-sdk.com/docs/workspaces_revoke_ip_rules/} for full documentation.
}
\keyword{internal}
