% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_start_workspaces}
\alias{workspaces_start_workspaces}
\title{Starts the specified WorkSpaces}
\usage{
workspaces_start_workspaces(StartWorkspaceRequests)
}
\arguments{
\item{StartWorkspaceRequests}{[required] The WorkSpaces to start. You can specify up to 25 WorkSpaces.}
}
\description{
Starts the specified WorkSpaces.

See \url{https://paws-r.github.io/docs/workspaces/start_workspaces.html} for full documentation.
}
\keyword{internal}
