% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_stop_workspaces_pool}
\alias{workspaces_stop_workspaces_pool}
\title{Stops the specified pool}
\usage{
workspaces_stop_workspaces_pool(PoolId)
}
\arguments{
\item{PoolId}{[required] The identifier of the pool.}
}
\description{
Stops the specified pool.

See \url{https://www.paws-r-sdk.com/docs/workspaces_stop_workspaces_pool/} for full documentation.
}
\keyword{internal}
