% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_terminate_workspaces}
\alias{workspaces_terminate_workspaces}
\title{Terminates the specified WorkSpaces}
\usage{
workspaces_terminate_workspaces(TerminateWorkspaceRequests)
}
\arguments{
\item{TerminateWorkspaceRequests}{[required] The WorkSpaces to terminate. You can specify up to 25 WorkSpaces.}
}
\description{
Terminates the specified WorkSpaces.
}
\details{
Terminating a WorkSpace is a permanent action and cannot be undone. The
user's data is destroyed. If you need to archive any user data, contact
Amazon Web Services before terminating the WorkSpace.

You can terminate a WorkSpace that is in any state except \code{SUSPENDED}.

This operation is asynchronous and returns before the WorkSpaces have
been completely terminated.
}
\section{Request syntax}{
\preformatted{svc$terminate_workspaces(
  TerminateWorkspaceRequests = list(
    list(
      WorkspaceId = "string"
    )
  )
)
}
}

\keyword{internal}
