% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_terminate_workspaces_pool_session}
\alias{workspaces_terminate_workspaces_pool_session}
\title{Terminates the pool session}
\usage{
workspaces_terminate_workspaces_pool_session(SessionId)
}
\arguments{
\item{SessionId}{[required] The identifier of the pool session.}
}
\description{
Terminates the pool session.

See \url{https://www.paws-r-sdk.com/docs/workspaces_terminate_workspaces_pool_session/} for full documentation.
}
\keyword{internal}
