% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_update_rules_of_ip_group}
\alias{workspaces_update_rules_of_ip_group}
\title{Replaces the current rules of the specified IP access control group with
the specified rules}
\usage{
workspaces_update_rules_of_ip_group(GroupId, UserRules)
}
\arguments{
\item{GroupId}{[required] The identifier of the group.}

\item{UserRules}{[required] One or more rules.}
}
\description{
Replaces the current rules of the specified IP access control group with
the specified rules.
}
\section{Request syntax}{
\preformatted{svc$update_rules_of_ip_group(
  GroupId = "string",
  UserRules = list(
    list(
      ipRule = "string",
      ruleDesc = "string"
    )
  )
)
}
}

\keyword{internal}
