% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_update_workspace_bundle}
\alias{workspaces_update_workspace_bundle}
\title{Updates a WorkSpace bundle with a new image}
\usage{
workspaces_update_workspace_bundle(BundleId = NULL, ImageId = NULL)
}
\arguments{
\item{BundleId}{The identifier of the bundle.}

\item{ImageId}{The identifier of the image.}
}
\description{
Updates a WorkSpace bundle with a new image. For more information about updating WorkSpace bundles, see \href{https://docs.aws.amazon.com/workspaces/latest/adminguide/update-custom-bundle.html}{Update a Custom WorkSpaces Bundle}.

See \url{https://www.paws-r-sdk.com/docs/workspaces_update_workspace_bundle/} for full documentation.
}
\keyword{internal}
