% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_update_workspaces_pool}
\alias{workspaces_update_workspaces_pool}
\title{Updates the specified pool}
\usage{
workspaces_update_workspaces_pool(
  PoolId,
  Description = NULL,
  BundleId = NULL,
  DirectoryId = NULL,
  Capacity = NULL,
  ApplicationSettings = NULL,
  TimeoutSettings = NULL,
  RunningMode = NULL
)
}
\arguments{
\item{PoolId}{[required] The identifier of the specified pool to update.}

\item{Description}{Describes the specified pool to update.}

\item{BundleId}{The identifier of the bundle.}

\item{DirectoryId}{The identifier of the directory.}

\item{Capacity}{The desired capacity for the pool.}

\item{ApplicationSettings}{The persistent application settings for users in the pool.}

\item{TimeoutSettings}{Indicates the timeout settings of the specified pool.}

\item{RunningMode}{The desired running mode for the pool. The running mode can only be
updated when the pool is in a stopped state.}
}
\description{
Updates the specified pool.

See \url{https://www.paws-r-sdk.com/docs/workspaces_update_workspaces_pool/} for full documentation.
}
\keyword{internal}
