% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_associate_network_settings}
\alias{workspacesweb_associate_network_settings}
\title{Associates a network settings resource with a web portal}
\usage{
workspacesweb_associate_network_settings(portalArn, networkSettingsArn)
}
\arguments{
\item{portalArn}{[required] The ARN of the web portal.}

\item{networkSettingsArn}{[required] The ARN of the network settings.}
}
\description{
Associates a network settings resource with a web portal.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_associate_network_settings/} for full documentation.
}
\keyword{internal}
