% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_create_data_protection_settings}
\alias{workspacesweb_create_data_protection_settings}
\title{Creates a data protection settings resource that can be associated with
a web portal}
\usage{
workspacesweb_create_data_protection_settings(
  displayName = NULL,
  description = NULL,
  tags = NULL,
  customerManagedKey = NULL,
  additionalEncryptionContext = NULL,
  inlineRedactionConfiguration = NULL,
  clientToken = NULL
)
}
\arguments{
\item{displayName}{The display name of the data protection settings.}

\item{description}{The description of the data protection settings.}

\item{tags}{The tags to add to the data protection settings resource. A tag is a
key-value pair.}

\item{customerManagedKey}{The custom managed key of the data protection settings.}

\item{additionalEncryptionContext}{Additional encryption context of the data protection settings.}

\item{inlineRedactionConfiguration}{The inline redaction configuration of the data protection settings that
will be applied to all sessions.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Idempotency ensures that an API request
completes only once. With an idempotent request, if the original request
completes successfully, subsequent retries with the same client token
returns the result from the original successful request.

If you do not specify a client token, one is automatically generated by
the Amazon Web Services SDK.}
}
\description{
Creates a data protection settings resource that can be associated with a web portal.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_create_data_protection_settings/} for full documentation.
}
\keyword{internal}
