% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_disassociate_trust_store}
\alias{workspacesweb_disassociate_trust_store}
\title{Disassociates a trust store from a web portal}
\usage{
workspacesweb_disassociate_trust_store(portalArn)
}
\arguments{
\item{portalArn}{[required] The ARN of the web portal.}
}
\description{
Disassociates a trust store from a web portal.

See \url{https://paws-r.github.io/docs/workspacesweb/disassociate_trust_store.html} for full documentation.
}
\keyword{internal}
