% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_browser_settings}
\alias{workspacesweb_get_browser_settings}
\title{Gets browser settings}
\usage{
workspacesweb_get_browser_settings(browserSettingsArn)
}
\arguments{
\item{browserSettingsArn}{[required] The ARN of the browser settings.}
}
\description{
Gets browser settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_browser_settings/} for full documentation.
}
\keyword{internal}
