% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_identity_provider}
\alias{workspacesweb_get_identity_provider}
\title{Gets the identity provider}
\usage{
workspacesweb_get_identity_provider(identityProviderArn)
}
\arguments{
\item{identityProviderArn}{[required] The ARN of the identity provider.}
}
\description{
Gets the identity provider.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_identity_provider/} for full documentation.
}
\keyword{internal}
