% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_trust_store}
\alias{workspacesweb_get_trust_store}
\title{Gets the trust store}
\usage{
workspacesweb_get_trust_store(trustStoreArn)
}
\arguments{
\item{trustStoreArn}{[required] The ARN of the trust store.}
}
\description{
Gets the trust store.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_trust_store/} for full documentation.
}
\keyword{internal}
