% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_trust_store_certificate}
\alias{workspacesweb_get_trust_store_certificate}
\title{Gets the trust store certificate}
\usage{
workspacesweb_get_trust_store_certificate(trustStoreArn, thumbprint)
}
\arguments{
\item{trustStoreArn}{[required] The ARN of the trust store certificate.}

\item{thumbprint}{[required] The thumbprint of the trust store certificate.}
}
\description{
Gets the trust store certificate.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_trust_store_certificate/} for full documentation.
}
\keyword{internal}
