% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_user_access_logging_settings}
\alias{workspacesweb_get_user_access_logging_settings}
\title{Gets user access logging settings}
\usage{
workspacesweb_get_user_access_logging_settings(userAccessLoggingSettingsArn)
}
\arguments{
\item{userAccessLoggingSettingsArn}{[required] The ARN of the user access logging settings.}
}
\description{
Gets user access logging settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_user_access_logging_settings/} for full documentation.
}
\keyword{internal}
