% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_get_user_settings}
\alias{workspacesweb_get_user_settings}
\title{Gets user settings}
\usage{
workspacesweb_get_user_settings(userSettingsArn)
}
\arguments{
\item{userSettingsArn}{[required] The ARN of the user settings.}
}
\description{
Gets user settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_get_user_settings/} for full documentation.
}
\keyword{internal}
