% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_list_identity_providers}
\alias{workspacesweb_list_identity_providers}
\title{Retrieves a list of identity providers for a specific web portal}
\usage{
workspacesweb_list_identity_providers(
  maxResults = NULL,
  nextToken = NULL,
  portalArn
)
}
\arguments{
\item{maxResults}{The maximum number of results to be included in the next page.}

\item{nextToken}{The pagination token used to retrieve the next page of results for this
operation.}

\item{portalArn}{[required] The ARN of the web portal.}
}
\description{
Retrieves a list of identity providers for a specific web portal.

See \url{https://paws-r.github.io/docs/workspacesweb/list_identity_providers.html} for full documentation.
}
\keyword{internal}
