% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_list_portals}
\alias{workspacesweb_list_portals}
\title{Retrieves a list or web portals}
\usage{
workspacesweb_list_portals(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to be included in the next page.}

\item{nextToken}{The pagination token used to retrieve the next page of results for this
operation.}
}
\description{
Retrieves a list or web portals.

See \url{https://paws-r.github.io/docs/workspacesweb/list_portals.html} for full documentation.
}
\keyword{internal}
