% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_list_trust_store_certificates}
\alias{workspacesweb_list_trust_store_certificates}
\title{Retrieves a list of trust store certificates}
\usage{
workspacesweb_list_trust_store_certificates(
  maxResults = NULL,
  nextToken = NULL,
  trustStoreArn
)
}
\arguments{
\item{maxResults}{The maximum number of results to be included in the next page.}

\item{nextToken}{The pagination token used to retrieve the next page of results for this
operation.}

\item{trustStoreArn}{[required] The ARN of the trust store}
}
\description{
Retrieves a list of trust store certificates.

See \url{https://paws-r.github.io/docs/workspacesweb/list_trust_store_certificates.html} for full documentation.
}
\keyword{internal}
