% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_list_trust_store_certificates}
\alias{workspacesweb_list_trust_store_certificates}
\title{Retrieves a list of trust store certificates}
\usage{
workspacesweb_list_trust_store_certificates(
  trustStoreArn,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{trustStoreArn}{[required] The ARN of the trust store}

\item{nextToken}{The pagination token used to retrieve the next page of results for this
operation.}

\item{maxResults}{The maximum number of results to be included in the next page.}
}
\description{
Retrieves a list of trust store certificates.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_list_trust_store_certificates/} for full documentation.
}
\keyword{internal}
