% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_list_trust_stores}
\alias{workspacesweb_list_trust_stores}
\title{Retrieves a list of trust stores}
\usage{
workspacesweb_list_trust_stores(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The pagination token used to retrieve the next page of results for this
operation.}

\item{maxResults}{The maximum number of results to be included in the next page.}
}
\description{
Retrieves a list of trust stores.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_list_trust_stores/} for full documentation.
}
\keyword{internal}
