% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspacesweb_operations.R
\name{workspacesweb_update_user_access_logging_settings}
\alias{workspacesweb_update_user_access_logging_settings}
\title{Updates the user access logging settings}
\usage{
workspacesweb_update_user_access_logging_settings(
  userAccessLoggingSettingsArn,
  kinesisStreamArn = NULL,
  clientToken = NULL
)
}
\arguments{
\item{userAccessLoggingSettingsArn}{[required] The ARN of the user access logging settings.}

\item{kinesisStreamArn}{The ARN of the Kinesis stream.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Idempotency ensures that an API request
completes only once. With an idempotent request, if the original request
completes successfully, subsequent retries with the same client token
return the result from the original successful request.

If you do not specify a client token, one is automatically generated by
the Amazon Web Services SDK.}
}
\description{
Updates the user access logging settings.

See \url{https://www.paws-r-sdk.com/docs/workspacesweb_update_user_access_logging_settings/} for full documentation.
}
\keyword{internal}
